### 2019-07-08 #23 @whedon
[PRE REVIEW]: StateSpaceModels.jl: a Julia Package for Time-Series Analysis in a State-Space Framework
**Submitting author:** @raphaelsaavedra (<a href="http://orcid.org/0000-0003-0095-4114">Raphael Saavedra</a>)
**Repository:** <a href="https://github.com/LAMPSPUC/StateSpaceModels.jl/" target ="_blank">https://github.com/LAMPSPUC/StateSpaceModels.jl/</a>

Thanks for submitting your paper to JuliaCon @raphaelsaavedra. The JuliaCon editor @matbesancon, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@raphaelsaavedra if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
