### 2019-07-09 #28 @whedon
[REVIEW]: StateSpaceModels.jl: a Julia Package for Time-Series Analysis in a State-Space Framework
**Submitting author:** @raphaelsaavedra (<a href="http://orcid.org/0000-0003-0095-4114">Raphael Saavedra</a>)
**Repository:** <a href="https://github.com/LAMPSPUC/StateSpaceModels.jl/" target ="_blank">https://github.com/LAMPSPUC/StateSpaceModels.jl/</a>
