+++ paper/ref.bib
@@ -6,14 +6,16 @@
   number={1},
   pages={65--98},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/141000671}
 }
 
 @book{durbin2012time,
   title={Time Series Analysis by State Space Methods},
   author={Durbin, James and Koopman, Siem Jan},
   year={2012},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/acprof:oso/9780199641178.001.0001}
 }
 
 @misc{statespacemodels,
@@ -45,7 +47,8 @@
   title={Linear System Theory: The State Space Approach},
   author={Zadeh, Lotfi and Desoer, Charles},
   year={2008},
-  publisher={Courier Dover Publications}
+  publisher={Courier Dover Publications},
+  doi={10.1109/proc.1965.4237}
 }
 
 @article{kfas,
@@ -88,7 +91,8 @@
   number={1},
   pages={35--45},
   year={1960},
-  publisher={American Society of Mechanical Engineers}
+  publisher={American Society of Mechanical Engineers},
+  doi={10.1115/1.3662552}
 }
 
 @article{koopman2000stamp,
@@ -113,7 +117,8 @@
   volume={3},
   number={24},
   year={2018},
-  publisher={Open Journals}
+  publisher={Open Journals},
+  doi={10.21105/joss.00615}
 }
 
 @article{liu1989limited,
@@ -131,5 +136,6 @@
   title={Simulating Low and High-Frequency Energy Demand Scenarios in a Unified Framework--{P}art {I}: Low-Frequency Simulation},
   author={Saavedra, Raphael and Bodin, Guilherme and others},
   booktitle={Proceedings of the L Simpósio Brasileiro de Pesquisa Operacional},
-  year={2018}
+  year={2018},
+  doi={10.17771/PUCRio.IRR.33804}
 }
