### 2019-07-09 d6cf688
update header print function
