### 2019-07-09 #28 @raphaelsaavedra
@matbesancon Thanks! I have included the suggested DOIs, except for "A new approach to linear filtering and prediction problems". I included what I believe is the correct DOI, 10.1115/1.3662552. The updated version is already in the master, do I need to make a new release?
