### 2019-07-09 #28 @matbesancon
No I don't think it is necessary now

On Tue, Jul 9, 2019, 21:10 Raphael Saavedra <notifications@github.com>
wrote:

> @matbesancon <https://github.com/matbesancon> Thanks! I have included the
> suggested DOIs, except for "A new approach to linear filtering and
> prediction problems". I included what I believe is the correct DOI,
> 10.1115/1.3662552. The updated version is already in the master, do I need
> to make a new release?
>
