+++ docs/src/index.md
@@ -1,34 +1,51 @@
-# StateSpaceModels.jl Documentation
+# StateSpaceModels.jl
+
+| **Build Status** | **Coverage** | **Documentation** |
+|:-----------------:|:-----------------:|:-----------------:|
+| [![Build Status][build-img]][build-url] | [![Codecov branch][codecov-img]][codecov-url] |[![](https://img.shields.io/badge/docs-latest-blue.svg)](https://lampspuc.github.io/StateSpaceModels.jl/latest/)
+
+StateSpaceModels.jl is a package for modeling, forecasting, and simulating time series in a state-space framework. Implementations were made based on the book "Time Series Analysis by State Space Methods" (2012) by James Durbin and Siem Jan Koopman. The notation of the variables in the code also follows the book.
 
 ## Installation
 
-This package is registered in METADATA so you will can do `Pkg.add` it as follows:
+This package is registered in METADATA so you can `Pkg.add` it as follows:
 ```julia
 pkg> add StateSpaceModels
 ```
 
-## Notes
+## Features
 
-StateSpaceModels.jl is a package for modeling, forecasting and simulating time series in a state space framework based on structural models of Harvey (1989), in which a time series is decomposed in trend, slope and seasonals. Implementations were made based on the book "Time Series Analysis by State Space Methods" (2012) by J. Durbin and S. J. Koopman.
+Current features:
+* Kalman filter and smoother
+* Square-root filter and smoother
+* Maximum likelihood estimation
+* Forecasting
+* Monte Carlo simulation
+* Multivariate modeling
+* User-defined models (input any `Z`, `T`, and `R`)
+* Several predefined models, including:
+  1. Basic structural model (trend, slope, seasonal)
+  2. Structural model with exogenous variables
+  3. Linear trend model
+  4. Local level model
+* Completion of missing values
+
+Planned features:
+* Exact initialization of the Kalman filter
+* EM algorithm for maximum likelihood estimation
+* Univariate treatment of multivariate models
+
+[build-img]: https://travis-ci.org/LAMPSPUC/StateSpaceModels.jl.svg?branch=master
+[build-url]: https://travis-ci.org/LAMPSPUC/StateSpaceModels.jl
+
+[codecov-img]: https://codecov.io/gh/LAMPSPUC/StateSpaceModels.jl/coverage.svg?branch=master
+[codecov-url]: https://codecov.io/gh/LAMPSPUC/StateSpaceModels.jl?branch=master
+
+## Works
 
 Works using this package:
 
 [Simulating Low and High-Frequency Energy
 Demand Scenarios in a Unified Framework – Part
-I: Low-Frequency Simulation](https://www.maxwell.vrac.puc-rio.br/33804/33804.PDF)
-
-## Features
-
-Current features:
-* Basic structural model (trend, slope, seasonal)
-* Exogenous variables
-* Square-root Kalman Filter and smoother
-* Big Kappa initialization
-* Monte Carlo simulation
-* Maximum likelihood estimation
-* Multivariate modeling
-
-Future features (work in progress):
-* User-defined model
-* Forecasting and confidence intervals
-* Completion of missing values
+I: Low-Frequency Simulation](https://proceedings.science/sbpo/papers/simulando-cenarios-de-demanda-em-baixa-e-alta-frequencia-em-um-framework-unificado---parte-i%3A-simulacao-em-baixa-frequen).
+In: L Simpósio Brasileiro de Pesquisa Operacional, Rio de Janeiro, Brazil.
