+++ docs/src/manual.md
@@ -7,8 +7,8 @@ In this package we consider the following state space model
 ```math
 \begin{gather*}
     \begin{aligned}
-        y_t &= Z_t \alpha_t  + \varepsilon_t \quad \quad \quad t = 1 \dots n\\
-        \alpha_{t+1} &= T_t \alpha_t + R_t \eta_t
+        y_t &= Z_t \alpha_t  + \varepsilon_t, \quad \quad \quad t = 1 \dots n, \\
+        \alpha_{t+1} &= T_t \alpha_t + R_t \eta_t,
     \end{aligned}
 \end{gather*}
 ```
@@ -36,13 +36,12 @@ NID
 \end{pmatrix}
 ```
 
-
 ## Data Structures
 
 ```@docs
 StateSpaceDimensions
 StateSpaceModel
-StateSpaceParameters
+StateSpaceCovariance
 SmoothedState
 FilterOutput
 StateSpace
@@ -86,7 +85,6 @@ The structural model is defined by
 ```@docs
 structuralmodel
 ```
-
 
 ## Estimation
 The model estimation is made using the function `statespace(y, s; X, nseeds)`. It receives as argument the time series `y` and the desired seasonality `s`. The user can input exogenous variables using optional argument `X` and specify the desired number of random seeds `nseeds` to perform the estimation.
