### 2019-07-27 eb5c4c8
improving manual
