+++ build/structures.jl
@@ -1,0 +1,212 @@
+export StateSpaceDimensions, StateSpaceModel, StateSpaceCovariance, SmoothedState,
+        FilterOutput, StateSpace, KalmanFilter, SquareRootFilter, RandomSeedsLBFGS
+
+# Abstract types
+"""
+TODO
+"""
+abstract type AbstractFilter end
+
+"""
+TODO
+"""
+abstract type AbstractSmoother end
+
+"""
+TODO
+"""
+abstract type AbstractOptimizationMethod end
+
+# Auxiliary structure for square-root Kalman filter
+mutable struct SquareRootFilter <: AbstractFilter
+    a::Matrix{Float64} # predictive state
+    v::Matrix{Float64} # innovations
+    sqrtP::Array{Float64, 3} # lower triangular matrix with sqrt-covariance of the predictive state
+    sqrtF::Array{Float64, 3} # lower triangular matrix with sqrt-covariance of the innovations
+    steadystate::Bool # flag that indicates if steady state was attained
+    tsteady::Int # instant when steady state was attained; in case it wasn't, tsteady = n+1
+    K::Array{Float64, 3} # Kalman gain
+end
+
+# Auxiliary structure for Kalman filter
+mutable struct KalmanFilter <: AbstractFilter
+    a::Matrix{Float64} # predictive state
+    v::Matrix{Float64} # innovations
+    P::Array{Float64, 3} # covariance matrix of the predictive state
+    F::Array{Float64, 3} # covariance matrix of the innovations
+    steadystate::Bool # flag that indicates if steady state was attained
+    tsteady::Int # instant when steady state was attained; in case it wasn't, tsteady = n+1
+    K::Array{Float64, 3} # Kalman gain
+end
+
+# Auxiliary structure for smoother
+mutable struct Smoother <: AbstractSmoother
+    alpha::Matrix{Float64} # smoothed state
+    V::Array{Float64, 3} # variance of smoothed state
+end
+
+
+# Concrete structs for optimization methods
+"""
+TODO
+"""
+mutable struct RandomSeedsLBFGS <: AbstractOptimizationMethod
+    f_tol::Float64
+    g_tol::Float64
+    iterations::Int
+    nseeds::Int
+    seeds::Array{Float64}
+
+    function RandomSeedsLBFGS(; f_tol::Float64 = 1e-6, g_tol::Float64 = 1e-6, iterations::Int = 10^5, nseeds::Int = 3)
+        return new(f_tol, g_tol, 1e5, nseeds)
+    end
+end
+
+
+"""
+    StateSpaceDimensions
+
+StateSpaceModel dimensions, following the notation of on the book
+\"Time Series Analysis by State Space Methods\" (2012) by J. Durbin and S. J. Koopman.
+
+* `n` is the number of observations
+* `p` is the dimension of the observation vector ``y_t``
+* `m` is the dimension of the state vector ``\\alpha_t``
+* `r` is the dimension of the state covariance matrix ``Q_t``
+"""
+struct StateSpaceDimensions
+    n::Int
+    p::Int
+    m::Int
+    r::Int
+end
+
+"""
+    StateSpaceModel
+
+Following the notation of on the book \"Time Series Analysis by State Space Methods\" (2012) by J. Durbin and S. J. Koopman.
+
+* `y` A ``n \\times p`` matrix containing observations
+* `Z` A ``p \\times m \\times n`` matrix
+* `T` A ``m \\times m`` matrix
+* `R` A ``m \\times r`` matrix
+
+A `StateSpaceModel` object can be defined using `StateSpaceModel(y::Matrix{Float64}, Z::Array{Float64, 3}, T::Matrix{Float64}, R::Matrix{Float64})`.
+
[…]
