### 2019-07-27 6bb1ec9
small fix in docs
