### 2019-07-29 e5ac19c
faster kalman filter
