+++ src/kalman.jl
@@ -8,6 +8,8 @@ function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; t
     # Load dimensions
     n, p, m, r = size(model)
 
+    time_invariant = model.mode == "time-invariant"
+
     # Load system
     y = model.y
     Z, T, R = ztr(model)
@@ -15,6 +17,8 @@ function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; t
     # Predictive state and its covariance
     a = Matrix{Float64}(undef, n+1, m)
     P = Array{Float64, 3}(undef, m, m, n+1)
+    att = Matrix{Float64}(undef, n, m)
+    Ptt = Array{Float64, 3}(undef, m, m, n)
 
     # Innovation and its covariance
     v = Matrix{Float64}(undef, n, p)
@@ -22,7 +26,7 @@ function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; t
 
     # Kalman gain
     K = Array{Float64, 3}(undef, m, p, n)
-
+    
     # Steady state initialization
     steadystate = false
     tsteady     = n+1
@@ -31,73 +35,43 @@ function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; t
     a[1, :]    = zeros(m, 1)
     P[:, :, 1] = 1e6 .* Matrix(I, m, m)
 
+    RQR = R * Q * R'
     # Kalman filter
     for t = 1:n
-        # Check for missing values
-        if any(isnan.(y[t, :]))
+        if check_missing_observation(y ,t)
             steadystate  = false
-            v[t, :]      = NaN*ones(p)
-            F[:, :, t]   = Z[:, :, t]*P[:, :, t]*Z[:, :, t]' + H
-            K[:, :, t]   = T * P[:, :, t] * Z[:, :, t]' * inv(F[:, :, t])
-            a[t+1, :]    = T*a[t, :]
-            P[:, :, t+1] = ensure_pos_sym(T*P[:, :, t]*T' + R*Q*R')
+            v[t, :]      = fill(NaN, p)
+            F[:, :, t]   = fill(NaN, (p, p))
+            att[t, :]    = a[t, :]
+            Ptt[:, :, t] = P[:, :, t]
+            update_a(a, att, T, t) # a_t+1 = T * att_t
+            P[:, :, t+1] = ensure_pos_sym(T * Ptt[:, :, t] * T' + RQR)
+        elseif steadystate
+            v[t, :]      = y[t, :] - Z[:, :, t] * a[t, :] # v_t = y_t - Z_t * a_t
+            repeat_matrix_t_plus_1(F, t-1) # F[:, :, t]   = F[:, :, t-1]
+            repeat_matrix_t_plus_1(K, t-1) # K[:, :, t]   = K[:, :, t-1]
+            att[t, :]    = a[t, :] + P[:, :, t] * Z[:, :, t]' * inv(F[:, :, t]) * v[t, :] # att_t = a_t + P_t * Z_t' * F^-1_t * v_t
+            repeat_matrix_t_plus_1(Ptt, t-1) # Ptt_t = Ptt_t-1
+            update_a(a, att, T, t) # a_t+1 = T * att_t
+            repeat_matrix_t_plus_1(P, t) # P_t+1 = P_t
         else
-            v[t, :] = y[t, :] - Z[:, :, t] * a[t, :]
-            if steadystate
-                F[:, :, t]   = F[:, :, t-1]
-                K[:, :, t]   = K[:, :, t-1]
-                a[t+1, :]    = T * a[t, :] + K[:, :, t] * v[t, :]
[…]
