+++ src/kalman.jl
@@ -5,66 +5,66 @@ Kalman filter with big Kappa initialization.
 """
 function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; tol::Typ = 1e-5) where Typ <: AbstractFloat
 
-    # Load dimensions
-    n, p, m, r = size(model)
-
     time_invariant = model.mode == "time-invariant"
 
     # Load system
-    y = model.y
     Z, T, R = ztr(model)
 
     # Predictive state and its covariance
-    a = Matrix{Float64}(undef, n+1, m)
-    P = Array{Float64, 3}(undef, m, m, n+1)
-    att = Matrix{Float64}(undef, n, m)
-    Ptt = Array{Float64, 3}(undef, m, m, n)
+    a = Matrix{Float64}(undef, model.dim.n+1, model.dim.m)
+    P = Array{Float64, 3}(undef, model.dim.m, model.dim.m, model.dim.n+1)
+    att = Matrix{Float64}(undef, model.dim.n, model.dim.m)
+    Ptt = Array{Float64, 3}(undef, model.dim.m, model.dim.m, model.dim.n)
 
     # Innovation and its covariance
-    v = Matrix{Float64}(undef, n, p)
-    F = Array{Float64, 3}(undef, p, p, n)
+    v = Matrix{Float64}(undef, model.dim.n, model.dim.p)
+    F = Array{Float64, 3}(undef, model.dim.p, model.dim.p, model.dim.n)
 
     # Kalman gain
-    K = Array{Float64, 3}(undef, m, p, n)
+    K = Array{Float64, 3}(undef, model.dim.m, model.dim.p, model.dim.n)
+    
+    # Auxiliary structures
+    ZP = Array{Float64, 2}(undef, model.dim.p, model.dim.m)
+    P_Ztransp_invF = Array{Float64, 2}(undef, model.dim.m, model.dim.p)
     
     # Steady state initialization
     steadystate = false
-    tsteady     = n+1
+    tsteady     = model.dim.n+1
 
     # Initial state: big Kappa initialization
-    a[1, :]    = zeros(m, 1)
-    P[:, :, 1] = 1e6 .* Matrix(I, m, m)
+    a[1, :]    = zeros(model.dim.m, 1)
+    P[:, :, 1] = 1e6 .* Matrix(I, model.dim.m, model.dim.m)
 
     RQR = R * Q * R'
     # Kalman filter
-    for t = 1:n
-        if check_missing_observation(y ,t)
+    for t = 1:model.dim.n
+        if check_missing_observation(model.y, t)
             steadystate  = false
-            v[t, :]      = fill(NaN, p)
-            F[:, :, t]   = fill(NaN, (p, p))
+            v[t, :]      = fill(NaN, model.dim.p)
+            F[:, :, t]   = fill(NaN, (model.dim.p, model.dim.p))
             att[t, :]    = a[t, :]
             Ptt[:, :, t] = P[:, :, t]
             update_a(a, att, T, t) # a_t+1 = T * att_t
             P[:, :, t+1] = ensure_pos_sym(T * Ptt[:, :, t] * T' + RQR)
         elseif steadystate
-            v[t, :]      = y[t, :] - Z[:, :, t] * a[t, :] # v_t = y_t - Z_t * a_t
+            v[t, :]      = model.y[t, :] - Z[:, :, t] * a[t, :] # v_t = y_t - Z_t * a_t
[…]
