### 2019-07-30 47d9f89
some function updates
