+++ src/kalman.jl
@@ -45,9 +45,9 @@ function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; t
             att[t, :]    = a[t, :]
             Ptt[:, :, t] = P[:, :, t]
             update_a(a, att, T, t) # a_t+1 = T * att_t
-            P[:, :, t+1] = ensure_pos_sym(T * Ptt[:, :, t] * T' + RQR)
+            update_P(P, T, Ptt, RQR, t)
         elseif steadystate
-            v[t, :]      = model.y[t, :] - Z[:, :, t] * a[t, :] # v_t = y_t - Z_t * a_t
+            update_v(v, model.y, Z, a, t) # v_t = y_t - Z_t * a_t
             repeat_matrix_t_plus_1(F, t-1) # F[:, :, t]   = F[:, :, t-1]
             repeat_matrix_t_plus_1(K, t-1) # K[:, :, t]   = K[:, :, t-1]
             att[t, :]    = a[t, :] + P[:, :, t] * Z[:, :, t]' * invF(F, t) * v[t, :] # att_t = a_t + P_t * Z_t' * F^-1_t * v_t
@@ -55,15 +55,15 @@ function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; t
             update_a(a, att, T, t) # a_t+1 = T * att_t
             repeat_matrix_t_plus_1(P, t) # P_t+1 = P_t
         else
-            v[t, :]      = model.y[t, :] - Z[:, :, t] * a[t, :] # v_t = y_t - Z_t * a_t
+            update_v(v, model.y, Z, a, t) # v_t = y_t - Z_t * a_t
             update_ZP(ZP, Z, P, t) # ZP = Z[:, :, t] * P[:, :, t]
-            F[:, :, t]   = ZP * Z[:, :, t]' + H # F_t = Z_t * P_t * Z_t + H
-            P_Ztransp_invF   = ZP' * invF(F, t)
+            update_F(F, ZP, Z, H, t) # F_t = Z_t * P_t * Z_t + H
+            update_P_Ztransp_Finv(P_Ztransp_invF, ZP, F, t) # P_Ztransp_invF   = ZP' * invF(F, t)
             update_K(K, P_Ztransp_invF, T, t) # K_t = T * P_t * Z_t * F^-1_t
-            att[t, :]    = a[t, :] + P_Ztransp_invF * v[t, :] # att_t = a_t + P_t * Z_t * F^-1_t * v_t
+            update_att(att, a, P_Ztransp_invF, v, t) # att_t = a_t + P_t * Z_t * F^-1_t * v_t
             Ptt[:, :, t] = P[:, :, t] - P_Ztransp_invF * ZP # Ptt_t = P_t - P_t * Z_t * F^-1_t * P_t * Z_t
             update_a(a, att, T, t) # a_t+1 = T * att_t
-            P[:, :, t+1] = ensure_pos_sym(T * Ptt[:, :, t] * T' + RQR) # P_t+1 = T * Ptt_t * T' + RQR'
+            update_P(P, T, Ptt, RQR, t) # P_t+1 = T * Ptt_t * T' + RQR'
             if check_steady_state(P, t, tol) && time_invariant
                 steadystate = true
                 tsteady     = t
@@ -191,7 +191,7 @@ end
 function repeat_matrix_t_plus_1(mat::AbstractArray{T}, t::Int) where T <: AbstractFloat
     @inbounds for j in axes(mat, 2), i in axes(mat, 1)
           mat[i, j, t+1] = mat[i, j, t]
[…]
