### 2019-07-31 743b3ab
change example directory
