+++ src/StateSpaceModels.jl
@@ -12,6 +12,7 @@ include("utils.jl")
 include("models.jl")
 include("estimation.jl")
 include("random_seeds_lbfgs.jl")
+include("kalman_utils.jl")
 include("kalman.jl")
 include("sqrt_kalman.jl")
 include("forecast.jl")
