+++ src/kalman.jl
@@ -46,7 +46,7 @@ function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; t
             update_att(att, a, t) # att_t = a_t
             update_Ptt(Ptt, P, t) # Ptt_t = P_t
             update_a(a, att, T, t) # a_t+1 = T * att_t
-            update_P(P, T, Ptt, RQR, t)
+            update_P(P, T, Ptt, RQR, t) # P_t+1 = T * Ptt_t * T' + RQR'
         elseif steadystate
             update_v(v, model.y, Z, a, t) # v_t = y_t - Z_t * a_t
             repeat_matrix_t_plus_1(F, t-1) # F[:, :, t]   = F[:, :, t-1]
@@ -186,89 +186,4 @@ function kalman_filter_and_smoother(model::StateSpaceModel, covariance::StateSpa
                         filtered_state.P, filtered_state.Ptt, filtered_state.F,
                         filtered_state.steadystate, filtered_state.tsteady),
            SmoothedState(smoothed_state.alpha, smoothed_state.V) 
-end
-
-# Utils for filter performance
-function repeat_matrix_t_plus_1(mat::AbstractArray{T}, t::Int) where T <: AbstractFloat
-    @inbounds for j in axes(mat, 2), i in axes(mat, 1)
-          mat[i, j, t+1] = mat[i, j, t]
-    end
-    return 
-end
-
-function update_a(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
-    @views @inbounds mul!(a[t+1, :], T, att[t, :])
-    return 
-end
-
-function update_ZP(ZP::AbstractArray{T}, Z::AbstractArray{T}, P::AbstractArray{T}, t::Int) where T <: AbstractFloat
-    @inbounds @views mul!(ZP, Z[:, :, t], P[:, :, t])
-    return 
-end
-
-function update_K(K::AbstractArray{Typ}, P_Ztransp_invF::AbstractArray{Typ}, T::AbstractArray{Typ}, t::Int) where Typ <: AbstractFloat
-    @inbounds @views mul!(K[:, :, t], T, P_Ztransp_invF)
-    return
-end
-
-function invF(F::AbstractArray{T}, t::Int) where T
-    return @inbounds @views inv(F[:, :, t]) 
-end
-
-function update_P(P::AbstractArray{Typ}, T::AbstractArray{Typ}, Ptt::AbstractArray{Typ}, RQR::AbstractArray{Typ}, t::Int)  where Typ <: AbstractFloat 
-    @views @inbounds LinearAlgebra.BLAS.gemm!('N', 'T', 1.0, T * Ptt[:, :, t], T, 0.0, P[:, :, t+1]) # P[:, :, t+1] = T * Ptt[:, :, t] * T'
-    sum_matrix(P, RQR, t, 1) # P[:, :, t+1] .+= RQR
-    ensure_pos_sym(P, t + 1)
-    return 
-end
-
-function update_v(v::AbstractArray{T}, y::AbstractArray{T}, Z::AbstractArray{T}, a::AbstractArray{T}, t::Int) where T <: AbstractFloat
-    @views @inbounds v[t, :] = y[t, :] - Z[:, :, t] * a[t, :]
-    return 
-end
-
-function update_att(att::AbstractArray{T}, a::AbstractArray{T}, P_Ztransp_invF::AbstractArray{T}, v::AbstractArray{T}, t::Int) where T <: AbstractFloat
[…]
