+++ src/utils.jl
@@ -55,10 +55,6 @@ end
 
 Ensure that matrix `M` is positive and symmetric to avoid numerical errors when numbers are small by doing `(M + M')/2 + ϵ*I`
 """
-function ensure_pos_sym(M::Matrix{T}; ϵ::T = 1e-8) where T <: AbstractFloat
-    return (M + M')/2 + ϵ*I
-end
-
 function ensure_pos_sym(M::AbstractArray{T}, t::Int; ϵ::T = 1e-8) where T <: AbstractFloat
     @inbounds for j in axes(M, 2), i in 1:j
         if i == j

[…]
