+++ src/kalman.jl
@@ -7,9 +7,6 @@ function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; t
 
     time_invariant = model.mode == "time-invariant"
 
-    # Load system
-    Z, T, R = ztr(model)
-
     # Predictive state and its covariance
     a = Matrix{Float64}(undef, model.dim.n+1, model.dim.m)
     P = Array{Float64, 3}(undef, model.dim.m, model.dim.m, model.dim.n+1)
@@ -36,7 +33,7 @@ function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; t
     a[1, :]    = zeros(model.dim.m, 1)
     P[:, :, 1] = 1e6 .* Matrix(I, model.dim.m, model.dim.m)
 
-    RQR = R * Q * R'
+    RQR = model.R * Q * model.R'
     # Kalman filter
     for t = 1:model.dim.n
         if check_missing_observation(model.y, t)
@@ -45,26 +42,26 @@ function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; t
             F[:, :, t]   = fill(NaN, (model.dim.p, model.dim.p))
             update_att(att, a, t) # att_t = a_t
             update_Ptt(Ptt, P, t) # Ptt_t = P_t
-            update_a(a, att, T, t) # a_t+1 = T * att_t
-            update_P(P, T, Ptt, RQR, t) # P_t+1 = T * Ptt_t * T' + RQR'
+            update_a(a, att, model.T, t) # a_t+1 = T * att_t
+            update_P(P, model.T, Ptt, RQR, t) # P_t+1 = T * Ptt_t * T' + RQR'
         elseif steadystate
-            update_v(v, model.y, Z, a, t) # v_t = y_t - Z_t * a_t
+            update_v(v, model.y, model.Z, a, t) # v_t = y_t - Z_t * a_t
             repeat_matrix_t_plus_1(F, t-1) # F[:, :, t]   = F[:, :, t-1]
             repeat_matrix_t_plus_1(K, t-1) # K[:, :, t]   = K[:, :, t-1]
-            update_att(att, a, P, Z, F, v, t) # att_t = a_t + P_t * Z_t' * F^-1_t * v_t
+            update_att(att, a, P, model.Z, F, v, t) # att_t = a_t + P_t * Z_t' * F^-1_t * v_t
             repeat_matrix_t_plus_1(Ptt, t-1) # Ptt_t = Ptt_t-1
-            update_a(a, att, T, t) # a_t+1 = T * att_t
+            update_a(a, att, model.T, t) # a_t+1 = T * att_t
             repeat_matrix_t_plus_1(P, t) # P_t+1 = P_t
         else
-            update_v(v, model.y, Z, a, t) # v_t = y_t - Z_t * a_t
-            update_ZP(ZP, Z, P, t) # ZP = Z[:, :, t] * P[:, :, t]
-            update_F(F, ZP, Z, H, t) # F_t = Z_t * P_t * Z_t + H
+            update_v(v, model.y, model.Z, a, t) # v_t = y_t - Z_t * a_t
+            update_ZP(ZP, model.Z, P, t) # ZP = Z[:, :, t] * P[:, :, t]
+            update_F(F, ZP, model.Z, H, t) # F_t = Z_t * P_t * Z_t + H
             update_P_Ztransp_Finv(P_Ztransp_invF, ZP, F, t) # P_Ztransp_invF   = ZP' * invF(F, t)
-            update_K(K, P_Ztransp_invF, T, t) # K_t = T * P_t * Z_t * F^-1_t
[…]
