### 2019-08-01 7750e2d
update ensure_pos_sym name
