+++ src/sqrt_kalman.jl
@@ -104,7 +104,7 @@ function filtered_state(model::StateSpaceModel, sqrt_filter::SquareRootFilter)
         PZF = P[:, :, t] * Z[:, :, t]' * inv(F[:, :, t])
         att[t, :]    = a[t, :] + PZF * v[t, :]
         Ptt[:, :, t] = P[:, :, t] - PZF * Z[:, :, t] * P[:, :, t]
-        ensure_pos_sym(Ptt, t)
+        ensure_pos_sym!(Ptt, t)
     end
 
     return att, Ptt
