+++ docs/src/reference.md
@@ -5,5 +5,5 @@
 size
 ztr
 check_steady_state
-ensure_pos_sym
+ensure_pos_sym!
 ```
