### 2019-08-01 a50b772
add sum_matrix!
