### 2019-08-01 4862a92
Kalman refact (#84)
* faster kalman filter

* updates

* some function updates

* optimized kalman filter

* change example directory

* take out line

* fixing requests

* remove ztr

* update ensure_pos_sym name

* add sum_matrix!

* more updates
