+++ src/kalman_utils.jl
@@ -1,0 +1,84 @@
+# Utils for filter performance
+function repeat_matrix_t_plus_1(mat::AbstractArray{T}, t::Int) where T <: AbstractFloat
+    @inbounds for j in axes(mat, 2), i in axes(mat, 1)
+          mat[i, j, t+1] = mat[i, j, t]
+    end
+    return 
+end
+
+function update_a(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
+    @views @inbounds mul!(a[t+1, :], T, att[t, :])
+    return 
+end
+
+function update_ZP(ZP::AbstractArray{T}, Z::AbstractArray{T}, P::AbstractArray{T}, t::Int) where T <: AbstractFloat
+    @inbounds @views mul!(ZP, Z[:, :, t], P[:, :, t])
+    return 
+end
+
+function update_K(K::AbstractArray{Typ}, P_Ztransp_invF::AbstractArray{Typ}, T::AbstractArray{Typ}, t::Int) where Typ <: AbstractFloat
+    @inbounds @views mul!(K[:, :, t], T, P_Ztransp_invF)
+    return
+end
+
+function invF(F::AbstractArray{T}, t::Int) where T
+    return @inbounds @views invertF(F[:, :, t]) 
+end
+
+function update_P(P::AbstractArray{Typ}, T::AbstractArray{Typ}, Ptt::AbstractArray{Typ}, RQR::AbstractArray{Typ}, t::Int)  where Typ <: AbstractFloat 
+    @views @inbounds LinearAlgebra.BLAS.gemm!('N', 'T', 1.0, T * Ptt[:, :, t], T, 0.0, P[:, :, t+1]) # P[:, :, t+1] = T * Ptt[:, :, t] * T'
+    sum_matrix!(P, RQR, t, 1) # P[:, :, t+1] .+= RQR
+    ensure_pos_sym!(P, t + 1)
+    return 
+end
+
+function update_v(v::AbstractArray{T}, y::AbstractArray{T}, Z::AbstractArray{T}, a::AbstractArray{T}, t::Int) where T <: AbstractFloat
+    @views @inbounds v[t, :] = y[t, :] - Z[:, :, t] * a[t, :]
+    return 
+end
+
+function update_att(att::AbstractArray{T}, a::AbstractArray{T}, P_Ztransp_invF::AbstractArray{T}, v::AbstractArray{T}, t::Int) where T <: AbstractFloat
+    @inbounds @views att[t, :] = a[t, :] + P_Ztransp_invF * v[t, :]
+    return 
+end
+
+function update_att(att::AbstractArray{T}, a::AbstractArray{T}, P::AbstractArray{T}, Z::AbstractArray{T},
+                    F::AbstractArray{T}, v::AbstractArray{T}, t::Int) where T <: AbstractFloat
+
+    @views @inbounds att[t, :] = a[t, :] + 
+        LinearAlgebra.BLAS.gemm('N', 'T', 1.0, P[:, :, t], Z[:, :, t]) *
+        invF(F, t) * v[t, :]
+    return
+end
+
+function update_att(att::AbstractArray{T}, a::AbstractArray{T}, t::Int) where T <: AbstractFloat
+    @inbounds for j in axes(att, 2)
+        att[t, j] = a[t, j]
+    end
+    return
+end
+
+function update_Ptt(Ptt::AbstractArray{T}, P::AbstractArray{T}, t::Int) where T <: AbstractFloat
+    @inbounds for i in axes(Ptt, 1), j in axes(Ptt, 2)
[…]
