### 2019-08-02 5884626
fix missing values performance regression
