+++ src/kalman.jl
@@ -33,10 +33,12 @@ function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; t
     a[1, :]    = zeros(model.dim.m, 1)
     P[:, :, 1] = 1e6 .* Matrix(I, model.dim.m, model.dim.m)
 
+    missing_obs = check_missing_observation(model.y)
+
     RQR = model.R * Q * model.R'
     # Kalman filter
     for t = 1:model.dim.n
-        if check_missing_observation(model.y, t)
+        if t in missing_obs
             steadystate  = false
             v[t, :]      = fill(NaN, model.dim.p)
             F[:, :, t]   = fill(NaN, (model.dim.p, model.dim.p))
@@ -103,8 +105,10 @@ function smoother(model::StateSpaceModel, kfilter::KalmanFilter)
     N[:, :, end] = zeros(m, m)
     r[end, :]    = zeros(m, 1)
 
+    missing_obs = check_missing_observation(v)
+
     @views @inbounds for t = n:-1:2
-        if check_missing_observation(v, t)
+        if t in missing_obs
             r[t-1, :]     = T' * r[t, :]
             N[:, :, t-1]  = T' * N[:, :, t] * T
         else
@@ -118,7 +122,7 @@ function smoother(model::StateSpaceModel, kfilter::KalmanFilter)
     end
 
     # Last iteration
-    if check_missing_observation(v, 1)
+    if 1 in missing_obs
         r0  = T' * r[1, :]
         N0  = T' * N[:, :, 1] * T
     else
