### 2019-08-05 70903de
add benchmark and move missing_obsrvations inside the model
