+++ src/structures.jl
@@ -103,6 +103,7 @@ struct StateSpaceModel
     T::Matrix{Float64} # state matrix
     R::Matrix{Float64} # state error matrix
     dim::StateSpaceDimensions
+    missing_observations::Vector{Int}
     mode::String
 
     function StateSpaceModel(y::Matrix{Float64}, Z::Array{Float64, 3}, T::Matrix{Float64}, R::Matrix{Float64})
@@ -116,7 +117,7 @@ struct StateSpaceModel
             error("StateSpaceModel dimension mismatch")
         end
         dim = StateSpaceDimensions(ny, py, mr, rr)
-        new(y, Z, T, R, dim, "time-variant")
+        new(y, Z, T, R, dim, find_missing_observations(y), "time-variant")
     end
 
     function StateSpaceModel(y::Matrix{Float64}, Z::Matrix{Float64}, T::Matrix{Float64}, R::Matrix{Float64})
@@ -133,10 +134,10 @@ struct StateSpaceModel
 
         # Build Z
         Zvar = Array{Float64, 3}(undef, pz, mz, ny)
-        for t = 1:ny
-            Zvar[:, :, t] = Z
+        for t in 1:ny, i in axes(Z, 1), j in axes(Z, 2)
+            Zvar[i, j, t] = Z[i, j] # Zvar[:, :, t] = Z
         end
-        new(y, Zvar, T, R, dim, "time-invariant")
+        new(y, Zvar, T, R, dim, find_missing_observations(y), "time-invariant")
     end
 end
 
