### 2019-08-05 #28 @raphaelsaavedra
@matbesancon are we allowed to upload a preprint of the juliacon paper on arxiv using a different template?
