### 2019-08-05 f56b61a
fix missing values performance regression (#85)
* fix missing values performance regression

* add benchmark and move missing_obsrvations inside the model
