+++ src/sqrt_kalman.jl
@@ -42,9 +42,10 @@ function sqrt_kalman_filter(model::StateSpaceModel, sqrtH::Matrix{Typ}, sqrtQ::M
     sqrtH_zeros_pr  = [sqrtH zeros_pr]
     zeros_mp_RsqrtQ = [zeros_mp R*sqrtQ]
 
+    !isempty(model.missing_observations) && error("Treatment of missing values not implemented for SquareRootFilter, please use KalmanFilter.")
+
     # Square-root Kalman filter
     for t = 1:n
-        any(isnan.(y[t, :])) && error("Treatment of missing values not implemented for SquareRootFilter, please use KalmanFilter.")
         v[t, :] = y[t, :] - Z[:, :, t]*a[t, :]
         if steadystate
             sqrtF[:, :, t] = sqrtF[:, :, t-1]
