### 2019-08-05 ded6ec9
fix check steady state for sqrt filter (#86)
