### 2019-08-05 b595787
minor performance update
