### 2019-08-06 0943425
minor performance update (#87)
