### 2019-08-06 471b30e
minor doc improvements
