### 2019-08-06 96b2e8e
update_v and update_att
