### 2019-08-06 e56fd27
adding filtered state to docstrings
