### 2019-08-06 6ee3e8d
update_v and update_att (#89)
