### 2019-08-07 c8361cd
update statespace_likelihood function
