### 2019-08-08 #28 @mschauer
Nice to meet you. So I should start stating that I share some interest in this topic, see https://github.com/mschauer/Kalman.jl This should be manageable according to PNAS criteria. Actually, my package is somewhat opinionated and small, so you would *expect* a different package to be the go-to package for this topic.
