+++ src/StateSpaceModels.jl
@@ -4,7 +4,7 @@ using Optim, Distributions, LinearAlgebra, StaticArrays, Dates, Printf
 
 import Base: size, show
 
-export statespace, kalman_filter_and_smoother
+export statespace, kfas
 
 include("prints.jl")
 include("structures.jl")
@@ -36,7 +36,7 @@ function statespace(model::StateSpaceModel; filter_type::DataType = KalmanFilter
     covariance = estimate_statespace(model, filter_type, optimization_method; verbose = verbose)
 
     # Kalman filter and smoothing
-    filter_output, smoothed_state = kalman_filter_and_smoother(model, covariance, filter_type)
+    filter_output, smoothed_state = kfas(model, covariance, filter_type)
 
     print_bottom(verbose)
 
@@ -44,15 +44,15 @@ function statespace(model::StateSpaceModel; filter_type::DataType = KalmanFilter
 end
 
 """
-    kalman_filter_and_smoother(model::StateSpaceModel, covariance::StateSpaceCovariance,
-                               filter_type::DataType)
+    kfas(model::StateSpaceModel, covariance::StateSpaceCovariance,
+         filter_type::DataType)
 
 Perform kalman filter and smoother according to the chosen `filter_type`.
 """
-function kalman_filter_and_smoother(model::StateSpaceModel, covariance::StateSpaceCovariance,
-                                    filter_type::DataType)
+function kfas(model::StateSpaceModel, covariance::StateSpaceCovariance,
+              filter_type::DataType)
     error(filter_type , " not implemented") # Returns an error if it cannot
-                                            # find a specialized kalman_filter_and_smoother
+                                            # find a specialized kfas
 end
 
 end # end module
