### 2019-08-08 70e7be9
Some docs updates (#90)
* update

* update statespace_likelihood function

* fixed typos and kfas
