+++ docs/src/reference.md
@@ -1,9 +1,9 @@
 # Reference
 
-<!-- utils.jl -->
 ```@docs
 size
 ztr
-check_steady_state
-ensure_pos_sym!
+StateSpaceModels.check_steady_state
+StateSpaceModels.ensure_pos_sym!
+StateSpaceModels.prepare_forecast
 ```
