### 2019-08-08 #28 @raphaelsaavedra
@mschauer thanks for taking your time! I should warn you that we have just released a significant performance update in v0.2.3. From what I've seen in the checklist, you have already installed the package, so you might want to update it :)
