### 2019-08-08 f94d4d0
updated filter recursion
