### 2019-08-09 039fd1a
added smoother and logdetF
