### 2019-08-09 4e19446
a small performance update
