### 2019-08-09 18b8a51
performance update of invertF
