+++ src/kalman_utils.jl
@@ -1,11 +1,11 @@
 # Utils for filter performance
-function fill_a1(a::Matrix{T}) where T <: AbstractFloat
+function fill_a1!(a::Matrix{T}) where T <: AbstractFloat
     @inbounds for i in axes(a, 2)
         a[1, i] = zero(T)
     end
 end
 
-function fill_P1(P::AbstractArray{T}; bigkappa::Float64 = 1e6) where T <: AbstractFloat
+function fill_P1!(P::AbstractArray{T}; bigkappa::Float64 = 1e6) where T <: AbstractFloat
     @inbounds for i in axes(P, 1), j in axes(P, 2)
         if i == j
             P[i, j, 1] = bigkappa
@@ -15,7 +15,7 @@ function fill_P1(P::AbstractArray{T}; bigkappa::Float64 = 1e6) where T <: Abstra
     end
 end
 
-function repeat_matrix_t_plus_1(mat::AbstractArray{T}, t::Int) where T <: AbstractFloat
+function repeat_matrix_t_plus_1!(mat::AbstractArray{T}, t::Int) where T <: AbstractFloat
     @inbounds for j in axes(mat, 2), i in axes(mat, 1)
           mat[i, j, t+1] = mat[i, j, t]
     end
@@ -37,7 +37,7 @@ function small_update_a(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int
     return 
 end
 
-function update_a(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
+function update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
     # Here there is a trade-off memory-speed, usually if the dimension of a is smaller than 15 
     # it is more performant to do the hard coded version (small_update_a)
     if size(a, 2) < 16
@@ -48,12 +48,12 @@ function update_a(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) wher
     return 
 end
 
-function update_ZP(ZP::AbstractArray{T}, Z::AbstractArray{T}, P::AbstractArray{T}, t::Int) where T <: AbstractFloat
+function update_ZP!(ZP::AbstractArray{T}, Z::AbstractArray{T}, P::AbstractArray{T}, t::Int) where T <: AbstractFloat
     @inbounds @views mul!(ZP, Z[:, :, t], P[:, :, t])
     return 
 end
 
-function update_K(K::AbstractArray{Typ}, P_Ztransp_invF::AbstractArray{Typ}, T::AbstractArray{Typ}, t::Int) where Typ <: AbstractFloat
+function update_K!(K::AbstractArray{Typ}, P_Ztransp_invF::AbstractArray{Typ}, T::AbstractArray{Typ}, t::Int) where Typ <: AbstractFloat
     @inbounds @views mul!(K[:, :, t], T, P_Ztransp_invF)
     return
 end
@@ -62,14 +62,14 @@ function invF(F::AbstractArray{T}, t::Int) where T
     return @inbounds @views invertF(F, t) 
 end
 
-function update_P(P::AbstractArray{Typ}, T::AbstractArray{Typ}, Ptt::AbstractArray{Typ}, RQR::AbstractArray{Typ}, t::Int)  where Typ <: AbstractFloat 
[…]
