+++ docs/src/manual.md
@@ -7,16 +7,16 @@ In this package we consider the following state-space model
 ```math
 \begin{gather*}
     \begin{aligned}
-        y_t &= Z_t \alpha_t  + \varepsilon_t, \quad \quad \quad t = 1 \dots n, \\
-        \alpha_{t+1} &= T \alpha_t + R \eta_t,
+        y_{t} &= Z_{t} \alpha_{t}  + \varepsilon_{t}, \quad \quad \quad t = 1 \dots n, \\
+        \alpha_{t+1} &= T \alpha_{t} + R \eta_{t},
     \end{aligned}
 \end{gather*}
 ```
 where
 ```math
 \begin{bmatrix}
-    \varepsilon_t \\
-    \eta_t \\
+    \varepsilon_{t} \\
+    \eta_{t} \\
     \alpha_1
 \end{bmatrix}
 \sim
@@ -53,8 +53,8 @@ The local level model is defined by
 ```math
 \begin{gather*}
     \begin{aligned}
-        y_t &=  \mu_t  + \varepsilon_t \quad \varepsilon_t \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\
-        \mu_{t+1} &= \mu_t + \eta_t \quad \eta_t \sim \mathcal{N}(0, \sigma^2_{\eta})\\
+        y_{t} &=  \mu_{t}  + \varepsilon_{t} \quad \varepsilon_{t} \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\
+        \mu_{t+1} &= \mu_{t} + \eta_{t} \quad \eta_{t} \sim \mathcal{N}(0, \sigma^2_{\eta})\\
     \end{aligned}
 \end{gather*}
 ```
@@ -68,9 +68,9 @@ The linear trend model is defined by
 ```math
 \begin{gather*}
     \begin{aligned}
-        y_t &=  \mu_t  + \varepsilon_t \quad &\varepsilon_t \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\
-        \mu_{t+1} &= \mu_t + \nu_t + \xi_t \quad &\xi_t \sim \mathcal{N}(0, \sigma^2_{\xi})\\
-        \nu_{t+1} &= \nu_t + \zeta_t \quad &\zeta_t \sim \mathcal{N}(0, \sigma^2_{\zeta})\\
+        y_{t} &=  \mu_{t}  + \varepsilon_{t} \quad &\varepsilon_{t} \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\
+        \mu_{t+1} &= \mu_{t} + \nu_{t} + \xi_{t} \quad &\xi_{t} \sim \mathcal{N}(0, \sigma^2_{\xi})\\
+        \nu_{t+1} &= \nu_{t} + \zeta_{t} \quad &\zeta_{t} \sim \mathcal{N}(0, \sigma^2_{\zeta})\\
     \end{aligned}
 \end{gather*}
 ```
@@ -118,13 +118,13 @@ The implementation of the Kalman Filter follows the recursion
 ```math
 \begin{gather*}
     \begin{aligned}
-        v_t &= y_t - Z_t a_t,  &F_t &= Z_t P_t Z^{\top}_t + H\\
-        a_{t|t} &= a_t - P_t Z^{\top}_t F_t^{-1} v_t, \quad \quad \quad &P_{t|t} &= P_t -  P_t Z^{\top}_t F_t^{-1}Z_t P_t\\
-        a_{t+1} &= T a_t - K_t v_t, \quad \quad \quad &P_{t+1} &= T P_t(T - K_t Z_t)^{\top} + R Q R\\
[…]
