+++ benchmark/bench.jl
@@ -19,31 +19,31 @@ bench_local_level = @benchmark statespace($locallevel, filter_type = $SquareRoot
 #   evals/sample:     1
 
 bench_local_level = @benchmark statespace($locallevel, filter_type = $KalmanFilter, verbose = $verbose)
-# 9th August 2019
+# 11th August 2019
 # BenchmarkTools.Trial: 
-#   memory estimate:  44.20 MiB
-#   allocs estimate:  564867
+#   memory estimate:  10.62 MiB
+#   allocs estimate:  92728
 #   --------------
-#   minimum time:     34.680 ms (11.09% GC)
-#   median time:      53.984 ms (10.75% GC)
-#   mean time:        53.772 ms (10.87% GC)
-#   maximum time:     69.858 ms (11.15% GC)
+#   minimum time:     11.002 ms (0.00% GC)
+#   median time:      19.476 ms (7.28% GC)
+#   mean time:        20.071 ms (5.96% GC)
+#   maximum time:     45.660 ms (3.44% GC)
 #   --------------
-#   samples:          93
+#   samples:          250
 #   evals/sample:     1
 
 H = fill(1.0, (1,1))
 Q = fill(1.0, (1,1))
 @benchmark StateSpaceModels.kalman_filter($locallevel, $H, $Q)
-# 9th August 2019
+# 11th August 2019
 # BenchmarkTools.Trial: 
-#   memory estimate:  66.80 KiB
-#   allocs estimate:  829
+#   memory estimate:  12.17 KiB
+#   allocs estimate:  93
 #   --------------
-#   minimum time:     41.687 μs (0.00% GC)
-#   median time:      42.620 μs (0.00% GC)
-#   mean time:        51.369 μs (11.04% GC)
-#   maximum time:     3.085 ms (97.27% GC)
+#   minimum time:     11.164 μs (0.00% GC)
+#   median time:      11.530 μs (0.00% GC)
+#   mean time:        13.175 μs (5.00% GC)
+#   maximum time:     1.549 ms (98.61% GC)
 #   --------------
 #   samples:          10000
 #   evals/sample:     1
