### 2019-08-12 2a07634
better gram function
