+++ src/utils.jl
@@ -29,7 +29,13 @@ function gram_in_time(mat::Array{Float64, 3})
 end
 
 function gram(mat::AbstractArray{T}) where T <: AbstractFloat
-    return LinearAlgebra.BLAS.gemm('N', 'T', mat, mat) # mat*mat'    
+    if size(mat, 1) == 1
+        gram_mat = Matrix{T}(undef, 1, 1)
+        gram_mat[1, 1] = mat[1, 1]^2
+        return gram_mat
+    else
+        return LinearAlgebra.BLAS.gemm('N', 'T', mat, mat) # mat*mat'    
+    end
 end
 
 """
