### 2019-08-12 ab24e18
remove p as an argument in loglik function
