### 2019-08-12 fb4a948
Improving performance and docs (#94)
* updated filter recursion

* added smoother and logdetF

* a small performance update

* performance update of invertF

* fixing requests

* fix forgotten line and _{t}

* nice performance update

* better gram function

* remove p as an argument in loglik function
