+++ src/kalman_utils.jl
@@ -1,12 +1,12 @@
 # Utils for filter performance
-function fill_a1(a::Matrix{T}) where T <: AbstractFloat
-    for i in axes(a, 2)
+function fill_a1!(a::Matrix{T}) where T <: AbstractFloat
+    @inbounds for i in axes(a, 2)
         a[1, i] = zero(T)
     end
 end
 
-function fill_P1(P::AbstractArray{T}; bigkappa::Float64 = 1e6) where T <: AbstractFloat
-    for i in axes(P, 1), j in axes(P, 2)
+function fill_P1!(P::AbstractArray{T}; bigkappa::Float64 = 1e6) where T <: AbstractFloat
+    @inbounds for i in axes(P, 1), j in axes(P, 2)
         if i == j
             P[i, j, 1] = bigkappa
         else
@@ -15,42 +15,63 @@ function fill_P1(P::AbstractArray{T}; bigkappa::Float64 = 1e6) where T <: Abstra
     end
 end
 
-function repeat_matrix_t_plus_1(mat::AbstractArray{T}, t::Int) where T <: AbstractFloat
+function repeat_matrix_t_plus_1!(mat::AbstractArray{T}, t::Int) where T <: AbstractFloat
     @inbounds for j in axes(mat, 2), i in axes(mat, 1)
           mat[i, j, t+1] = mat[i, j, t]
     end
     return 
 end
 
-function update_a(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
+function big_update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
     @views @inbounds mul!(a[t+1, :], T, att[t, :])
     return 
 end
 
-function update_ZP(ZP::AbstractArray{T}, Z::AbstractArray{T}, P::AbstractArray{T}, t::Int) where T <: AbstractFloat
+function small_update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
+    @inbounds for i in axes(a, 2)
+        a[t+1, i] = zero(Typ)
+        for j in axes(T, 2)
+            a[t+1, i] += T[i, j]*att[t, j]
+        end
+    end
+    return 
+end
+
+function update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
+    # Here there is a trade-off memory-speed, usually if the dimension of a is smaller than 15 
+    # it is more performant to do the hard coded version (small_update_a)
+    if size(a, 2) < 16
+        small_update_a!(a, att, T, t)
+    else
+        big_update_a!(a, att, T, t)
+    end
+    return 
+end
+
+function update_ZP!(ZP::AbstractArray{T}, Z::AbstractArray{T}, P::AbstractArray{T}, t::Int) where T <: AbstractFloat
     @inbounds @views mul!(ZP, Z[:, :, t], P[:, :, t])
     return 
 end
 
-function update_K(K::AbstractArray{Typ}, P_Ztransp_invF::AbstractArray{Typ}, T::AbstractArray{Typ}, t::Int) where Typ <: AbstractFloat
+function update_K!(K::AbstractArray{Typ}, P_Ztransp_invF::AbstractArray{Typ}, T::AbstractArray{Typ}, t::Int) where Typ <: AbstractFloat
[…]
