### 2019-08-12 54b34f3
starting to bench structural models
