+++ benchmark/bench.jl
@@ -47,3 +47,22 @@ Q = fill(1.0, (1,1))
 #   --------------
 #   samples:          10000
 #   evals/sample:     1
+
+# Benchmark structural model
+AP = CSV.read("./examples/AirPassengers.csv")
+logAP = log.(Vector{Float64}(AP[:Passengers]))
+structural_model = structural(logAP, 12)
+
+bench_structural = @benchmark statespace($structural_model, filter_type = $KalmanFilter, verbose = $verbose)
+# 12th August 2019
+# BenchmarkTools.Trial: 
+#   memory estimate:  5.45 GiB
+#   allocs estimate:  11321246
+#   --------------
+#   minimum time:     4.361 s (8.57% GC)
+#   median time:      4.464 s (8.60% GC)
+#   mean time:        4.464 s (8.60% GC)
+#   maximum time:     4.567 s (8.62% GC)
+#   --------------
+#   samples:          2
+#   evals/sample:     1
