+++ src/univariate_kalman.jl
@@ -1,0 +1,179 @@
+"""
+    kalman_filter(model::StateSpaceModel, H::Typ, Q::Matrix{Typ}; tol::Float64 = 1e-5) where Typ <: AbstractFloat
+
+Kalman filter with big Kappa initialization, i.e., initializing state variances as 1e6.
+"""
+function univariate_kalman_filter(model::StateSpaceModel, H::Typ, Q::Matrix{Typ}; tol::Typ = 1e-5) where Typ <: AbstractFloat
+
+    time_invariant = model.mode == "time-invariant"
+
+    # Predictive state and its covariance
+    a = Matrix{Float64}(undef, model.dim.n+1, model.dim.m)
+    P = Array{Float64, 3}(undef, model.dim.m, model.dim.m, model.dim.n+1)
+    att = Matrix{Float64}(undef, model.dim.n, model.dim.m)
+    Ptt = Array{Float64, 3}(undef, model.dim.m, model.dim.m, model.dim.n)
+
+    # Innovation and its covariance
+    v = Vector{Float64}(undef, model.dim.n)
+    F = Vector{Float64}(undef, model.dim.n)
+
+    # Kalman gain
+    K = Matrix{Float64}(undef, model.dim.m, model.dim.n)
+    
+    # Auxiliary structures
+    ZP = Vector{Float64}(undef, model.dim.m)
+    P_Ztransp_invF = Vector{Float64}(undef, model.dim.m)
+    RQR = Matrix{Float64}(undef, model.dim.m, model.dim.m)
+
+    # Steady state initialization
+    steadystate = false
+    tsteady     = model.dim.n+1
+
+    # Initial state: big Kappa initialization
+    fill_a1!(a)
+    fill_P1!(P; bigkappa = 1e6)
+
+    mul!(RQR, model.R, LinearAlgebra.BLAS.gemm('N', 'T', 1.0, Q, model.R)) # RQR = R Q R'
+    # Kalman filter
+    for t = 1:model.dim.n
+        if t in model.missing_observations
+            steadystate  = false
+            v[t] = NaN
+            F[t] = NaN
+            update_att!(att, a, t) # att_t = a_t
+            update_Ptt!(Ptt, P, t) # Ptt_t = P_t
+            update_a!(a, att, model.T, t) # a_t+1 = T * att_t
+            update_P!(P, model.T, Ptt, RQR, t) # P_t+1 = T * Ptt_t * T' + RQR'
+        elseif steadystate
+            update_v!(v, model.y, model.Z, a, t) # v_t = y_t - Z_t * a_t
+            F[t] = F[t-1] # F[:, :, t]   = F[:, :, t-1]
+            repeat_vector_t_plus_1!(K, t-1) # K[:, t]   = K[:, t-1]
+            update_att!(att, a, P_Ztransp_invF, v, t) # att_t = a_t + P_t * Z_t * F^-1_t * v_t
+            repeat_matrix_t_plus_1!(Ptt, t-1) # Ptt_t = Ptt_t-1
+            update_a!(a, att, model.T, t) # a_t+1 = T * att_t
+            repeat_matrix_t_plus_1!(P, t) # P_t+1 = P_t
+        else
+            update_v!(v, model.y, model.Z, a, t) # v_t = y_t - Z_t * a_t
+            update_ZP!(ZP, model.Z, P, t) # ZP = Z[:, :, t] * P[:, :, t]
+            update_F!(F, ZP, model.Z, H, t) # F_t = Z_t * P_t * Z_t + H
+            update_P_Ztransp_Finv!(P_Ztransp_invF, ZP, F, t) # P_Ztransp_invF   = ZP' * invF(F, t)
+            update_K!(K, P_Ztransp_invF, model.T, t) # K_t = T * P_t * Z_t * F^-1_t
[…]
