+++ src/kalman_utils.jl
@@ -4,7 +4,6 @@ function fill_a1!(a::Matrix{T}) where T <: AbstractFloat
         a[1, i] = zero(T)
     end
 end
-
 function fill_P1!(P::AbstractArray{T}; bigkappa::Float64 = 1e6) where T <: AbstractFloat
     @inbounds for i in axes(P, 1), j in axes(P, 2)
         if i == j
@@ -21,12 +20,17 @@ function repeat_matrix_t_plus_1!(mat::AbstractArray{T}, t::Int) where T <: Abstr
     end
     return 
 end
+function repeat_vector_t_plus_1!(mat::Matrix{T}, t::Int) where T <: AbstractFloat
+    @inbounds for i in axes(mat, 1)
+        mat[i, t+1] = mat[i, t]
+    end
+    return 
+end
 
 function big_update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
     @views @inbounds mul!(a[t+1, :], T, att[t, :])
     return 
 end
-
 function small_update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
     @inbounds for i in axes(a, 2)
         a[t+1, i] = zero(Typ)
@@ -36,7 +40,6 @@ function small_update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::In
     end
     return 
 end
-
 function update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
     # Here there is a trade-off memory-speed, usually if the dimension of a is smaller than 15 
     # it is more performant to do the hard coded version (small_update_a)
@@ -48,13 +51,21 @@ function update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) whe
     return 
 end
 
-function update_ZP!(ZP::AbstractArray{T}, Z::AbstractArray{T}, P::AbstractArray{T}, t::Int) where T <: AbstractFloat
+function update_ZP!(ZP::Matrix{T}, Z::Array{T, 3}, P::Array{T, 3}, t::Int) where T <: AbstractFloat
     @inbounds @views mul!(ZP, Z[:, :, t], P[:, :, t])
     return 
 end
+function update_ZP!(ZP::Vector{T}, Z::Array{T, 3}, P::Array{T, 3}, t::Int) where T <: AbstractFloat
+    @inbounds @views ZP = Z[:, :, t] * P[:, :, t]
+    return 
+end
 
-function update_K!(K::AbstractArray{Typ}, P_Ztransp_invF::AbstractArray{Typ}, T::AbstractArray{Typ}, t::Int) where Typ <: AbstractFloat
+function update_K!(K::AbstractArray{Typ}, P_Ztransp_invF::AbstractMatrix{Typ}, T::AbstractMatrix{Typ}, t::Int) where Typ <: AbstractFloat
     @inbounds @views mul!(K[:, :, t], T, P_Ztransp_invF)
+    return
+end
+function update_K!(K::AbstractMatrix{Typ}, P_Ztransp_invF::AbstractVector{Typ}, T::AbstractMatrix{Typ}, t::Int) where Typ <: AbstractFloat
+    @inbounds K[:, t] = P_Ztransp_invF' * T
     return
 end
 
[…]

[…]
