### 2019-08-20 5f1b2b8
fixing small details and leaving #todos
