### 2019-08-21 cefa30d
some updates
