+++ src/univariate_kalman.jl
@@ -122,7 +122,7 @@ function univariate_smoother(model::StateSpaceModel, kfilter::UnivariateKalmanFi
         r0  = T' * r[1, :]
         N0  = T' * N[:, :, 1] * T
     else
-        Z_transp_invF = Z[1, :, 1]' * invertF(F, 1)
+        Z_transp_invF = Z[:, :, 1]' * invertF(F, 1)
         L[:, :, 1]    = T - K[:, 1] * Z[:, :, 1]
         r0            = Z_transp_invF * v[1] + L[:, :, 1]' * r[1, :]
         N0            = Z_transp_invF * Z[:, :, 1] + L[:, :, 1]' * N[:, :, 1] * L[:, :, 1]
