### 2019-08-23 9bf3ec0
Univariate filter (#96)
* fixing small details and leaving #todos

* added univariate_kalman_filter with tests
