+++ benchmark/bench.jl
@@ -32,6 +32,20 @@ bench_local_level = @benchmark statespace($locallevel, filter_type = $KalmanFilt
 #   samples:          250
 #   evals/sample:     1
 
+bench_local_level = @benchmark statespace($locallevel, filter_type = $UnivariateKalmanFilter, verbose = $verbose)
+# 22th August 2019
+# BenchmarkTools.Trial:
+#   memory estimate:  7.09 MiB
+#   allocs estimate:  42550
+#   --------------
+#   minimum time:     5.756 ms (0.00% GC)
+#   median time:      10.094 ms (0.00% GC)
+#   mean time:        11.012 ms (12.64% GC)
+#   maximum time:     41.820 ms (24.80% GC)
+#   --------------
+#   samples:          454
+#   evals/sample:     1
+
 H = fill(1.0, (1,1))
 Q = fill(1.0, (1,1))
 @benchmark StateSpaceModels.kalman_filter($locallevel, $H, $Q)
@@ -47,6 +61,22 @@ Q = fill(1.0, (1,1))
 #   --------------
 #   samples:          10000
 #   evals/sample:     1
+
+H = fill(1.0, (1,1))
+Q = fill(1.0, (1,1))
+@benchmark StateSpaceModels.univariate_kalman_filter($locallevel, $H, $Q)
+# 22th August 2019
+# BenchmarkTools.Trial:
+#   memory estimate:  8.33 KiB
+#   allocs estimate:  36
+#   --------------
+#   minimum time:     5.066 μs (0.00% GC)
+#   median time:      5.544 μs (0.00% GC)
+#   mean time:        6.965 μs (12.17% GC)
+#   maximum time:     1.124 ms (98.75% GC)
+#   --------------
+#   samples:          10000
+#   evals/sample:     6
 
 # Benchmark structural model
 AP = CSV.read("./examples/AirPassengers.csv")
@@ -82,4 +112,36 @@ Q[[2; 3; 4; 6; 7; 8]] .= 0 # Zero on off diagonal elements
 #   maximum time:     4.704 ms (86.14% GC)
 #   --------------
 #   samples:          8182
+#   evals/sample:     1
+
+
+bench_structural = @benchmark statespace($structural_model, filter_type = $UnivariateKalmanFilter, verbose = $verbose)
+# 21th August 2019
+# BenchmarkTools.Trial:
+#   memory estimate:  3.93 GiB
+#   allocs estimate:  2888472
+#   --------------
+#   minimum time:     1.923 s (10.59% GC)
+#   median time:      2.082 s (10.59% GC)
+#   mean time:        2.066 s (10.59% GC)
+#   maximum time:     2.194 s (10.60% GC)
+#   --------------
+#   samples:          3
+#   evals/sample:     1
+
+H = fill(1e-4, (1,1))
+Q = fill(1e-4, (3,3))
+Q[[2; 3; 4; 6; 7; 8]] .= 0 # Zero on off diagonal elements
+@benchmark StateSpaceModels.univariate_kalman_filter($structural_model, $H, $Q)
+# 21th August 2019
+# BenchmarkTools.Trial:
+#   memory estimate:  658.19 KiB
+#   allocs estimate:  446
+#   --------------
+#   minimum time:     243.650 μs (0.00% GC)
+#   median time:      253.331 μs (0.00% GC)
+#   mean time:        302.104 μs (11.47% GC)
+#   maximum time:     3.784 ms (83.94% GC)
+#   --------------
+#   samples:          10000
 #   evals/sample:     1
