### 2019-08-26 #28 @matbesancon
ping @dehann, @mschauer to continue the review :)
