### 2019-08-29 96a370c
tests to compare filters
