+++ test/test_lineartrend.jl
@@ -27,10 +27,10 @@
         @test ss3.smoother.alpha[:, 2] ≈ ones(15) rtol = 1e-4
         compare_forecast_simulation(ss3, 20, 1000, 1e-3)
 
-        @test ss2.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-4
-        @test ss3.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-4 # !!! failing
-        @test ss2.filter.a ≈ ss1.filter.a rtol = 1e-4
-        @test ss3.filter.a ≈ ss1.filter.a rtol = 1e-4
+        @test ss2.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-3
+        @test ss3.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-3 # !!! failing
+        @test ss2.filter.a ≈ ss1.filter.a rtol = 1e-3
+        @test ss3.filter.a ≈ ss1.filter.a rtol = 1e-3
     end
 
     @testset "Linear trend model with missing values" begin
