+++ test/test_results.jl
@@ -25,8 +25,8 @@ logAP = log.(Vector{Float64}(AP[:Passengers]))
     @test isa(ss3, StateSpace)
     compare_forecast_simulation(ss3, 20, 1000, 1e-2)
 
-    @test ss2.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-2
-    @test ss3.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-2 # !!! univariate failing
-    @test ss2.filter.a ≈ ss1.filter.a rtol = 1e-2
-    @test ss3.filter.a ≈ ss1.filter.a rtol = 1e-2
+    @test ss2.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-3
+    @test ss3.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-3 # !!! univariate failing
+    @test ss2.filter.a ≈ ss1.filter.a rtol = 1e-3
+    @test ss3.filter.a ≈ ss1.filter.a rtol = 1e-3
 end
