### 2019-08-30 b259384
Adding tests to compare filters (#98)
* tests to compare filters

* adding local level tests

* univariate smoother working

* fix path
