### 2019-09-27 fc3d35d
Add diagnostics
