### 2019-09-27 030b766
adding diagnostics to module
