### 2019-09-27 4aae712
fix docstring formats
