### 2019-09-27 5e2792f
update readme, index and manual
