### 2019-09-27 72f4062
adding diagnostics tests
