+++ test/test_lineartrend.jl
@@ -12,14 +12,14 @@
         @test isa(ss1, StateSpace)
         @test ss1.smoother.alpha[:, 2] ≈ ones(15) rtol = 1e-4
         compare_forecast_simulation(ss1, 20, 1000, 1e-3)
-                
+
         ss2 = statespace(unimodel; filter_type = SquareRootFilter)
 
         @test ss2.filter_type == SquareRootFilter
         @test isa(ss2, StateSpace)
         @test ss2.smoother.alpha[:, 2] ≈ ones(15) rtol = 1e-4
         compare_forecast_simulation(ss2, 20, 1000, 1e-3)
-                
+
         ss3 = statespace(unimodel; filter_type = UnivariateKalmanFilter)
 
         @test ss3.filter_type == UnivariateKalmanFilter
@@ -31,6 +31,13 @@
         @test ss3.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-3
         @test ss2.filter.a ≈ ss1.filter.a rtol = 1e-3
         @test ss3.filter.a ≈ ss1.filter.a rtol = 1e-3
+
+        diags1 = diagnostics(ss1)
+        diags2 = diagnostics(ss2)
+        diags3 = diagnostics(ss3)
+
+        @test diags1.p_jarquebera ≈ diags2.p_jarquebera atol = 1e-2
+        @test diags1.p_jarquebera ≈ diags3.p_jarquebera atol = 1e-2
     end
 
     @testset "Linear trend model with missing values" begin
