### 2019-09-28 7b67531
Add diagnostics (#100)
* Add diagnostics

* update version on print

* adding diagnostics to module

* fix include

* add statsbase to deps

* fix docstring formats

* update readme, index and manual

* adding diagnostics tests

* add multivariate diagnostics test
