+++ test/test_results.jl
@@ -1,5 +1,5 @@
 AP = CSV.read("../examples/AirPassengers.csv")
-logAP = log.(Vector{Float64}(AP[:Passengers]))
+logAP = log.(Vector{Float64}(AP[!, :Passengers]))
 
 @testset "Air passengers with Kalman filter" begin
     model = structural(logAP, 12)
@@ -29,4 +29,15 @@ logAP = log.(Vector{Float64}(AP[:Passengers]))
     @test ss3.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-3
     @test ss2.filter.a ≈ ss1.filter.a rtol = 1e-3
     @test ss3.filter.a ≈ ss1.filter.a rtol = 1e-3
+
+    diags1 = diagnostics(ss1)
+    diags2 = diagnostics(ss2)
+    diags3 = diagnostics(ss3)
+
+    @test diags1.p_jarquebera ≈ diags2.p_jarquebera atol = 1e-2
+    @test diags1.p_jarquebera ≈ diags3.p_jarquebera atol = 1e-2
+    @test diags1.p_ljungbox ≈ diags2.p_ljungbox atol = 1e-2
+    @test diags1.p_ljungbox ≈ diags3.p_ljungbox atol = 1e-2
+    @test diags1.p_homo ≈ diags2.p_homo atol = 1e-2
+    @test diags1.p_homo ≈ diags3.p_homo atol = 1e-2
 end
