+++ src/StateSpaceModels.jl
@@ -1,10 +1,10 @@
 module StateSpaceModels
 
-using Optim, Distributions, LinearAlgebra, StaticArrays, Dates, Printf
+using Optim, Distributions, LinearAlgebra, StaticArrays, Dates, Printf, StatsBase
 
 import Base: size, show
 
-export statespace, kfas
+export statespace, kfas, diagnostics
 
 include("prints.jl")
 include("structures.jl")
@@ -17,6 +17,7 @@ include("univariate_kalman.jl")
 include("kalman.jl")
 include("sqrt_kalman.jl")
 include("forecast.jl")
+include("diagnostics.jl")
 
 """
     statespace(model::StateSpaceModel; filter_type::DataType = KalmanFilter, optimization_method::AbstractOptimizationMethod = RandomSeedsLBFGS(), verbose::Int = 1)
