### 2019-09-30 #28 @mschauer
Dear authors,

I have some feedback regarding the design. Some of your implementation choices appear not very idiomatic to me, but you might have good reasons for them. On the other hand, this might be helpful feedback otherwise.

Let's start looking at
```julia
struct StateSpaceModel
    y::Matrix{Float64} # observations
    Z::Array{Float64, 3} # observation matrix
    T::Matrix{Float64} # state matrix
    R::Matrix{Float64} # state error matrix
    dim::StateSpaceDimensions
    missing_observations::Vector{Int}
    mode::String
end
```
The first thing to notice is that you restrict your users basically throughout to `Float64` matrices of type `Array`. What about embedded systems with `Float32` accuracy? What if their filtering problems are large (and can only be handeled with say banded or block-diagonal matrices? What about static arrays (might be helpful in realtime applications?)
At some places you seem to have it right,
```
function big_update_K!(K::AbstractMatrix{Typ}, P_Ztransp_invF::AbstractVector{Typ}, T::AbstractMatrix{Typ}, t::Int) where Typ <: AbstractFloat
    @inbounds @views K[:, t] = T * P_Ztransp_invF
    return
end
```
tolerates a range of different inputs. This flexibility in input (container and value-) types is also important for interaction with other packages, e.g. if you want to replace maximum-likelihood estimation by an MCMC method.

Another issue: Does your package support the online application of the filtering algorithm? The Kalman filter is a filter, after all... You mention the focus on time-series applications, but in the end, all the implementations, even the square root filter, are there (you wrote them), it would be nice to make them available in a more generic fashion.
