### 2019-09-30 #28 @raphaelsaavedra
Hello @mschauer, thank you for taking your time to do this review.

Regarding the generality of the `StateSpaceModel` structure, we believe you are absolutely right. The package would certainly benefit from allowing a wider range of types. In fact, we already do this inside the code, as you pointed out. We will try to address this point as soon as possible, as it seems like a simple change, at least in terms of allowing other Float types.

With respect to your enquiry about the online application: this is indeed something that we do not allow at the moment. We do have a function called `kfas` that runs the filtering and smoothing steps, but we should definitely allow users to call only the filter if they so desire. We will make filtering available so that online applications can become a possibility. This should also be a simple change.

Once again, thank you for your valuable feedback. We will make the changes and comment here once they are done.
