+++ src/sqrt_kalman.jl
@@ -1,9 +1,9 @@
 """
-    sqrt_kalman_filter(model::StateSpaceModel, sqrtH::Matrix{Typ}, sqrtQ::Matrix{Typ}; tol::Float64 = 1e-5) where Typ <: AbstractFloat
+    sqrt_kalman_filter(model::StateSpaceModel, sqrtH::Matrix{Typ}, sqrtQ::Matrix{Typ}; tol::Typ = 1e-5) where Typ <: AbstractFloat
 
 Square-root Kalman filter with big Kappa initialization.
 """
-function sqrt_kalman_filter(model::StateSpaceModel, sqrtH::Matrix{Typ}, sqrtQ::Matrix{Typ}; tol::Typ = 1e-5) where Typ <: AbstractFloat
+function sqrt_kalman_filter(model::StateSpaceModel{Typ}, sqrtH::Matrix{Typ}, sqrtQ::Matrix{Typ}; tol::Typ = 1e-5) where Typ <: AbstractFloat
 
     time_invariant = model.mode == "time-invariant"
 
@@ -15,18 +15,18 @@ function sqrt_kalman_filter(model::StateSpaceModel, sqrtH::Matrix{Typ}, sqrtQ::M
     Z, T, R = ztr(model)
 
     # Predictive state and its sqrt-covariance
-    a     = Matrix{Float64}(undef, n+1, m)
-    sqrtP = Array{Float64, 3}(undef, m, m, n+1)
+    a     = Matrix{Typ}(undef, n+1, m)
+    sqrtP = Array{Typ, 3}(undef, m, m, n+1)
 
     # Innovation and its sqrt-covariance
-    v     = Matrix{Float64}(undef, n, p)
-    sqrtF = Array{Float64, 3}(undef, p, p, n)
+    v     = Matrix{Typ}(undef, n, p)
+    sqrtF = Array{Typ, 3}(undef, p, p, n)
 
     # Kalman gain
-    K     = Array{Float64, 3}(undef, m, p, n)
+    K     = Array{Typ, 3}(undef, m, p, n)
 
     # Auxiliary matrices
-    U2star = Array{Float64, 3}(undef, m, p, n)
+    U2star = Array{Typ, 3}(undef, m, p, n)
 
     # Steady state initialization
     steadystate  = false
@@ -85,7 +85,7 @@ end
 
 Obtain the filtered state estimates and their covariance matrices.
 """
-function filtered_state(model::StateSpaceModel, sqrt_filter::SquareRootFilter)
+function filtered_state(model::StateSpaceModel{Typ}, sqrt_filter::SquareRootFilter{Typ}) where Typ <: AbstractFloat
 
     # Load dimensions data
     n, p, m, r = size(model)
@@ -100,8 +100,8 @@ function filtered_state(model::StateSpaceModel, sqrt_filter::SquareRootFilter)
     P = gram_in_time(sqrt_filter.sqrtP)
 
     # Filtered state and its covariance
-    att = Matrix{Float64}(undef, n, m)
-    Ptt = Array{Float64, 3}(undef, m, m, n)
+    att = Matrix{Typ}(undef, n, m)
+    Ptt = Array{Typ, 3}(undef, m, m, n)
 
     for t = 1:n
         PZF = P[:, :, t] * Z[:, :, t]' * inv(F[:, :, t])
@@ -119,7 +119,7 @@ end
 
 Square-root smoother for state space model.
 """
-function sqrt_smoother(model::StateSpaceModel, sqrt_filter::SquareRootFilter)
+function sqrt_smoother(model::StateSpaceModel{Typ}, sqrt_filter::SquareRootFilter{Typ}) where Typ <: AbstractFloat
 
     # Load dimensions data
     n, p, m, r = size(model)
[…]
