+++ src/kalman.jl
@@ -1,29 +1,29 @@
 """
-    kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; tol::Float64 = 1e-5) where Typ <: AbstractFloat
+    kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; tol::Typ = 1e-5) where Typ <: AbstractFloat
 
 Kalman filter with big Kappa initialization, i.e., initializing state variances as 1e6.
 """
-function kalman_filter(model::StateSpaceModel, H::Matrix{Typ}, Q::Matrix{Typ}; tol::Typ = 1e-5) where Typ <: AbstractFloat
+function kalman_filter(model::StateSpaceModel{Typ}, H::Matrix{Typ}, Q::Matrix{Typ}; tol::Typ = Typ(1e-5)) where Typ <: AbstractFloat
 
     time_invariant = model.mode == "time-invariant"
 
     # Predictive state and its covariance
-    a = Matrix{Float64}(undef, model.dim.n+1, model.dim.m)
-    P = Array{Float64, 3}(undef, model.dim.m, model.dim.m, model.dim.n+1)
-    att = Matrix{Float64}(undef, model.dim.n, model.dim.m)
-    Ptt = Array{Float64, 3}(undef, model.dim.m, model.dim.m, model.dim.n)
+    a = Matrix{Typ}(undef, model.dim.n+1, model.dim.m)
+    P = Array{Typ, 3}(undef, model.dim.m, model.dim.m, model.dim.n+1)
+    att = Matrix{Typ}(undef, model.dim.n, model.dim.m)
+    Ptt = Array{Typ, 3}(undef, model.dim.m, model.dim.m, model.dim.n)
 
     # Innovation and its covariance
-    v = Matrix{Float64}(undef, model.dim.n, model.dim.p)
-    F = Array{Float64, 3}(undef, model.dim.p, model.dim.p, model.dim.n)
+    v = Matrix{Typ}(undef, model.dim.n, model.dim.p)
+    F = Array{Typ, 3}(undef, model.dim.p, model.dim.p, model.dim.n)
 
     # Kalman gain
-    K = Array{Float64, 3}(undef, model.dim.m, model.dim.p, model.dim.n)
+    K = Array{Typ, 3}(undef, model.dim.m, model.dim.p, model.dim.n)
     
     # Auxiliary structures
-    ZP = Array{Float64, 2}(undef, model.dim.p, model.dim.m)
-    P_Ztransp_invF = Array{Float64, 2}(undef, model.dim.m, model.dim.p)
-    RQR = Array{Float64, 2}(undef, model.dim.m, model.dim.m)
+    ZP = Array{Typ, 2}(undef, model.dim.p, model.dim.m)
+    P_Ztransp_invF = Array{Typ, 2}(undef, model.dim.m, model.dim.p)
+    RQR = Array{Typ, 2}(undef, model.dim.m, model.dim.m)
 
     # Steady state initialization
     steadystate = false
@@ -77,7 +77,7 @@ end
 
 Smoother for state-space model.
 """
-function smoother(model::StateSpaceModel, kfilter::KalmanFilter)
+function smoother(model::StateSpaceModel{Typ}, kfilter::KalmanFilter{Typ}) where Typ <: AbstractFloat
 
     # Load dimensions data
     n, p, m, r = size(model)
@@ -93,11 +93,11 @@ function smoother(model::StateSpaceModel, kfilter::KalmanFilter)
     K       = kfilter.K
 
     # Smoothed state and its covariance
-    alpha = Matrix{Float64}(undef, n, m)
-    V     = Array{Float64, 3}(undef, m, m, n)
-    L     = Array{Float64, 3}(undef, m, m, n)
[…]

[…]
