### 2019-10-01 e93e229
finished imporve type system
