+++ src/utils.jl
@@ -62,7 +62,7 @@ end
 
 Ensure that matrix `M` is positive and symmetric to avoid numerical errors when numbers are small by doing `(M + M')/2 + ϵ*I`
 """
-function ensure_pos_sym!(M::AbstractArray{T}, t::Int; ϵ::T = 1e-8) where T <: AbstractFloat
+function ensure_pos_sym!(M::AbstractArray{T}, t::Int; ϵ::T = T(1e-8)) where T <: AbstractFloat
     @inbounds for j in axes(M, 2), i in 1:j
         if i == j
             M[i, i, t] = (M[i, i, t] + M[i, i, t])/2 + ϵ
