### 2019-10-02 80a5e32
Allow any float type in SSM (#101)
* improve type system

* finished imporve type system

* fix tests

* fix tests

* fix DataFrames
