+++ src/models.jl
@@ -1,14 +1,14 @@
 export structural, local_level, linear_trend
 
 """
-    structural(y::VecOrMat{Typ}, s::Int; X::VecOrMat{Typ} = Matrix{Float64}(undef, 0, 0)) where Typ <: AbstractFloat
+    structural(y::VecOrMat{Typ}, s::Int; X::VecOrMat{Typ} = Matrix{Typ}(undef, 0, 0)) where Typ <: AbstractFloat
 
 Build state-space system for a given structural model with observations y, seasonality s, and, optionally, exogenous variables X.
 
 If `y` is provided as an `Array{Typ, 1}` it will be converted to an `Array{Typ, 2}` inside the `StateSpaceModel`. The same will happen to X, 
 if an `Array{Typ, 1}` it will be converted to an `Array{Typ, 2}` inside the `StateSpaceModel`.
 """
-function structural(y::VecOrMat{Typ}, s::Int; X::VecOrMat{Typ} = Matrix{Float64}(undef, 0, 0)) where Typ <: AbstractFloat
+function structural(y::VecOrMat{Typ}, s::Int; X::VecOrMat{Typ} = Matrix{Typ}(undef, 0, 0)) where Typ <: AbstractFloat
 
     # Number of observations and endogenous variables
     y = y[:, :]
@@ -28,10 +28,10 @@ function structural(y::VecOrMat{Typ}, s::Int; X::VecOrMat{Typ} = Matrix{Float64}
     # Observation equation
     N = max(n, n_exp)
     if p_exp > 0 # exogenous variables: Z is time-variant
-        Z = Array{Float64, 3}(undef, p, m, N)
+        Z = Array{Typ, 3}(undef, p, m, N)
         for t = 1:N
             Z[:, :, t] = kron(
-                Matrix{Float64}(I, p, p),
+                Matrix{Typ}(I, p, p),
                 [
                     X[t, :]' 1 0 1 zeros(1, s - 2)
                 ]
@@ -39,7 +39,7 @@ function structural(y::VecOrMat{Typ}, s::Int; X::VecOrMat{Typ} = Matrix{Float64}
         end
     else # no exogenous variables: Z is time-invariant
         Z = kron(
-            Matrix{Float64}(I, p, p),
+            Matrix{Typ}(I, p, p),
             [
                 1 0 1 zeros(1, s - 2)
             ]
@@ -48,34 +48,34 @@ function structural(y::VecOrMat{Typ}, s::Int; X::VecOrMat{Typ} = Matrix{Float64}
     
     # State equation
     if p_exp > 0
-        T0 = [Matrix{Float64}(I, p_exp, p_exp) zeros(p_exp, 1 + s)]
+        T0 = [Matrix{Typ}(I, p_exp, p_exp) zeros(p_exp, 1 + s)]
         T = kron(
-            Matrix{Float64}(I, p, p),
+            Matrix{Typ}(I, p, p),
             [
                 T0; 
                 zeros(1, p_exp) 1 1 zeros(1, s - 1); 
                 zeros(1, p_exp) 0 1 zeros(1, s - 1);
                 zeros(1, p_exp) 0 0 -ones(1, s - 1);
-                zeros(s - 2, p_exp) zeros(s - 2, 2) Matrix{Float64}(I, s - 2, s - 2) zeros(s - 2)
+                zeros(s - 2, p_exp) zeros(s - 2, 2) Matrix{Typ}(I, s - 2, s - 2) zeros(s - 2)
             ]
             )
     else
         T = kron(
-            Matrix{Float64}(I, p, p),
[…]
