+++ src/structures.jl
@@ -26,46 +26,46 @@ Abstract type used to implement an interface for generic optimization methods.
 abstract type AbstractOptimizationMethod end
 
 # Auxiliary structure for square-root Kalman filter
-mutable struct SquareRootFilter <: AbstractFilter
-    a::Matrix{Float64} # predictive state
-    v::Matrix{Float64} # innovations
-    sqrtP::Array{Float64, 3} # lower triangular matrix with sqrt-covariance of the predictive state
-    sqrtF::Array{Float64, 3} # lower triangular matrix with sqrt-covariance of the innovations
+mutable struct SquareRootFilter{T <: AbstractFloat} <: AbstractFilter
+    a::Matrix{T} # predictive state
+    v::Matrix{T} # innovations
+    sqrtP::Array{T, 3} # lower triangular matrix with sqrt-covariance of the predictive state
+    sqrtF::Array{T, 3} # lower triangular matrix with sqrt-covariance of the innovations
     steadystate::Bool # flag that indicates if steady state was attained
     tsteady::Int # instant when steady state was attained; in case it wasn't, tsteady = n+1
-    K::Array{Float64, 3} # Kalman gain
+    K::Array{T, 3} # Kalman gain
 end
 
 # Auxiliary structure for Kalman filter
-mutable struct KalmanFilter <: AbstractFilter
-    a::Matrix{Float64} # predictive state
-    att::Matrix{Float64}
-    v::Matrix{Float64} # innovations
-    P::Array{Float64, 3} # covariance matrix of the predictive state
-    Ptt::Array{Float64, 3}
-    F::Array{Float64, 3} # covariance matrix of the innovations
+mutable struct KalmanFilter{T <: AbstractFloat} <: AbstractFilter
+    a::Matrix{T} # predictive state
+    att::Matrix{T}
+    v::Matrix{T} # innovations
+    P::Array{T, 3} # covariance matrix of the predictive state
+    Ptt::Array{T, 3}
+    F::Array{T, 3} # covariance matrix of the innovations
     steadystate::Bool # flag that indicates if steady state was attained
     tsteady::Int # instant when steady state was attained; in case it wasn't, tsteady = n+1
-    K::Array{Float64, 3} # Kalman gain
+    K::Array{T, 3} # Kalman gain
 end
 
 # Auxiliary structure for univariate Kalman filter
-mutable struct UnivariateKalmanFilter <: AbstractFilter
-    a::Matrix{Float64} # predictive state
-    att::Matrix{Float64}
-    v::Vector{Float64} # innovations
-    P::Array{Float64, 3} # covariance matrix of the predictive state
-    Ptt::Array{Float64, 3}
-    F::Vector{Float64} # covariance matrix of the innovations
+mutable struct UnivariateKalmanFilter{T <: AbstractFloat} <: AbstractFilter
+    a::Matrix{T} # predictive state
+    att::Matrix{T}
+    v::Vector{T} # innovations
+    P::Array{T, 3} # covariance matrix of the predictive state
+    Ptt::Array{T, 3}
+    F::Vector{T} # covariance matrix of the innovations
     steadystate::Bool # flag that indicates if steady state was attained
     tsteady::Int # instant when steady state was attained; in case it wasn't, tsteady = n+1
-    K::Matrix{Float64} # Kalman gain
+    K::Matrix{T} # Kalman gain
 end
 
 # Auxiliary structure for smoother
-mutable struct Smoother <: AbstractSmoother
[…]
