### 2019-10-07 c4b7b15
Definition of the structural model
