### 2019-10-08 963cd26
Adjusted structural model
