+++ docs/make.jl
@@ -1,6 +1,6 @@
-push!(LOAD_PATH, "C:\\Users\\mdietze\\Downloads\\cmder\\StateSpaceModels.jl\\src")
+push!(LOAD_PATH, "C:\\Users\\mdietze\\Documents\\StateSpaceModels.jl\\src")
 using Documenter, StateSpaceModels
-cd("C:\\Users\\mdietze\\Downloads\\cmder\\StateSpaceModels.jl\\docs")
+cd("C:\\Users\\mdietze\\Documents\\StateSpaceModels.jl\\docs")
 makedocs(
     modules = [StateSpaceModels],
     doctest  = false,
