### 2019-10-09 a76092e
Corrected structural model
