+++ docs/src/manual.md
@@ -84,7 +84,7 @@ The structural model is defined by
 ```math
 \begin{gather*}
     \begin{aligned}
-        y_{t+1} &=  \mu_{t+1} + \gamma_{t+1} + \varepsilon_{t} \quad &\varepsilon_{t} \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\
+        y_{t} &=  \mu_{t} + \gamma_{t} + \varepsilon_{t} \quad &\varepsilon_{t} \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\
         \mu_{t+1} &= \mu_{t} + \nu_{t} + \eta_{t} \quad &\eta_{t} \sim \mathcal{N}(0, \sigma^2_{\eta})\\
         \nu_{t+1} &= \nu_{t} + \zeta_{t} \quad &\zeta_{t} \sim \mathcal{N}(0, \sigma^2_{\zeta})\\
         \gamma_{t+1} &= \sum_{j=1}^{s/2} \gamma_{j,t}\\
