### 2019-10-09 3f9a355
Updated structural model
