### 2019-10-10 ddc64e7
Definition of the structural model (#103)
* Definition of the structural model
