### 2019-10-18 ea394d1
updating paper
