+++ paper/ref.bib
@@ -139,3 +139,25 @@
   year={2018},
   doi={10.17771/PUCRio.IRR.33804}
 }
+
+@article{jarque1980efficient,
+  title={Efficient tests for normality, homoscedasticity and serial independence of regression residuals},
+  author={Jarque, Carlos M and Bera, Anil K},
+  journal={Economics letters},
+  volume={6},
+  number={3},
+  pages={255--259},
+  year={1980},
+  publisher={Elsevier}
+}
+
+@article{box1970distribution,
+  title={Distribution of residual autocorrelations in autoregressive-integrated moving average time series models},
+  author={Box, George EP and Pierce, David A},
+  journal={Journal of the American statistical Association},
+  volume={65},
+  number={332},
+  pages={1509--1526},
+  year={1970},
+  publisher={Taylor \& Francis Group}
+}
