### 2019-10-18 #28 @raphaelsaavedra
Hello @mschauer,

@guilhermebodin and I have discussed and addressed some of your concerns. We have modified the package to allow any float accuracy within the model definition and the filters (https://github.com/LAMPSPUC/StateSpaceModels.jl/pull/101).

With respect to the online application of the filter, after some discussion we have decided not to implement it in the current form of the package. Within our package, the filter data (state information, covariances, Kalman gain etc.) are kept in memory and saved in their full form (all time periods). We do this to allow for studies that can be pertinent within the context of time series analysis. In order to make an online implementation of the filter, we would have to create new filter functions and structures from scratch. While we believe there is definitely value in doing that, we think this might be outside of the current scope of our package, which is mainly focused in time series. This might end up motivating a new package that can deal with online filtering applications.

As an unrelated note, we have also implemented statistical diagnostics to evaluate the model specification and estimation (https://github.com/LAMPSPUC/StateSpaceModels.jl/pull/100).

Thank you for the feedback and do not hesitate to comment again if you have further observations.
