### 2019-10-21 #28 @mschauer
Hi, this is a good change.
Can you relax to `Real` in `struct SquareRootFilter{T <: AbstractFloat}` as `AbstractFloat` is to narrow, e.g. dual numbers are only `Real`? Then you can also remove all `<: AbstractFloat` in function definitions as
```
function univariate_smoother(model::StateSpaceModel{Typ}, kfilter::UnivariateKalmanFilter{Typ}) where Typ <: AbstractFloat
```
it is enough to contain in one place (the struct itself).
I am looking at the estimation next.
