### 2019-10-21 0883701
change AbstractFloat to Real, fix some docstrings
