+++ src/kalman_utils.jl
@@ -1,10 +1,10 @@
 # Utils for filter performance
-function fill_a1!(a::Matrix{T}) where T <: AbstractFloat
+function fill_a1!(a::Matrix{T}) where T
     @inbounds for i in axes(a, 2)
         a[1, i] = zero(T)
     end
 end
-function fill_P1!(P::AbstractArray{T}; bigkappa::T = 1e6) where T <: AbstractFloat
+function fill_P1!(P::AbstractArray{T}; bigkappa::T = 1e6) where T
     @inbounds for i in axes(P, 1), j in axes(P, 2)
         if i == j
             P[i, j, 1] = bigkappa
@@ -14,24 +14,24 @@ function fill_P1!(P::AbstractArray{T}; bigkappa::T = 1e6) where T <: AbstractFlo
     end
 end
 
-function repeat_matrix_t_plus_1!(mat::AbstractArray{T}, t::Int) where T <: AbstractFloat
+function repeat_matrix_t_plus_1!(mat::AbstractArray{T}, t::Int) where T
     @inbounds for j in axes(mat, 2), i in axes(mat, 1)
           mat[i, j, t+1] = mat[i, j, t]
     end
     return 
 end
-function repeat_vector_t_plus_1!(mat::Matrix{T}, t::Int) where T <: AbstractFloat
+function repeat_vector_t_plus_1!(mat::Matrix{T}, t::Int) where T
     @inbounds for i in axes(mat, 1)
         mat[i, t+1] = mat[i, t]
     end
     return 
 end
 
-function big_update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
+function big_update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ
     @views @inbounds mul!(a[t+1, :], T, att[t, :])
     return 
 end
-function small_update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
+function small_update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ
     @inbounds for i in axes(a, 2)
         a[t+1, i] = zero(Typ)
         for j in axes(T, 2)
@@ -40,9 +40,9 @@ function small_update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::In
     end
     return 
 end
-function update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ <: AbstractFloat
-    # Here there is a trade-off memory-speed, usually if the dimension of a is smaller than 15 
-    # it is more performant to do the hard coded version (small_update_a)
+function update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) where Typ
+    # Here there is a trade-off between memory and speed, usually if the dimension of a is smaller than 15 
+    # it is more performant to do the hard-coded version (small_update_a)
     if size(a, 2) < 16
         small_update_a!(a, att, T, t)
     else
@@ -51,11 +51,11 @@ function update_a!(a::Matrix{Typ}, att::Matrix{Typ}, T::Matrix{Typ}, t::Int) whe
[…]
