### 2019-10-21 432822a
change AbstractFloat to Real, fix some docstrings (#107)
