+++ src/diagnostics.jl
@@ -33,7 +33,7 @@ end
 
 Obtain standardized residuals from a state-space model
 """
-function residuals(ss::StateSpace{T}) where T <: AbstractFloat
+function residuals(ss::StateSpace{T}) where T
     e = Matrix{T}(undef, ss.model.dim.n, ss.model.dim.p)
     for j = 1:ss.model.dim.p
         for t = 1:ss.model.dim.n
@@ -52,7 +52,7 @@ end
 
 Run Jarque-Bera normality test and return p-values
 """
-function jarquebera(e::Matrix{T}) where T <: AbstractFloat
+function jarquebera(e::Matrix{T}) where T
     n, p = size(e)
     pvalue = Vector{T}(undef, p)
     for j = 1:p
@@ -74,7 +74,7 @@ end
 
 Run Ljung-Box independence test and return p-values
 """
-function ljungbox(e::Matrix{T}; maxlag::Int = 20) where T <: AbstractFloat
+function ljungbox(e::Matrix{T}; maxlag::Int = 20) where T
     n, p = size(e)
     n <= maxlag && (maxlag = n-1)
     pvalue = Vector{T}(undef, p)
@@ -93,7 +93,7 @@ end
 
 Run homoscedasticity test and return p-values
 """
-function homoscedast(e::Matrix{T}) where T <: AbstractFloat
+function homoscedast(e::Matrix{T}) where T
     n, p = size(e)
     t1 = floor(Int, n/3)
     t2 = ceil(Int, 2*(n/3)+1)
