+++ src/structures.jl
@@ -26,7 +26,7 @@ Abstract type used to implement an interface for generic optimization methods.
 abstract type AbstractOptimizationMethod end
 
 # Auxiliary structure for square-root Kalman filter
-mutable struct SquareRootFilter{T <: AbstractFloat} <: AbstractFilter
+mutable struct SquareRootFilter{T <: Real} <: AbstractFilter
     a::Matrix{T} # predictive state
     v::Matrix{T} # innovations
     sqrtP::Array{T, 3} # lower triangular matrix with sqrt-covariance of the predictive state
@@ -37,7 +37,7 @@ mutable struct SquareRootFilter{T <: AbstractFloat} <: AbstractFilter
 end
 
 # Auxiliary structure for Kalman filter
-mutable struct KalmanFilter{T <: AbstractFloat} <: AbstractFilter
+mutable struct KalmanFilter{T <: Real} <: AbstractFilter
     a::Matrix{T} # predictive state
     att::Matrix{T}
     v::Matrix{T} # innovations
@@ -50,7 +50,7 @@ mutable struct KalmanFilter{T <: AbstractFloat} <: AbstractFilter
 end
 
 # Auxiliary structure for univariate Kalman filter
-mutable struct UnivariateKalmanFilter{T <: AbstractFloat} <: AbstractFilter
+mutable struct UnivariateKalmanFilter{T <: Real} <: AbstractFilter
     a::Matrix{T} # predictive state
     att::Matrix{T}
     v::Vector{T} # innovations
@@ -63,24 +63,23 @@ mutable struct UnivariateKalmanFilter{T <: AbstractFloat} <: AbstractFilter
 end
 
 # Auxiliary structure for smoother
-mutable struct Smoother{T <: AbstractFloat} <: AbstractSmoother
+mutable struct Smoother{T <: Real} <: AbstractSmoother
     alpha::Matrix{T} # smoothed state
     V::Array{T, 3} # variance of smoothed state
 end
-
 
 # Concrete structs for optimization methods
 """
 TODO
 """
-mutable struct RandomSeedsLBFGS{T <: AbstractFloat} <: AbstractOptimizationMethod
+mutable struct RandomSeedsLBFGS{T <: Real} <: AbstractOptimizationMethod
     f_tol::T
     g_tol::T
     iterations::Int
     nseeds::Int
     seeds::Array{T}
 
-    function RandomSeedsLBFGS(; f_tol::T = 1e-6, g_tol::T = 1e-6, iterations::Int = 10^5, nseeds::Int = 3) where T <: AbstractFloat
+    function RandomSeedsLBFGS(; f_tol::T = 1e-6, g_tol::T = 1e-6, iterations::Int = 10^5, nseeds::Int = 3) where T <: Real
         return new{T}(f_tol, g_tol, 1e5, nseeds)
     end
 end
@@ -118,7 +117,7 @@ A `StateSpaceModel` object can be defined using `StateSpaceModel(y::Matrix{Typ},
 
 Alternatively, if `Z` is time-invariant, it can be input as a single ``p \\times m`` matrix.
 """
-struct StateSpaceModel{Typ <: AbstractFloat}
+struct StateSpaceModel{Typ <: Real}
     y::Matrix{Typ} # observations
     Z::Array{Typ, 3} # observation matrix
     T::Matrix{Typ} # state matrix
@@ -127,7 +126,7 @@ struct StateSpaceModel{Typ <: AbstractFloat}
     missing_observations::Vector{Int}
     mode::String
 
-    function StateSpaceModel(y::Matrix{Typ}, Z::Array{Typ, 3}, T::Matrix{Typ}, R::Matrix{Typ}) where Typ <: AbstractFloat
[…]
