### 2019-10-21 42f4c8c
add julia 1.1 and 1.2 to travis
