### 2019-10-21 #28 @raphaelsaavedra
We agree this makes sense and it has been changed in https://github.com/LAMPSPUC/StateSpaceModels.jl/pull/107.
