### 2019-10-21 ca5cba3
improving some docstrings
