+++ src/models.jl
@@ -3,7 +3,7 @@ export structural, local_level, linear_trend
 """
     structural(y::VecOrMat{Typ}, s::Int; X::VecOrMat{Typ} = Matrix{Typ}(undef, 0, 0)) where Typ <: Real
 
-Build state-space system for a given structural model with observations y, seasonality s, and, optionally, exogenous variables X.
+Build state-space system for a given structural model with observations `y`, seasonality `s`, and, optionally, exogenous variables `X`.
 
 If `y` is provided as an `Array{Typ, 1}` it will be converted to an `Array{Typ, 2}` inside the `StateSpaceModel`. The same will happen to X, 
 if an `Array{Typ, 1}` it will be converted to an `Array{Typ, 2}` inside the `StateSpaceModel`.
@@ -89,7 +89,7 @@ end
 
 Build state-space system for a local level model with observations y.
 
-If `y` is proided as an `Array{Typ, 1}` it will be converted to an `Array{Typ, 2}` inside the `StateSpaceModel`.
+If `y` is provided as an `Array{Typ, 1}` it will be converted to an `Array{Typ, 2}` inside the `StateSpaceModel`.
 """
 function local_level(y::VecOrMat{Typ}) where Typ <: Real
 
@@ -115,7 +115,7 @@ end
 
 Build state-space system for a linear trend model with observations y.
 
-If `y` is proided as an `Array{Typ, 1}` it will be converted to an `Array{Typ, 2}` inside the `StateSpaceModel`.
+If `y` is provided as an `Array{Typ, 1}` it will be converted to an `Array{Typ, 2}` inside the `StateSpaceModel`.
 """
 function linear_trend(y::VecOrMat{Typ}) where Typ <: Real
 
