### 2019-10-21 869d5a4
improving some docstrings (#109)
