### 2019-10-22 ab6dc10
fixing multivariate and adding test
