+++ test/runtests.jl
@@ -1,4 +1,4 @@
-using Test, StateSpaceModels, Statistics, CSV, LinearAlgebra, Random
+using Test, StateSpaceModels, Statistics, CSV, LinearAlgebra, Random, Distributions
 
 const SSM = StateSpaceModels
 
