### 2019-10-22 de4695f
added best seed optimization result
