### 2019-10-22 bc98eb6
removing old multivariate test
