### 2019-10-23 cdd6345
fixing multivariate and adding test (#110)
* fixing multivariate and adding test

* adding distributions dependency on runtests

* removing old multivariate test
