+++ test/test_locallevel.jl
@@ -99,49 +99,78 @@ y = [0.0     # Deterministic local level generated time series
 6.165810835928076
 8.366964896750876]
 
-@testset "Local level" begin
-    @testset "Local level model with Kalman filter" begin
-        unimodel = local_level(y)
+@testset "Local level model with Kalman filter" begin
+    unimodel = local_level(y)
 
-        @test isa(unimodel, StateSpaceModel)
-        @test unimodel.mode == "time-invariant"
-        
-        ss1 = statespace(unimodel)
-        @test ss1.filter_type <: KalmanFilter
-        @test isa(ss1, StateSpace)
-        
-        ss2 = statespace(unimodel; filter_type = SquareRootFilter{Float64})
-        @test ss2.filter_type <: SquareRootFilter
-        @test isa(ss2, StateSpace)
-        
-        ss3 = statespace(unimodel; filter_type = UnivariateKalmanFilter{Float64})
-        @test ss3.filter_type <: UnivariateKalmanFilter
-        @test isa(ss3, StateSpace)
+    @test isa(unimodel, StateSpaceModel)
+    @test unimodel.mode == "time-invariant"
+    
+    ss1 = statespace(unimodel)
+    @test ss1.filter_type <: KalmanFilter
+    @test isa(ss1, StateSpace)
+    
+    ss2 = statespace(unimodel; filter_type = SquareRootFilter{Float64})
+    @test ss2.filter_type <: SquareRootFilter
+    @test isa(ss2, StateSpace)
+    
+    ss3 = statespace(unimodel; filter_type = UnivariateKalmanFilter{Float64})
+    @test ss3.filter_type <: UnivariateKalmanFilter
+    @test isa(ss3, StateSpace)
 
-        @test ss2.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-3
-        @test ss3.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-3
-        @test ss2.filter.a ≈ ss1.filter.a rtol = 1e-3
-        @test ss3.filter.a ≈ ss1.filter.a rtol = 1e-3
+    @test ss2.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-3
+    @test ss3.smoother.alpha ≈ ss1.smoother.alpha rtol = 1e-3
+    @test ss2.filter.a ≈ ss1.filter.a rtol = 1e-3
+    @test ss3.filter.a ≈ ss1.filter.a rtol = 1e-3
 
-        unimodel32 = local_level(Float32.(y))
-        ss4 = statespace(unimodel32; filter_type = KalmanFilter{Float32})
-        @test ss4.filter_type <: KalmanFilter
-        @test isa(ss4, StateSpace)
-        
-        ss5 = statespace(unimodel32; filter_type = SquareRootFilter{Float32})
-        @test ss5.filter_type <: SquareRootFilter
-        @test isa(ss5, StateSpace)
-        
-        ss6 = statespace(unimodel32; filter_type = UnivariateKalmanFilter{Float32})
-        @test ss6.filter_type <: UnivariateKalmanFilter
-        @test isa(ss6, StateSpace)
+    unimodel32 = local_level(Float32.(y))
+    ss4 = statespace(unimodel32; filter_type = KalmanFilter{Float32})
+    @test ss4.filter_type <: KalmanFilter
+    @test isa(ss4, StateSpace)
+    
+    ss5 = statespace(unimodel32; filter_type = SquareRootFilter{Float32})
[…]
