### 2019-10-24 302838f
Add best seed optimization result as an option in verbose (#111)
* added best seed optimization result

* update paper.pdf and paper.tex
