+++ test/test_locallevel.jl
@@ -104,15 +104,15 @@ y = [0.0     # Deterministic local level generated time series
 
     @test isa(unimodel, StateSpaceModel)
     @test unimodel.mode == "time-invariant"
-    
-    ss1 = statespace(unimodel)
+
+    ss1 = statespace(unimodel; verbose = 2)
     @test ss1.filter_type <: KalmanFilter
     @test isa(ss1, StateSpace)
-    
+
     ss2 = statespace(unimodel; filter_type = SquareRootFilter{Float64})
     @test ss2.filter_type <: SquareRootFilter
     @test isa(ss2, StateSpace)
-    
+
     ss3 = statespace(unimodel; filter_type = UnivariateKalmanFilter{Float64})
     @test ss3.filter_type <: UnivariateKalmanFilter
     @test isa(ss3, StateSpace)
