### 2019-10-24 69ab83d
eliminate statespacecovariance
