+++ src/StateSpaceModels.jl
@@ -9,6 +9,7 @@ export statespace, kfas, diagnostics
 include("prints.jl")
 include("structures.jl")
 include("utils.jl")
+include("model_unknowns.jl")
 include("models.jl")
 include("estimation.jl")
 include("random_seeds_lbfgs.jl")
@@ -35,24 +36,22 @@ function statespace(model::StateSpaceModel{T}; filter_type::DataType = KalmanFil
     print_header(verbose)
 
     # Maximum likelihood estimation
-    covariance = estimate_statespace(model, filter_type, optimization_method; verbose = verbose)
+    estimate_statespace!(model, filter_type, optimization_method; verbose = verbose)
 
     # Kalman filter and smoothing
-    filter_output, smoothed_state = kfas(model, covariance, filter_type)
+    filter_output, smoothed_state = kfas(model, filter_type)
 
     print_bottom(verbose)
 
-    return StateSpace{T}(model, filter_output, smoothed_state, covariance, filter_type, optimization_method)
+    return StateSpace{T}(model, filter_output, smoothed_state, filter_type, optimization_method)
 end
 
 """
-    kfas(model::StateSpaceModel, covariance::StateSpaceCovariance,
-         filter_type::DataType)
+    kfas(model::StateSpaceModel{T}, filter_type::DataType) where T
 
 Perform Kalman filter and smoother according to the chosen `filter_type`.
 """
-function kfas(model::StateSpaceModel{T}, covariance::StateSpaceCovariance{T},
-              filter_type::DataType) where T
+function kfas(model::StateSpaceModel{T}, filter_type::DataType) where T
     error(filter_type , " not implemented") # Returns an error if it cannot
                                             # find a specialized kfas
 end
