### 2019-10-25 9109fd5
fixed bug of H and Q exploding
