+++ src/random_seeds_lbfgs.jl
@@ -1,3 +1,18 @@
+"""
+TODO
+"""
+mutable struct RandomSeedsLBFGS{T <: Real} <: AbstractOptimizationMethod
+    f_tol::T
+    g_tol::T
+    iterations::Int
+    nseeds::Int
+    seeds::Array{T}
+
+    function RandomSeedsLBFGS(; f_tol::T = 1e-6, g_tol::T = 1e-6, iterations::Int = 10^5, nseeds::Int = 3) where T <: Real
+        return new{T}(f_tol, g_tol, 1e5, nseeds)
+    end
+end
+
 function estimate_statespace!(model::StateSpaceModel{T}, filter_type::DataType,
                              opt_method::RandomSeedsLBFGS; verbose::Int = 1) where T
 
@@ -42,16 +57,21 @@ function estimate_statespace!(model::StateSpaceModel{T}, filter_type::DataType,
 
     # Optimization
     for iseed = 1:nseeds
-        optseed = optimize(psitilde -> statespace_likelihood(psitilde, model, unknowns, valid_insts, filter_type), seeds[:, iseed],
-                            LBFGS(), Optim.Options(f_tol = opt_method.f_tol, 
-                                                   g_tol = opt_method.g_tol, 
-                                                   iterations = opt_method.iterations,
-                                                   show_trace = (verbose == 3 ? true : false) ))
-        loglikelihood[iseed] = -optseed.minimum
-        psi[:, iseed] = optseed.minimizer
-        push!(optseeds, optseed)
+        model_opt = copy(model)
+        try
+            optseed = optimize(psitilde -> statespace_likelihood(psitilde, model_opt, unknowns, valid_insts, filter_type), seeds[:, iseed],
+                                LBFGS(), Optim.Options(f_tol = opt_method.f_tol, 
+                                                       g_tol = opt_method.g_tol, 
+                                                       iterations = opt_method.iterations,
+                                                       show_trace = (verbose == 3 ? true : false) ))
+            loglikelihood[iseed] = -optseed.minimum
+            psi[:, iseed] = optseed.minimizer
+            push!(optseeds, optseed)
 
-        print_loglikelihood(verbose, iseed, loglikelihood, t0)
+            print_loglikelihood(verbose, iseed, loglikelihood, t0)
+        catch
+            println("seed diverged")
+        end
     end
 
     best_seed = argmax(loglikelihood)
