+++ src/utils.jl
@@ -1,4 +1,4 @@
-export size, ztr, copy
+export size, ztr, copy, statespace_recursion
 
 """
     size(model::StateSpaceModel)
@@ -16,6 +16,42 @@ Return the state space model arrays `Z`, `T` and `R` of the `StateSpaceModel`
 """
 function ztr(model::StateSpaceModel)
     return model.Z, model.T, model.R
+end
+
+"""
+    statespace_recursion(model::StateSpaceModel{Typ}, N::Int, initial_a::Matrix{Typ}) where Typ
+
+Runs the recursion of a state space model ignoring the y.
+"""
+function statespace_recursion(model::StateSpaceModel{Typ}, initial_a::Matrix{Typ}) where Typ
+
+    # If model has any unknows is not possible to perform the recursion
+    unknowns = Unknowns(model)
+    if unknowns.n_unknowns != 0
+        error("StateSpaceModel has unkown parameters.")
+    end
+
+    # Save the distributions H and Q
+    dist_H = MvNormal(model.H)
+    dist_Q = MvNormal(model.Q)
+
+    n, p = size(model.y)
+
+    if size(initial_a, 2) != p
+        error("intial_a must be a 1 by $p matrix.")
+    end
+
+    y = Matrix{Float64}(undef, n, p)
+    α = Matrix{Float64}(undef, n, p)
+    α[1, :] = initial_a
+
+    for t = 1:n-1
+        y[t, :]   = model.Z[:, :, t]*α[t, :] + rand(dist_H)
+        α[t+1, :] = model.T*α[t, :] + model.R*rand(dist_Q)
+    end
+    y[n, :] = model.Z[:, :, n]*α[n, :] + rand(dist_H)
+
+    return y, α
 end
 
 
