### 2019-10-28 109a0ec
test estimating params in all matrices
