### 2019-10-29 11df648
better docs for statespace function
