+++ src/StateSpaceModels.jl
@@ -23,7 +23,8 @@ include("diagnostics.jl")
 """
     statespace(model::StateSpaceModel; filter_type::DataType = KalmanFilter, optimization_method::AbstractOptimizationMethod = RandomSeedsLBFGS(), verbose::Int = 1)
 
-Estimate the pre-specified state-space model.
+Estimate the pre-specified state-space model. The function will only estimate the entries that are declared `NaN`. If there are no NaNs in the `model` it will
+only perform the filter and smoother computations.
 """
 function statespace(model::StateSpaceModel{T}; filter_type::DataType = KalmanFilter{T}, 
                     optimization_method::AbstractOptimizationMethod = RandomSeedsLBFGS(), verbose::Int = 1) where T
