### 2019-10-29 92eec1d
eliminate Float64 declaration
