### 2019-10-30 8a8b844
added tests and fixed some structures
