+++ src/utils.jl
@@ -35,14 +35,14 @@ function statespace_recursion(model::StateSpaceModel{Typ}, initial_a::Matrix{Typ
     dist_H = MvNormal(model.H)
     dist_Q = MvNormal(model.Q)
 
-    n, p = size(model.y)
+    p, m, n = size(model.Z)
 
-    if size(initial_a, 2) != p
+    if size(initial_a, 2) != m
         error("intial_a must be a 1 by $p matrix.")
     end
 
     y = Matrix{Typ}(undef, n, p)
-    α = Matrix{Typ}(undef, n, p)
+    α = Matrix{Typ}(undef, n, m)
     α[1, :] = initial_a
 
     for t = 1:n-1
@@ -175,10 +175,6 @@ function Base.show(io::IO, model::StateSpaceModel)
     return nothing
 end
 
-function copy(ssm::StateSpaceModel)
-    return StateSpaceModel(copy(ssm.y), copy(ssm.Z), copy(ssm.T), copy(ssm.R), copy(ssm.H), copy(ssm.Q), ssm.mode)
-end
-
 function build_H(p::Int, T)
     H = fill(NaN, p, p)
     return T.(H)
@@ -188,4 +184,26 @@ function build_Q(r::Int, p::Int, T)
     Q = kron(ones(p, p), Matrix{T}(I, Int(r/p), Int(r/p)))
     Q[findall(isequal(1), Q)] .= NaN
     return Q
+end
+
+function build_ss_dim(y::Matrix{Typ}, Z::Array{Typ, 3}, T::Matrix{Typ}, R::Matrix{Typ}) where Typ <: Real
+    ny, py = size(y)
+    pz, mz, nz = size(Z)
+    mt1, mt2 = size(T)
+    mr, rr = size(R)
+    if !((mz == mt1 == mt2 == mr) && (pz == py))
+        error("StateSpaceModel dimension mismatch")
+    end
+    return StateSpaceDimensions(ny, py, mr, rr)
+end
+
+function build_ss_dim(y::Matrix{Typ}, Z::Matrix{Typ}, T::Matrix{Typ}, R::Matrix{Typ}) where Typ <: Real
+    ny, py = size(y)
+    pz, mz = size(Z)
+    mt, mt = size(T)
+    mr, rr = size(R)
+    if !((mz == mt == mr) && (pz == py))
+        error("StateSpaceModel dimension mismatch")
+    end
+    return StateSpaceDimensions(ny, py, mr, rr)
 end

[…]
