### 2019-10-30 7bc9d8d
Eliminate StateSpaceCovariance (#113)
* eliminate statespacecovariance

* fixed bug of H and Q exploding

* added statespace_recursion function

* test estimating params in all matrices

* better docs for statespace function

* eliminate Float64 declaration

* added tests and fixed some structures

* add section about custom models

* remove copy

* last updates

* fix phrase
