+++ src/structures.jl
@@ -1,4 +1,4 @@
-export StateSpaceDimensions, StateSpaceModel, StateSpaceCovariance, SmoothedState,
+export StateSpaceDimensions, StateSpaceModel, SmoothedState,
         FilterOutput, StateSpace
 
 export KalmanFilter, SquareRootFilter, UnivariateKalmanFilter
@@ -68,23 +68,6 @@ mutable struct Smoother{T <: Real} <: AbstractSmoother
     V::Array{T, 3} # variance of smoothed state
 end
 
-# Concrete structs for optimization methods
-"""
-TODO
-"""
-mutable struct RandomSeedsLBFGS{T <: Real} <: AbstractOptimizationMethod
-    f_tol::T
-    g_tol::T
-    iterations::Int
-    nseeds::Int
-    seeds::Array{T}
-
-    function RandomSeedsLBFGS(; f_tol::T = 1e-6, g_tol::T = 1e-6, iterations::Int = 10^5, nseeds::Int = 3) where T <: Real
-        return new{T}(f_tol, g_tol, 1e5, nseeds)
-    end
-end
-
-
 """
     StateSpaceDimensions
 
@@ -112,76 +95,68 @@ Following the notation of on the book \"Time Series Analysis by State Space Meth
 * `Z` A ``p \\times m \\times n`` matrix
 * `T` A ``m \\times m`` matrix
 * `R` A ``m \\times r`` matrix
+* `H` A ``p \\times p`` matrix
+* `Q` A ``r \\times r`` matrix
 
 A `StateSpaceModel` object can be defined using `StateSpaceModel(y::Matrix{Typ}, Z::Array{Typ, 3}, T::Matrix{Typ}, R::Matrix{Typ})`.
-
 Alternatively, if `Z` is time-invariant, it can be input as a single ``p \\times m`` matrix.
 """
 struct StateSpaceModel{Typ <: Real}
-    y::Matrix{Typ} # observations
-    Z::Array{Typ, 3} # observation matrix
-    T::Matrix{Typ} # state matrix
-    R::Matrix{Typ} # state error matrix
+    y::Matrix{Typ} # Observations
+    Z::Array{Typ, 3} # Observation matrix
+    T::Matrix{Typ} # State matrix
+    R::Matrix{Typ} # State error matrix
+    H::Matrix{Typ} # Covariance matrix of the observation vector
+    Q::Matrix{Typ} # Covariance matrix of the state vector
     dim::StateSpaceDimensions
     missing_observations::Vector{Int}
     mode::String
 
+    function StateSpaceModel(y::Matrix{Typ}, Z::Array{Typ, 3}, T::Matrix{Typ}, R::Matrix{Typ}, 
+                             H::Matrix{Typ}, Q::Matrix{Typ}) where Typ <: Real
+        # Build StateSpaceDimensions
+        dim = build_ss_dim(y, Z, T, R)
+        return new{Typ}(y, Z, T, R, H, Q, dim, find_missing_observations(y), "time-variant")
+    end
+
+    function StateSpaceModel(y::Matrix{Typ}, Z::Matrix{Typ}, T::Matrix{Typ}, R::Matrix{Typ}, 
+                             H::Matrix{Typ}, Q::Matrix{Typ}) where Typ <: Real
+        # Build StateSpaceDimensions
+        dim = build_ss_dim(y, Z, T, R)
+        Zvar = Array{Typ, 3}(undef, dim.p, dim.m, dim.n)
+        for t in 1:dim.n, i in axes(Z, 1), j in axes(Z, 2)
+            Zvar[i, j, t] = Z[i, j]
+        end
+
+        return new{Typ}(y, Zvar, T, R, H, Q, dim, find_missing_observations(y), "time-invariant")
+    end
+
[…]
