### 2019-10-30 bbcebec
added mean adjustments
