+++ src/utils.jl
@@ -78,7 +78,7 @@ end
     check_steady_state(P_t1::Matrix{T}, P_t::Matrix{T}, tol::T) where T
 
 Return `true` if steady state was attained with respect to tolerance `tol`, `false` otherwise.
-The steady state is checked by the following equation. TODO
+The steady state is checked by the following equation `maximum(abs.((P_t1 - P_t)./P_t1)) < tol`.
 """
 function check_steady_state(P_t1::Matrix{T}, P_t::Matrix{T}, tol::T) where T
     return maximum(abs.((P_t1 - P_t)./P_t1)) < tol ? true : false
