+++ test/test_locallevel.jl
@@ -181,8 +181,8 @@ function test_model_estimation(nan_pos::Int, n; rtol = 1e-1, nseeds = 3, seed = 
     Z = [1.0][:, :]
     T = [1.0][:, :]
     R = [1.0][:, :]
-    d = zeros(n, 1)
-    c = zeros(n, 1)
+    d = 0.01*ones(n, 1)
+    c = 0.01*ones(n, 1)
     H = [1.0][:, :]
     Q = [1.0][:, :]
     model = SSM.StateSpaceModel(y, Z, T, R, d, c, H, Q)
@@ -216,8 +216,8 @@ function test_model_estimation(nan_pos::Int, n; rtol = 1e-1, nseeds = 3, seed = 
     @test ss.model.Z[end] ≈ 1.0 rtol = rtol
     @test ss.model.T[1] ≈ 1.0 rtol = rtol
     @test ss.model.R[1] ≈ 1.0 rtol = rtol
-    @test ss.model.d[1] ≈ 0.0 rtol = rtol
-    @test ss.model.c[1] ≈ 0.0 rtol = rtol
+    @test ss.model.d[1] ≈ 0.01 rtol = rtol
+    @test ss.model.c[1] ≈ 0.01 rtol = rtol
     @test ss.model.H[1] ≈ 1.0 rtol = rtol
     @test ss.model.Q[1] ≈ 1.0 rtol = rtol
 end
