### 2019-11-01 e7716e9
remove StateSpaceCovariance from docs
