### 2019-11-04 3138914
Added mean adjustments (#117)
* added mean adjustments

* added c

* fix state space recursion

* update unknowns and test

* added a small value to c and d

* remove StateSpaceCovariance from docs
