+++ test/runtests.jl
@@ -9,4 +9,5 @@ include("test_locallevel.jl")
 include("test_lineartrend.jl")
 include("test_structural.jl")
 include("test_simulate.jl")
+include("test_opt_methods.jl")
 include("test_results.jl")
