### 2019-11-04 c3a4431
fixed LBFGS and added BFGS (#119)
