### 2019-11-05 859b8cd
added regression as method
