### 2019-11-05 2d42c07
Update docs/src/manual.md
Co-Authored-By: Raphael Saavedra <raphael.saavedra93@gmail.com>
