### 2019-11-05 2bcbafe
Update test/test_regression.jl
Co-Authored-By: Raphael Saavedra <raphael.saavedra93@gmail.com>
