+++ test/test_regression.jl
@@ -1,5 +1,5 @@
 @testset "regression" begin
-    Random.seed!(10)
+    Random.seed!(14)
     beta = [10; 20; 30]
     n = 1000
     X = rand(n, 3)
@@ -10,5 +10,5 @@
     ss = statespace(model)
 
     @test ss.smoother.alpha[end, :] ≈ [10; 20; 30] rtol = 1e-2
-    @test ss.model.H[1] ≈ 1.0 rtol = 1e-1
+    @test ss.model.H[1] ≈ 1.0 rtol = 1e-2
 end
