### 2019-11-06 0d8b648
Added regression as method (#120)
* added regression as method

* Update docs/src/manual.md

Co-Authored-By: Raphael Saavedra <raphael.saavedra93@gmail.com>

* Update src/models.jl

Co-Authored-By: Raphael Saavedra <raphael.saavedra93@gmail.com>

* Update test/test_regression.jl

Co-Authored-By: Raphael Saavedra <raphael.saavedra93@gmail.com>

* better test

* fix test

* fix test
