### 2019-11-18 cc4d9a8
added aic and bic, also improved opt_method
