+++ src/StateSpaceModels.jl
@@ -12,8 +12,7 @@ include("utils.jl")
 include("model_unknowns.jl")
 include("models.jl")
 include("estimation.jl")
-include("lbfgs.jl")
-include("bfgs.jl")
+include("opt_methods.jl")
 include("kalman_utils.jl")
 include("univariate_kalman.jl")
 include("kalman.jl")
