### 2019-11-18 4ea7350
added aic and bic, also improved opt_method (#124)
