### 2019-11-19 27fb6f6
fix forecast and simulate
