+++ src/forecast.jl
@@ -13,11 +13,12 @@ function forecast(ss::StateSpace{Typ}, N::Int) where Typ
     Q = ss.model.Q
 
     # Load system
-    n, p, m, r = size(ss.model)
-    Z, T, R    = prepare_forecast(ss, N)
+    n, p, m, r    = size(ss.model)
+    Z, T, R, d, c = prepare_forecast(ss, N)
 
-    # Load a, P, and F at last in-sample instant
+    # Load a, c, P, and F at last in-sample instant
     a0 = ss.smoother.alpha[end, :]
+    c0 = ss.model.c[end, :]
     P0 = ss.filter.P[:, :, end]
     F0 = ss.filter.F[:, :, end]
     
@@ -30,18 +31,18 @@ function forecast(ss::StateSpace{Typ}, N::Int) where Typ
     dist = Vector{Distribution}(undef, N)
 
     # Initialization
-    a[1, :]    = T*a0
+    a[1, :]    = T*a0 + c0
     P[:, :, 1] = T*P0*T' + R*Q*R'
     F[:, :, 1] = Z[:, :, 1]*P[:, :, 1]*Z[:, :, 1]' + H
     ensure_pos_sym!(F, 1)
-    dist[1]    = MvNormal(vec(Z[:, :, 1]*a[1, :]), F[:, :, 1])
+    dist[1]    = MvNormal(vec(Z[:, :, 1]*a[1, :] + d[1, :]), F[:, :, 1])
 
     for t = 2:N
-        a[t, :]    = T*a[t-1, :]
+        a[t, :]    = T*a[t-1, :] + c[t-1, :]
         P[:, :, t] = T*P[:, :, t-1]*T' + R*Q*R'
         F[:, :, t] = Z[:, :, t]*P[:, :, t]*Z[:, :, t]' + H
         ensure_pos_sym!(F, t)
-        dist[t]    = MvNormal(vec(Z[:, :, t]*a[t, :]), F[:, :, t])
+        dist[t]    = MvNormal(vec(Z[:, :, t]*a[t, :] + d[t, :]), F[:, :, t])
     end
 
     forec = Matrix{Typ}(undef, N, p)
@@ -65,8 +66,8 @@ function simulate(ss::StateSpace{Typ}, N::Int, S::Int) where Typ
     Q = ss.model.Q
 
     # Load system
-    n, p, m, r = size(ss.model)
-    Z, T, R    = prepare_forecast(ss, N)
+    n, p, m, r    = size(ss.model)
+    Z, T, R, d, c = prepare_forecast(ss, N)
 
     # Distribution of the state space errors
     dist_ϵ = MvNormal(zeros(p), H)
@@ -82,13 +83,13 @@ function simulate(ss::StateSpace{Typ}, N::Int, S::Int) where Typ
         η = rand(dist_η, N)'
 
         # Initializing simulation
-        αsim[1, :, s] = T*ss.smoother.alpha[n, :] + R*η[1, :]
-        ysim[1, :, s] = Z[:, :, 1]*αsim[1, :, s] + ϵ[1, :]
+        αsim[1, :, s] = T*ss.smoother.alpha[n, :] + c[1, :] + R*η[1, :]
+        ysim[1, :, s] = Z[:, :, 1]*αsim[1, :, s] + d[1, :] + ϵ[1, :]
 
         # Simulating future scenarios
         for t = 2:N
-            αsim[t, :, s] = T*αsim[t-1, :, s] + R*η[t, :]
-            ysim[t, :, s] = Z[:, :, t]*αsim[t, :, s] + ϵ[t, :]
[…]
