### 2019-11-25 69d2436
Update examples.md
add mention to example folder
