### 2019-11-25 45461ba
simplify MvNormal calls
