+++ docs/src/examples.md
@@ -96,10 +96,10 @@ In this example, we define the noise variances ``H`` and ``Q``, generate the noi
 ```julia
 # Generate random actuators
 Q = .5 * Matrix{Float64}(I, q, q)
-η = MvNormal(zeros(q), Q)
+η = MvNormal(Q)
 # Generate random measurement noise
 H = 2. * Matrix{Float64}(I, p, p)
-ε = MvNormal(zeros(p), H)
+ε = MvNormal(H)
 # Simulate vehicle trajectory
 α = zeros(n + 1, m)
 y = zeros(n, p)
