### 2019-11-25 57b4add
Bump version (#125)
* bump version

* add mention to example folder

* simplify MvNormal calls
