+++ Project.toml
@@ -14,6 +14,10 @@ StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [compat]
 julia = "1"
+Optim = "~0.19"
+Distributions = "~0.21"
+StatsBase = "~0.32"
+StaticArrays = "~0.12"
 
 [extras]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
