### 2019-11-25 e19840b
fix docstrings
