### 2019-11-25 bce8bd0
fix docstrings (#129)
* fix docstrings
