### 2019-11-29 8118c78
return to MathJax, fix some issues with docs
