### 2019-11-29 5e6fb50
fix structural model definition
