### 2019-11-29 a3f1ecc
return to MathJax, fix some issues with docs (#133)
* return to MathJax, fix some issues with docs

* fix structural model definition
