### 2019-12-09 1754bef
big kappa should be 1e3 for sqrt filter
