### 2020-02-04 01e92cd
update travis-ci script
