### 2020-02-04 a310bd7
Update Travis to travis-ci.com (#140)
* Update Travis to travis-ci.com

* fix docs

* update travis-ci script
