+++ README.md
@@ -37,8 +37,8 @@ Current features:
 Planned features:
 * Univariate treatment of multivariate models
 
-[build-img]: https://travis-ci.org/LAMPSPUC/StateSpaceModels.jl.svg?branch=master
-[build-url]: https://travis-ci.org/LAMPSPUC/StateSpaceModels.jl
+[build-img]: https://travis-ci.com/LAMPSPUC/StateSpaceModels.jl.svg?branch=master
+[build-url]: https://travis-ci.com/LAMPSPUC/StateSpaceModels.jl
 
 [codecov-img]: https://codecov.io/gh/LAMPSPUC/StateSpaceModels.jl/coverage.svg?branch=master
 [codecov-url]: https://codecov.io/gh/LAMPSPUC/StateSpaceModels.jl?branch=master
