### 2020-02-06 c777d22
imporove docs and fix diagnostics
