### 2020-02-06 394bdc4
Improve docs and fix diagnostics (#141)
* improve docs and fix diagnostics

* adress suggestions
