+++ .travis.yml
@@ -4,6 +4,7 @@ julia:
   - 1.0
   - 1.1
   - 1.2
+  - 1.3
 
 os:
   - linux
