### 2020-02-14 53c8b9c
Update dependency versions (#144)
* Update dependency versions

* Add Julia .13 to Travis

* Update Project.toml
