### 2020-02-17 e58e8d8
add gif to docs
