+++ examples/vehicle_tracking.jl
@@ -1,7 +1,10 @@
 using StateSpaceModels, Distributions, LinearAlgebra, Plots, Random
 
+# Fix seed
+Random.seed!(1)
+
 # Number of observations
-n = 300
+n = 400
 # State dimension (2d position + 2d speed)
 m = 4
 # Measurements dimension (2d noisy position)
@@ -10,23 +13,23 @@ p = 2
 q = 2
 
 # Damping ratio
-ρ = .05
+ρ = 0.05
 # Time delta
-Δ = 1.
+Δ = 1.0
 
 # State transition matrix
-T = kron(Matrix{Float64}(I, p, p), [1. (1. - ρ * Δ / 2.) * Δ; 0. (1. - ρ * Δ)])
+T = kron(Matrix{Float64}(I, p, p), [1.0 (1.0 - ρ * Δ / 2.0) * Δ; 0.0 (1.0 - ρ * Δ)])
 # Input matrix
-R = kron(Matrix{Float64}(I, p, p), [.5 * Δ^2.; Δ])
+R = kron(Matrix{Float64}(I, p, p), [0.5 * Δ^2; Δ])
 # Output (measurement) matrix
-Z = kron(Matrix{Float64}(I, p, p), [1. 0])
+Z = kron(Matrix{Float64}(I, p, p), [1.0 0.0])
 
 # Generate random actuators
-Q = .5 * Matrix{Float64}(I, q, q)
+Q = 0.5 * Matrix{Float64}(I, q, q)
 η = MvNormal(Q)
 
 # Generate random measurement noise
-H = 2. * Matrix{Float64}(I, p, p)
+H = 2.0 * Matrix{Float64}(I, p, p)
 ε = MvNormal(H)
 
 # Simulate vehicle trajectory
@@ -40,10 +43,16 @@ end
 
 # User defined model
 model = StateSpaceModel(y, Z, T, R)
-
 # Estimate vehicle speed and position
 ss = statespace(model)
 
-plot(y[:, 1], y[:, 2], label="Measured position", line=:scatter, lw=2, markeralpha=0.2, color=:black, title="Vehicle tracking")
-plot!(α[:, 1], α[:, 3], label="True position", lw=3, color=:indianred)
-plot!(ss.smoother.alpha[:, 1], ss.smoother.alpha[:, 3], label="Estimated position", lw=2, color=:forestgreen)
+ss.model.H
+ss.model.Q
+
+anim = @animate for i in 1:n
+    plot(y[1:i, 1], y[1:i, 2], label="Measured position", line=:scatter, lw=2, markeralpha=0.2, color=:black, title="Vehicle tracking")
+    plot!(α[1:i, 1], α[1:i, 3], label="True position", lw=3, color=:indianred)
+    plot!(ss.smoother.alpha[1:i, 1], ss.smoother.alpha[1:i, 3], label="Estimated position", lw=2, color=:forestgreen)
+end
+
+gif(anim, "vehicle_tracking.gif", fps = 15)
