### 2020-02-18 e2a1ff2
fix d and c sizes
