### 2020-02-18 554eca7
update gif and add it to README
