+++ examples/vehicle_tracking.jl
@@ -52,7 +52,7 @@ ss.model.Q
 anim = @animate for i in 1:n
     plot(y[1:i, 1], y[1:i, 2], label="Measured position", line=:scatter, lw=2, markeralpha=0.2, color=:black, title="Vehicle tracking")
     plot!(α[1:i, 1], α[1:i, 3], label="True position", lw=3, color=:indianred)
-    plot!(ss.smoother.alpha[1:i, 1], ss.smoother.alpha[1:i, 3], label="Estimated position", lw=2, color=:forestgreen)
+    plot!(ss.filter.a[2:i+1, 1], ss.filter.a[2:i+1, 3], label="Estimated position", lw=2, color=:forestgreen)
 end
 
 gif(anim, "vehicle_tracking.gif", fps = 15)
