### 2020-02-20 #28 @dehann
Dear Authors,

i liked the paper as well as the aim of the package.  I understand the package as standardizing linearized state-space models for JuliaLang, and found the structure familiar from existing literature.  I did have a little trouble switching to the nomenclature in the paper, but that is not serious.

Specific comments:
- p1.,right column (r.) second paragraph (g.2), this you can drop or rephrase the sentence "Furthermore, some have not been updated for a long time and are not functional in Julia1.0."
- Sec. 3.1.1 perhaps give a bit more detail about "stochastic level component" -- perhaps an example of something one try estimate in the wild.  E.g. estimating the temperature of object without a direct measurement, etc.
- Sec. 3.1.2 Relate slope term directly to eq. 3 - 7.
- Sec. 3.1.3, assuming you mean seasonality in weather sense?  Perhaps just elaborate a little more.
- Section 4, g.1 common literature in state estimation has measurement noise matrix as `R`.
- Sec 4.1, 4.2, please clarify meaning of "derive" do you perhaps mean "predict"?
- Sec 4.2, clarify use of predict and update cycles, e.g. can you do multiple predictions and then a measurement, or can you do multiple measurements at current value?
- Sec. 6.1, g.1 "well-known" ... perhaps add a reference.
- g.2, "we can analyzed the individual" -> "we can analyze the individual...".
- Sec 6.2, by quantiles, you likely mean 95% the same as 2 sigma since all Gaussian beliefs, not in the sense that multiple filters are running or some other output conversion is done to arrive at quantiles.
- Consider adding a reference to the plotting library used (likely one from Julia :-)

Great, and sorry again for not uploading sooner!
