+++ paper/ref.bib
@@ -161,3 +161,14 @@
   year={1970},
   publisher={Taylor \& Francis Group}
 }
+
+@article{lutkepohl2012role,
+  title={The role of the log transformation in forecasting economic variables},
+  author={L{\"u}tkepohl, Helmut and Xu, Fang},
+  journal={Empirical Economics},
+  volume={42},
+  number={3},
+  pages={619--638},
+  year={2012},
+  publisher={Springer}
+}
