+++ paper/paper.tex
@@ -33,7 +33,7 @@ There are several packages focused on state-space models for time-series analysi
     \item ~\texttt{ControlSystems.jl} \cite{controlsystemsjl}, a package aimed at control applications with numerous features, including state-space modeling.
 \end{itemize}
 
-The contributions brought by these packages notwithstanding, there are currently no state-space model packages aimed at time-series analysis in Julia. Furthermore, some have not been updated for a long time and are not functional in Julia 1.0. Thus, the main objectives of \texttt{StateSpaceModels.jl} \cite{statespacemodels} are:
+The contributions brought by these packages notwithstanding, there are currently no state-space model packages aimed at time-series analysis in Julia. Thus, the main objectives of \texttt{StateSpaceModels.jl} \cite{statespacemodels} are:
 
 \begin{enumerate}
     \item to fill this gap by implementing a general and intuitive framework in Julia for modeling, estimating, forecasting and simulating time series with state-space models;
@@ -99,11 +99,13 @@ model = local_level(y)
 \end{lstlisting}
 %
 \noindent where \texttt{y} is an observation vector in the univariate case or a matrix in column-wise fashion in the multivariate case, i.e., each column representing a variable and each line representing a time period.
+
+While very simple, the local level model can be successfully applied to a wide array of problems wherein one wants to estimate some magnitude that varies over time. For instance, in \cite{koopman2000stamp}, an illustrative example utilizes water inflow data from the river Nile.
 %
 
 \subsubsection{Linear trend model}
 
-The linear trend model consists of a local level model with the addition of a stochastic slope term, as seen in Eq. \eqref{eq:lineartrend1}--\eqref{eq:lineartrend3}. A linear trend \texttt{StateSpaceModel} can be created using:
+The linear trend model consists of a local level model with the addition of a stochastic slope term, as seen in Eq. \eqref{eq:lineartrend1}--\eqref{eq:lineartrend3}. Note that it differs from the local level model due to the inclusion of the slope $\nu_{t}$ in \eqref{eq:lineartrend2}. A linear trend \texttt{StateSpaceModel} can be created using:
 %
 \begin{lstlisting}[language = Julia]
 model = linear_trend(y)
@@ -120,11 +122,13 @@ The basic structural model consists of three stochastic components: trend, slope
     \gamma_{t+1} &= \sum_{j=1}^{s-1} \gamma_{t+1-j} + \omega_{t}, \quad \omega_{t} \sim N(0, \sigma^{2}_\omega). \label{eq:structural4}
 \end{align}
 %
-In this model, there is a stochastic trend that contains a slope component, similarly to the linear trend model. Furthermore, a stochastic seasonal component is modeled so that its sum over the seasonality period $s$ is equal to zero except for a noise factor. A basic structural \texttt{StateSpaceModel} can be created using:
+In this model, there is a stochastic trend that contains a slope component, exactly as the linear trend model. Furthermore, a stochastic seasonal component is modeled such that its sum over the seasonality period $s$ is equal to zero except for a noise factor. A basic structural \texttt{StateSpaceModel} can be created using:
 %
 \begin{lstlisting}[language = Julia]
 model = structural(y, s)
 \end{lstlisting}
+
+The idea of the seasonal component is to account for effects that have a periodic nature, such as monthly temperature over each year.
 
 Additionally, a structural model can contain exogenous variables, also known as explanatory variables. Exogenous variables represent external factors that are correlated to the phenomenon in question and can, thus, be used to improve the estimation and forecasting by increasing the available information. This model is identical to \eqref{eq:structural1}--\eqref{eq:structural4} except for the observation equation, which has an additional exogenous factor:
 %
[…]
