### 2020-02-20 #28 @raphaelsaavedra
Hi @dehann, thank you for your feedback.

Please find below my answers to your comments:

* Regarding the notation of the noise matrix: we have tried to adhere as much as possible to the notation of Durbin & Koopman's book, which was our main reference in this work. In this case, `R` is already reserved for the matrix multiplying the measurement errors.
* By "derive" I mean simply "compute". We have changed it to compute to make it clearer.
* We apologize but your comment regarding Sec. 4.2 wasn't clear to us. In Sec. 4.2, we discuss the estimation procedure and state that we are trying to estimate the error covariances by maximising the likelihood. Since the likelihood is a product of the filter, we need to run the filter at each step of the optimisation procedure until we attain optimality.
* Thank you for pointing out that a reference was missing. We have added the following reference: Helmut Lütkepohl and Fang Xu. The role of the log transformation in forecasting economic variables. Empirical Economics, 42(3):619–638, 2012.
* By quantiles, we mean actual quantiles of the simulated scenarios. Please note that we have separate functions for forecasting and simulating. When forecasting, the confidence intervals are indeed computed based on the variances as you mentioned. However, when simulating, errors are randomly sampled from their estimated distributions at each step and each scenario has temporal dependency so that they are coherent. Therefore, it is a slightly different procedure and quantiles are computed rather than confidence intervals.
* I would love to add a citation to the plotting library but there is no Plots.jl paper. I could just reference the repository but listing the correct authors is tricky.

Your other suggestions that weren't explicitly addressed above were implemented in the paper.

Furthermore, please note that since the initial pdf was generated, we have added a new section regarding basic statistical diagnostics.

We hope you're satisfied with the new version. If you require any further changes, do not hesitate to ask 🙂
