### 2020-02-20 #28 @dehann
Done thanks, one more suggestion is to add links to the demo data in the Docs too (already in Readme).
