+++ paper/paper.tex
@@ -346,7 +346,7 @@ scenarios = simulate(ss, N, S)
 
 \section{Applications} \label{sec:applications}
 
-In this section, we present several practical applications which can be addressed via state-space modeling. We utilize \texttt{StateSpaceModels.jl} to tackle the presented problems and provide the results. Furthermore, for the sake of reproducibility, the code of some of the examples is in the folder \texttt{examples} in the \texttt{StateSpaceModels.jl} repository.
+In this section, we present several practical applications which can be addressed via state-space modeling. We utilize \texttt{StateSpaceModels.jl} to tackle the presented problems and provide the results. Furthermore, for the sake of reproducibility, the code of some of the examples is in the folder \texttt{examples} in the \texttt{StateSpaceModels.jl} repository \cite{statespacemodels}.
 
 \subsection{Airline passengers}
 
@@ -377,7 +377,7 @@ pred, dist = forecast(ss, N)
 \begin{figure}[h]
 	\centering
 	\includegraphics[width=\columnwidth]{Figures/log_airline_forecast.pdf}
-	\caption{Log-airline passengers historical data and forecasting.}
+	\caption{Log-airline passengers historical data and forecast for the following two years.}
 	\label{fig:log_airline_forecast}
 \end{figure}
 %
