### 2020-02-25 #28 @raphaelsaavedra
Thanks for the suggestion. We have added reference [14] to the repo in the Applications section, where we mention that the code and data for examples can be found in the `examples` folder.
