### 2020-03-19 20471c3
Add asserts and additional tests
