### 2020-03-19 2de7aea
fixing assertions
