### 2020-03-24 3c98f1d
Add asserts and additional tests (#153)
* Add reference to repo

* Add asserts and additional tests

* fixing assertions
