### 2020-04-08 7092a5f
Add nile example (#154)
* add nile example

* update docs
