### 2020-04-24 e2f36cc
Add 1.4 and nightly to CI
