+++ .travis.yml
@@ -5,6 +5,12 @@ julia:
   - 1.1
   - 1.2
   - 1.3
+  - 1.4
+  - nightly
+  
+matrix:
+  allow_failures:
+    - julia: nightly
 
 os:
   - linux
