### 2020-06-08 #28 @matbesancon
Hi @mschauer, we had some delays on the JuliaCon  proceedings handling, would you be ok with continuing this review?
