+++ src/kalman.jl
@@ -1,154 +1,0 @@
-"""
-    kalman_filter(model::StateSpaceModel{Typ}; tol::Typ = Typ(1e-5)) where Typ
-
-Kalman filter with big Kappa initialization, i.e., initializing state variances as 1e6.
-"""
-function kalman_filter(model::StateSpaceModel{Typ}; tol::Typ = Typ(1e-5)) where Typ
-
-    time_invariant = model.mode == "time-invariant"
-
-    # Predictive state and its covariance
-    a = Matrix{Typ}(undef, model.dim.n+1, model.dim.m)
-    P = Array{Typ, 3}(undef, model.dim.m, model.dim.m, model.dim.n+1)
-    att = Matrix{Typ}(undef, model.dim.n, model.dim.m)
-    Ptt = Array{Typ, 3}(undef, model.dim.m, model.dim.m, model.dim.n)
-
-    # Innovation and its covariance
-    v = Matrix{Typ}(undef, model.dim.n, model.dim.p)
-    F = Array{Typ, 3}(undef, model.dim.p, model.dim.p, model.dim.n)
-
-    # Kalman gain
-    K = Array{Typ, 3}(undef, model.dim.m, model.dim.p, model.dim.n)
-    
-    # Auxiliary structures
-    ZP = Array{Typ, 2}(undef, model.dim.p, model.dim.m)
-    P_Ztransp_invF = Array{Typ, 2}(undef, model.dim.m, model.dim.p)
-    RQR = Array{Typ, 2}(undef, model.dim.m, model.dim.m)
-
-    # Steady state initialization
-    steadystate = false
-    tsteady     = model.dim.n+1
-
-    # Initial state: big Kappa initialization
-    fill_a1!(a)
-    fill_P1!(P; bigkappa = Typ(1e6))
-
-    RQR = model.R * LinearAlgebra.BLAS.gemm('N', 'T', Typ(1.0), model.Q, model.R) # RQR = R Q R'
-    # Kalman filter
-    for t = 1:model.dim.n
-        if t in model.missing_observations
-            steadystate  = false
-            v[t, :]      = fill(NaN, model.dim.p)
-            F[:, :, t]   = fill(NaN, (model.dim.p, model.dim.p))
-            update_att!(att, a, t) # att_t = a_t
-            update_Ptt!(Ptt, P, t) # Ptt_t = P_t
-            update_a!(a, att, model.T, model.c, t) # a_t+1 = T * att_t + c_t
-            update_P!(P, model.T, Ptt, RQR, t) # P_t+1 = T * Ptt_t * T' + RQR'
-        elseif steadystate
-            update_v!(v, model.y, model.Z, model.d, a, t) # v_t = y_t - Z_t * a_t - d_t
-            repeat_matrix_t_plus_1!(F, t-1) # F[:, :, t]   = F[:, :, t-1]
-            repeat_matrix_t_plus_1!(K, t-1) # K[:, :, t]   = K[:, :, t-1]
-            update_att!(att, a, P_Ztransp_invF, v, t) # att_t = a_t + P_t * Z_t * F^-1_t * v_t
-            repeat_matrix_t_plus_1!(Ptt, t-1) # Ptt_t = Ptt_t-1
-            update_a!(a, att, model.T, model.c, t) # a_t+1 = T * att_t + c_t
-            repeat_matrix_t_plus_1!(P, t) # P_t+1 = P_t
-        else
-            update_v!(v, model.y, model.Z, model.d, a, t) # v_t = y_t - Z_t * a_t - d_t
-            update_ZP!(ZP, model.Z, P, t) # ZP = Z[:, :, t] * P[:, :, t]
-            update_F!(F, ZP, model.Z, model.H, t) # F_t = Z_t * P_t * Z_t + H
[…]
