+++ src/smoothers/kalman_smoother.jl
@@ -1,0 +1,5 @@
+function kalman_smoother!(smoother_output::SmootherOutput, 
+                          model.system, 
+                          filter_output::FilterOutput)
+
+end
