+++ src/structures.jl
@@ -1,262 +1,0 @@
-export StateSpaceDimensions, StateSpaceModel, SmoothedState,
-        FilterOutput, StateSpace
-
-export KalmanFilter, SquareRootFilter, UnivariateKalmanFilter
-
-# Abstract types
-"""
-    AbstractFilter
-
-Abstract type used to implement an interface for generic filters.
-"""
-abstract type AbstractFilter end
-
-"""
-    AbstractSmoother
-
-Abstract type used to implement an interface for generic smoothers.
-"""
-abstract type AbstractSmoother end
-
-"""
-    AbstractOptimizationMethod{T}
-
-Abstract type used to implement an interface for generic optimization methods.
-"""
-abstract type AbstractOptimizationMethod{T} end
-
-# Auxiliary structure for square-root Kalman filter
-mutable struct SquareRootFilter{T <: Real} <: AbstractFilter
-    a::Matrix{T} # predictive state
-    v::Matrix{T} # innovations
-    sqrtP::Array{T, 3} # lower triangular matrix with sqrt-covariance of the predictive state
-    sqrtF::Array{T, 3} # lower triangular matrix with sqrt-covariance of the innovations
-    steadystate::Bool # flag that indicates if steady state was attained
-    tsteady::Int # instant when steady state was attained; in case it wasn't, tsteady = n+1
-    K::Array{T, 3} # Kalman gain
-end
-
-# Auxiliary structure for Kalman filter
-mutable struct KalmanFilter{T <: Real} <: AbstractFilter
-    a::Matrix{T} # predictive state
-    att::Matrix{T}
-    v::Matrix{T} # innovations
-    P::Array{T, 3} # covariance matrix of the predictive state
-    Ptt::Array{T, 3}
-    F::Array{T, 3} # covariance matrix of the innovations
-    steadystate::Bool # flag that indicates if steady state was attained
-    tsteady::Int # instant when steady state was attained; in case it wasn't, tsteady = n+1
-    K::Array{T, 3} # Kalman gain
-end
-
-# Auxiliary structure for univariate Kalman filter
-mutable struct UnivariateKalmanFilter{T <: Real} <: AbstractFilter
-    a::Matrix{T} # predictive state
-    att::Matrix{T}
-    v::Vector{T} # innovations
-    P::Array{T, 3} # covariance matrix of the predictive state
-    Ptt::Array{T, 3}
-    F::Vector{T} # covariance matrix of the innovations
-    steadystate::Bool # flag that indicates if steady state was attained
-    tsteady::Int # instant when steady state was attained; in case it wasn't, tsteady = n+1
-    K::Matrix{T} # Kalman gain
-end
-
-# Auxiliary structure for smoother
-mutable struct Smoother{T <: Real} <: AbstractSmoother
-    alpha::Matrix{T} # smoothed state
-    V::Array{T, 3} # variance of smoothed state
-end
-
-"""
-    StateSpaceDimensions
-
-StateSpaceModel dimensions, following the notation of on the book
-\"Time Series Analysis by State Space Methods\" (2012) by J. Durbin and S. J. Koopman.
-
-* `n` is the number of observations
-* `p` is the dimension of the observation vector ``y_t``
-* `m` is the dimension of the state vector ``\\alpha_t``
-* `r` is the dimension of the state covariance matrix ``Q``
-"""
-struct StateSpaceDimensions
-    n::Int
[…]
