### 2020-07-14 188218c
add local level tests
