### 2020-07-14 dc74083
add local level tests (#159)
