### 2020-07-14 ae1fd6f
add more tests to the local level model
