### 2020-07-15 053a94e
add more tests to the local level model (#160)
