### 2020-07-15 dc6ed1a
Added tests for basic structural and linear regression
