### 2020-07-15 ed8ebf9
Added tests for basic structural and linear regression (#161)
