+++ docs/src/manual.md
@@ -1,0 +1,16 @@
+# Manual
+
+## Models
+
+The package provides a variaty of pre-defined models, fell free to contribute if you want to add 
+one to the list.
+
+### Implementing a custom StateSpaceModel
+
+...
+
+### Local Level
+
+```@docs
+LocalLevel
+```
