+++ Project.toml
@@ -4,6 +4,7 @@ authors = ["raphaelsaavedra <raphael.saavedra93@gmail.com>, guilhermebodin <guil
 version = "0.4.0"
 
 [deps]
+DelimitedFiles = "8bb1440f-4735-579b-a4ab-409b98df4dab"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Optim = "429524aa-4258-5aef-a3af-852621145aeb"
