+++ src/models/common.jl
@@ -1,3 +1,6 @@
+export is_valid_statespacemodel
+
+
 """
 Unobserved component models are state space models 
 that you can name components, each component has a proper meaning
@@ -7,27 +10,29 @@ abstract type UnobservedComponentModel <: StateSpaceModel end
 typeof_model_elements(model::StateSpaceModel) = eltype(model.system.y)
 
 """
+TODO
 """
-get_num_states(model::StateSpaceModel) = get_num_states(get_system(model))
+num_states(model::StateSpaceModel) = num_states(system(model))
 
 """
+TODO
 """
-get_system(model::StateSpaceModel) = model.system
+system(model::StateSpaceModel) = model.system
 
 """
 TODO
 
 Verifica se o model tem todas as funções implementadas
 """
-function validate_statespacemodel(model::StateSpaceModel)
-    # TODO colocar mensagem de erro caso não tenha os métodos necessários
-    tuple_with_model_type = Tuple{typeof(model)}
+function is_valid_statespacemodel(model_type::Type{<:StateSpaceModel})
+    # TODO error message to indicate which method is missing
+    tuple_with_model_type = Tuple{model_type}
     hasmethod(default_filter, tuple_with_model_type)
     hasmethod(initial_hyperparameters!, tuple_with_model_type)
     hasmethod(constraint_hyperparameters!, tuple_with_model_type)
     hasmethod(unconstraint_hyperparameters!, tuple_with_model_type)
     hasmethod(update!, tuple_with_model_type)
-    return 
+    return true
 end
 
 # standard print 
