+++ src/hyperparameters.jl
@@ -181,6 +181,7 @@ end
 
 # Some special constraint functions
 """
+TODO
 """
 function constrain_box(model::StateSpaceModel, str::String, lb::Fl, ub::Fl) where Fl
     update_constrained_value!(model, str, lb + ((ub - lb)/(1 + exp(-get_unconstrained_value(model, str)))))
@@ -188,6 +189,7 @@ function constrain_box(model::StateSpaceModel, str::String, lb::Fl, ub::Fl) wher
 end
 
 """
+TODO
 """
 function unconstrain_box(model::StateSpaceModel, str::String, lb::Fl, ub::Fl) where Fl
     update_unconstrained_value!(model, str, -log((ub - lb)/(get_constrained_value(model, str) - lb - 1)))
@@ -195,6 +197,7 @@ function unconstrain_box(model::StateSpaceModel, str::String, lb::Fl, ub::Fl) wh
 end
 
 """
+TODO
 """
 function constrain_variance(model::StateSpaceModel, str::String)
     update_constrained_value!(model, str, (get_unconstrained_value(model, str))^2)
@@ -202,6 +205,7 @@ function constrain_variance(model::StateSpaceModel, str::String)
 end
 
 """
+TODO
 """
 function unconstrain_variance(model::StateSpaceModel, str::String)
     update_unconstrained_value!(model, str, sqrt(get_constrained_value(model, str)))
@@ -209,6 +213,7 @@ function unconstrain_variance(model::StateSpaceModel, str::String)
 end
 
 """
+TODO
 """
 function constrain_identity(model::StateSpaceModel, str::String)
     update_constrained_value!(model, str, get_unconstrained_value(model, str))
@@ -216,6 +221,7 @@ function constrain_identity(model::StateSpaceModel, str::String)
 end
 
 """
+TODO
 """
 function unconstrain_identity(model::StateSpaceModel, str::String)
     update_unconstrained_value!(model, str, get_constrained_value(model, str))
