+++ test/models/locallevel.jl
@@ -1,6 +1,9 @@
-@testset "LocalLevel" begin 
-    y = StateSpaceModels.NILE
-    model = LocalLevel(y)
+@testset "LocalLevel" begin
+    nile = read_csv(StateSpaceModels.NILE)
+
+    @assert is_valid_statespacemodel(LocalLevel)
+
+    model = LocalLevel(nile.flow)
     fit(model)
     @test loglike(model) ≈ -632.5376 atol = 1e-5 rtol = 1e-5
 
@@ -8,7 +11,7 @@
     a1 = 0.0
     P1 = 1e7
     scalar_filter = ScalarKalmanFilter(a1, P1, 1)
-    model = LocalLevel(y)
+    model = LocalLevel(nile.flow)
     fit(model; filter = scalar_filter)
 
     # Without the concentrated filter and score calculation this is close enough
@@ -26,8 +29,8 @@
     @test get_constrained_value(model, "sigma2_η") ≈ 1469.1 atol = 1
 
     # Estimate with Float32
-    y = Float32.(StateSpaceModels.NILE)
-    model = LocalLevel(y)
+    nile32 = Float32.(nile.flow)
+    model = LocalLevel(nile32)
     fit(model)
     @test loglike(model) ≈ -632.53766f0 atol = 1e-5 rtol = 1e-5
 end
