+++ src/StateSpaceModels.jl
@@ -7,7 +7,6 @@ import Base.show
 using LinearAlgebra
 using Statistics
 using Optim
-using DelimitedFiles
 
 include("datasets/datasets.jl")
 
@@ -19,6 +18,8 @@ include("kalman_filter_and_smoother.jl")
 include("filters/univariate_kalman_filter.jl")
 include("filters/scalar_kalman_filter.jl")
 include("filters/regression_kalman_filter.jl")
+
+include("smoothers/kalman_smoother.jl")
 
 include("models/common.jl")
 include("models/locallevel.jl")
