+++ docs/src/manual.md
@@ -1,16 +1,19 @@
 # Manual
 
+## Systems
+
+## Hyperparameters
+
 ## Models
 
-The package provides a variaty of pre-defined models, fell free to contribute if you want to add 
-one to the list.
+The package provides a variaty of pre-defined models, fell free to contribute if you want to add one to the list.
 
 ### Implementing a custom StateSpaceModel
-
-...
 
 ### Local Level
 
 ```@docs
 LocalLevel
 ```
+
+## Optim interface

[…]
