### 2020-07-17 a1dc186
Add docs, smoother and datasets (#162)
* Added tests for basic structural and linear regression

* add first docs

* added smoother
