+++ src/models/basicstructural.jl
@@ -35,9 +35,9 @@ end
 
 function default_filter(model::BasicStructural{Fl}) where Fl
     steadystate_tol = 1e-5
-    a1 = zeros(Fl, get_num_states(model))
-    P1 = 1e6 .* Matrix{Fl}(I, get_num_states(model), get_num_states(model))
-    return UnivariateKalmanFilter(a1, P1, get_num_states(model), steadystate_tol)
+    a1 = zeros(Fl, num_states(model))
+    P1 = 1e6 .* Matrix{Fl}(I, num_states(model), num_states(model))
+    return UnivariateKalmanFilter(a1, P1, num_states(model), steadystate_tol)
 end
 function initial_hyperparameters!(model::BasicStructural{Fl}) where Fl
     initial_hyperparameters = Dict{String, Fl}(
