+++ src/systems.jl
@@ -1,5 +1,7 @@
 abstract type StateSpaceSystem end
 
+"""
+"""
 mutable struct LinearUnivariateTimeInvariant{Fl <: Real} <: StateSpaceSystem
     y::Vector{Fl}
     Z::Vector{Fl}
@@ -31,8 +33,10 @@ mutable struct LinearUnivariateTimeInvariant{Fl <: Real} <: StateSpaceSystem
     end
 end
 
-get_num_states(system::LinearUnivariateTimeInvariant) = size(system.T, 1)
+num_states(system::LinearUnivariateTimeInvariant) = size(system.T, 1)
 
+"""
+"""
 mutable struct LinearUnivariateTimeVariant{Fl <: Real} <: StateSpaceSystem
     y::Vector{Fl}
     Z::Vector{Vector{Fl}}
@@ -54,24 +58,132 @@ mutable struct LinearUnivariateTimeVariant{Fl <: Real} <: StateSpaceSystem
     end
 end
 
-get_num_states(system::LinearUnivariateTimeVariant) = size(system.T[1], 1)
+num_states(system::LinearUnivariateTimeVariant) = size(system.T[1], 1)
+
+"""
+"""
+mutable struct LinearMultivariateTimeInvariant{Fl <: Real} <: StateSpaceSystem
+    y::Matrix{Fl}
+    Z::Matrix{Fl}
+    T::Matrix{Fl}
+    R::Matrix{Fl}
+    d::Vector{Fl}
+    c::Vector{Fl}
+    H::Matrix{Fl}
+    Q::Matrix{Fl}
+
+    function LinearMultivariateTimeInvariant{Fl}(y::Matrix{Fl}, Z::Matrix{Fl}, 
+                                                   T::Matrix{Fl}, R::Matrix{Fl}, 
+                                                   d::Vector{Fl}, c::Vector{Fl}, H::Matrix{Fl},
+                                                   Q::Matrix{Fl}) where Fl
+
+        # TODO assert dimensions
+
+        return new{Fl}(y, Z, T, R, d, c, H, Q)
+    end
+end
+
+"""
+TODO
+"""
+mutable struct LinearMultivariateTimeVariant{Fl <: Real} <: StateSpaceSystem
+    y::Matrix{Fl}
+    Z::Vector{Matrix{Fl}}
+    T::Vector{Matrix{Fl}}
+    R::Vector{Matrix{Fl}}
+    d::Vector{Vector{Fl}}
+    c::Vector{Vector{Fl}}
+    H::Vector{Matrix{Fl}}
+    Q::Vector{Matrix{Fl}}
+
+    function LinearMultivariateTimeVariant{Fl}(y::Matrix{Fl}, Z::Vector{Matrix{Fl}}, 
+                                                   T::Vector{Matrix{Fl}}, R::Vector{Matrix{Fl}}, 
+                                                   d::Vector{Vector{Fl}}, c::Vector{Vector{Fl}}, H::Vector{Matrix{Fl}},
+                                                   Q::Vector{Matrix{Fl}}) where Fl
+
+        # TODO assert dimensions
+
+        return new{Fl}(y, Z, T, R, d, c, H, Q)
+    end
+end
+
+# Some useful functions for dealing with these systems
 function fill_system_matrice_with_value_in_time(vector::Vector{Fl}, value::Fl) where Fl
     for t in axes(vector, 1)
         vector[t] = value
     end
 end
 
-
-# Functions to bridge StateSpaceSystems
-"""
-TODO
-"""
-function to_time_variant(system::LinearUnivariateTimeInvariant)
-
+function repeat_system_matrice_n_times(value::T, n::Int) where T
[…]

[…]
