+++ src/models/basicstructural.jl
@@ -2,9 +2,10 @@ export BasicStructural
 
 """
 """
-mutable struct BasicStructural{Fl} <: StateSpaceModel
-    hyperparameters::HyperParameters{Fl}
-    system::LinearUnivariateTimeInvariant{Fl}
+mutable struct BasicStructural <: StateSpaceModel
+    hyperparameters::HyperParameters
+    system::LinearUnivariateTimeInvariant
+    seasonality::Int
 
     function BasicStructural(y::Vector{Fl}, s::Int) where Fl
 
@@ -29,17 +30,19 @@ mutable struct BasicStructural{Fl} <: StateSpaceModel
         names = ["sigma2_ε", "sigma2_μ", "sigma2_β", "sigma2_γ"]
         hyperparameters = HyperParameters{Fl}(names)
 
-        return new{Fl}(hyperparameters, system)
+        return new(hyperparameters, system, s)
     end
 end
 
-function default_filter(model::BasicStructural{Fl}) where Fl
-    steadystate_tol = 1e-5
+function default_filter(model::BasicStructural)
+    Fl = typeof_model_elements(model)
+    steadystate_tol = Fl(1e-5)
     a1 = zeros(Fl, num_states(model))
     P1 = 1e6 .* Matrix{Fl}(I, num_states(model), num_states(model))
     return UnivariateKalmanFilter(a1, P1, num_states(model), steadystate_tol)
 end
-function initial_hyperparameters!(model::BasicStructural{Fl}) where Fl
+function initial_hyperparameters!(model::BasicStructural)
+    Fl = typeof_model_elements(model)
     initial_hyperparameters = Dict{String, Fl}(
         "sigma2_ε" => var(model.system.y),
         "sigma2_μ" => var(model.system.y),
@@ -49,24 +52,27 @@ function initial_hyperparameters!(model::BasicStructural{Fl}) where Fl
     set_initial_hyperparameters!(model, initial_hyperparameters)
     return
 end
-function constraint_hyperparameters!(model::BasicStructural{Fl}) where Fl
+function constraint_hyperparameters!(model::BasicStructural)
     constrain_variance(model, "sigma2_ε")
     constrain_variance(model, "sigma2_μ")
     constrain_variance(model, "sigma2_β")
     constrain_variance(model, "sigma2_γ")
     return
 end
-function unconstraint_hyperparameters!(model::BasicStructural{Fl}) where Fl
+function unconstraint_hyperparameters!(model::BasicStructural)
     unconstrain_variance(model, "sigma2_ε")
     unconstrain_variance(model, "sigma2_μ")
     unconstrain_variance(model, "sigma2_β")
     unconstrain_variance(model, "sigma2_γ")
     return
 end
-function update!(model::BasicStructural{Fl}) where Fl
+function fill_model_system!(model::BasicStructural)
     model.system.H = get_constrained_value(model, "sigma2_ε")
     model.system.Q[1] = get_constrained_value(model, "sigma2_μ")
     model.system.Q[5] = get_constrained_value(model, "sigma2_β")
     model.system.Q[end] = get_constrained_value(model, "sigma2_γ")
     return 
+end
+function reinstantiate(model::BasicStructural, y::Vector{Fl}) where Fl
+    return BasicStructural(y, model.seasonality)
 end
