+++ test/models/basicstructural.jl
@@ -7,4 +7,7 @@
     model = BasicStructural(log_air_passengers, 12)
     fit(model)
     @test loglike(model) ≈ 234.33641 atol = 1e-5 rtol = 1e-5
+
+    # forecasting
+    forec = forecast(model, 10)
 end
