### 2020-07-21 4ff2a56
implement suggestions
