+++ src/models/linear_regression.jl
@@ -66,14 +66,14 @@ function initial_hyperparameters!(model::LinearRegression{Fl}) where Fl
     set_initial_hyperparameters!(model, initial_hyperparameters)
     return
 end
-function constraint_hyperparameters!(model::LinearRegression{Fl}) where {Fl}
+function constrain_hyperparameters!(model::LinearRegression{Fl}) where {Fl}
     for i in 1:num_states(model)
         constrain_identity(model, get_beta_name(model, i))
     end
     constrain_variance(model, "sigma2_ε")
     return
 end
-function unconstraint_hyperparameters!(model::LinearRegression{Fl}) where Fl
+function unconstrain_hyperparameters!(model::LinearRegression{Fl}) where Fl
     for i in 1:num_states(model)
         unconstrain_identity(model, get_beta_name(model, i))
     end
