+++ Project.toml
@@ -20,4 +20,4 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "CSV", "Statistics", "DataFrames", "Random"]
+test = ["Test", "CSV", "DataFrames", "Random"]
