### 2020-07-21 84a1498
Working example (#163)
* Working example

* implement suggestions

* fix Project.toml
