+++ src/filters/scalar_kalman_filter.jl
@@ -113,20 +113,22 @@ end
 
 function scalar_update_kalman_state!(kalman_state, y, Z, T, H, RQR, 
                                      d, c, skip_llk_instants, tol, t)
-    if kalman_state.steady_state
-        scalar_update_v!(kalman_state, y, Z, d)
-        scalar_update_att!(kalman_state, Z)
-        scalar_update_a!(kalman_state, T, c)
-        if t > skip_llk_instants
-            update_llk!(kalman_state)
-        end
-    elseif isnan(y)
+
+    if isnan(y)
         kalman_state.v = NaN
         scalar_update_F!(kalman_state, Z, H)
         kalman_state.att = kalman_state.a
         scalar_update_a!(kalman_state, T, c)
         kalman_state.Ptt = kalman_state.P
         scalar_update_P!(kalman_state, T, RQR)
+        kalman_state.steady_state = false # Not on steadystate anymore
+    elseif kalman_state.steady_state
+        scalar_update_v!(kalman_state, y, Z, d)
+        scalar_update_att!(kalman_state, Z)
+        scalar_update_a!(kalman_state, T, c)
+        if t > skip_llk_instants
+            update_llk!(kalman_state)
+        end
     else
         scalar_update_v!(kalman_state, y, Z, d)
         scalar_update_F!(kalman_state, Z, H)
