+++ src/models/lineartrend.jl
@@ -39,13 +39,13 @@ function initial_hyperparameters!(model::LinearTrend{Fl}) where Fl
     set_initial_hyperparameters!(model, initial_hyperparameters)
     return
 end
-function constraint_hyperparameters!(model::LinearTrend{Fl}) where {Fl}
+function constrain_hyperparameters!(model::LinearTrend{Fl}) where {Fl}
     update_constrained_value!(model, "sigma2_ε", get_unconstrained_value(model, "sigma2_ε")^2)
     update_constrained_value!(model, "sigma2_η", get_unconstrained_value(model, "sigma2_η")^2)
     update_constrained_value!(model, "sigma2_β", get_unconstrained_value(model, "sigma2_β")^2)
     return
 end
-function unconstraint_hyperparameters!(model::LinearTrend{Fl}) where Fl
+function unconstrain_hyperparameters!(model::LinearTrend{Fl}) where Fl
     update_unconstrained_value!(model, "sigma2_ε", sqrt(get_constrained_value(model, "sigma2_ε")))
     update_unconstrained_value!(model, "sigma2_η", sqrt(get_constrained_value(model, "sigma2_η")))
     update_unconstrained_value!(model, "sigma2_β", sqrt(get_constrained_value(model, "sigma2_β")))
