+++ src/kalman_filter_and_smoother.jl
@@ -29,11 +29,11 @@ end
 function update_model_hyperparameters!(model::StateSpaceModel, 
                                        unconstrained_hyperparameters::Vector{Fl}) where Fl
     register_unconstrained_values!(model, unconstrained_hyperparameters)
-    constraint_hyperparameters!(model)
-    update!(model)
+    constrain_hyperparameters!(model)
+    fill_model_system!(model)
     return nothing
 end
-# @time StateSpaceModels2.
+
 function update_filter_hyperparameters!(filter::KalmanFilter, model::StateSpaceModel)
     update!(filter, model)
     return nothing
