+++ Project.toml
@@ -18,6 +18,8 @@ CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+
 
 [targets]
-test = ["Test", "CSV", "DataFrames", "Random"]
+test = ["Test", "CSV", "DataFrames", "Documenter", "Random"]
