### 2020-07-23 57a7b5b
Update docs/src/examples.md
Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
