### 2020-07-23 1253b1c
Update src/optimizers.jl
Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
