### 2020-07-23 6c64567
Update src/optimizers.jl
Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
