### 2020-07-23 ee4b0bb
Updated some docs (#164)
* updated some docs

* fix wrong using Plots

* Update docs/src/examples.md

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* Update src/optimizers.jl

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* Update src/optimizers.jl

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* update datasets

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
