+++ docs/make.jl
@@ -18,6 +18,6 @@ makedocs(
     ]
 )
 
-# deploydocs(
-#     repo = "github.com/LAMPSPUC/StateSpaceModels.jl.git",
-# )
+deploydocs(
+    repo = "github.com/LAMPSPUC/StateSpaceModels.jl.git",
+)
