### 2020-08-03 6a0b725
update functions to have the ! syntax
