+++ src/models/linear_regression.jl
@@ -68,16 +68,16 @@ function initial_hyperparameters!(model::LinearRegression{Fl}) where Fl
 end
 function constrain_hyperparameters!(model::LinearRegression{Fl}) where {Fl}
     for i in 1:num_states(model)
-        constrain_identity(model, get_beta_name(model, i))
+        constrain_identity!(model, get_beta_name(model, i))
     end
-    constrain_variance(model, "sigma2_ε")
+    constrain_variance!(model, "sigma2_ε")
     return
 end
 function unconstrain_hyperparameters!(model::LinearRegression{Fl}) where Fl
     for i in 1:num_states(model)
-        unconstrain_identity(model, get_beta_name(model, i))
+        unconstrain_identity!(model, get_beta_name(model, i))
     end
-    unconstrain_variance(model, "sigma2_ε")
+    unconstrain_variance!(model, "sigma2_ε")
     return
 end
 function fill_model_system!(model::LinearRegression{Fl}) where Fl
