### 2020-08-03 e1533eb
updates (#165)
* updates

* Update src/models/common.jl

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* Update src/systems.jl

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* Update src/systems.jl

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* update functions to have the ! syntax

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
