+++ src/fit.jl
@@ -11,18 +11,13 @@ function fit(model::StateSpaceModel,
              optimizer::Optimizer)
     
     initial_unconstrained_hyperparameter = handle_optim_initial_hyperparameters(model)
-    # Optimization
-    # try
-        func = TwiceDifferentiable(x -> -optim_loglike(model, filter, x), 
-                                        initial_unconstrained_hyperparameter)
-        opt = optimize(func, initial_unconstrained_hyperparameter, 
-                       optimizer.method, optimizer.options)
-        opt_loglikelihood = -opt.minimum
-        opt_hyperparameters = opt.minimizer
-    # catch err
-    #     println(err)
-    #     println("seed diverged")
-    # end
+    # Optimization a try catch is missing
+    func = TwiceDifferentiable(x -> -optim_loglike(model, filter, x), 
+                                    initial_unconstrained_hyperparameter)
+    opt = optimize(func, initial_unconstrained_hyperparameter, 
+                    optimizer.method, optimizer.options)
+    opt_loglikelihood = -opt.minimum
+    opt_hyperparameters = opt.minimizer
 
     log_lik = opt_loglikelihood
     update_model_hyperparameters!(model, opt_hyperparameters)
