+++ src/models/linear_regression.jl
@@ -68,16 +68,16 @@ function initial_hyperparameters!(model::LinearRegression{Fl}) where Fl
 end
 function constrain_hyperparameters!(model::LinearRegression{Fl}) where {Fl}
     for i in 1:num_states(model)
-        constrain_identity(model, get_beta_name(model, i))
+        constrain_identity!(model, get_beta_name(model, i))
     end
-    constrain_variance(model, "sigma2_ε")
+    constrain_variance!(model, "sigma2_ε")
     return
 end
 function unconstrain_hyperparameters!(model::LinearRegression{Fl}) where Fl
     for i in 1:num_states(model)
-        unconstrain_identity(model, get_beta_name(model, i))
+        unconstrain_identity!(model, get_beta_name(model, i))
     end
-    unconstrain_variance(model, "sigma2_ε")
+    unconstrain_variance!(model, "sigma2_ε")
     return
 end
 function fill_model_system!(model::LinearRegression{Fl}) where Fl
@@ -86,7 +86,7 @@ function fill_model_system!(model::LinearRegression{Fl}) where Fl
     fill_H_in_time(model, H)
     return 
 end
-function update!(filter::KalmanFilter, model::LinearRegression{Fl}) where Fl
+function fill_model_filter!(filter::KalmanFilter, model::LinearRegression{Fl}) where Fl
     for i in axes(filter.kalman_state.a, 1)
         filter.kalman_state.a[i] = get_constrained_value(model, get_beta_name(model, i))
     end
