+++ test/runtests.jl
@@ -14,7 +14,12 @@ include("utils.jl")
 @testset "Models" begin
     include("models/locallevel.jl")
     include("models/basicstructural.jl")
+    include("models/arima.jl")
     include("models/linear_regression.jl")
+end
+
+@testset "Systems" begin
+    include("systems.jl")
 end
 
 @testset "Documentation examples" begin
