### 2020-08-12 d57dbd2
Update docs/src/examples.md
Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
