### 2020-08-13 b78c44d
simulate observations (#172)
* simulate observations

* Update src/systems.jl

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* Update docs/src/examples.md

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* updaate size(T, 1)

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
