### 2020-08-28 #28 @raphaelsaavedra
Any updates here? I see the proceedings are already out (https://proceedings.juliacon.org/).
