+++ .travis.yml
@@ -1,38 +1,31 @@
 language: julia
 
-julia:
-  - 1.0
-  # - 1.1
-  # - 1.2
-  # - 1.3
-  - 1.4
-  # - nightly
-  
-# matrix:
-#   allow_failures:
-#     - julia: nightly
-
 os:
   - linux
+  - osx
 
-after_success:
-  - julia --project -e 'import Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(Codecov.process_folder())'
-  
-notifications:
-    email: false
-
-branches:
-  only:
-    - master
-    - refactoring-to-0.4
+julia:
+  - 1.0 # Latest LTS release
+  - 1.4
+  - 1   # Latest stable release
+  - nightly
 
 jobs:
+  allow_failures:
+    - julia: nightly
   include:
     - stage: "Documentation"
-      julia: 1.0
+      julia: 1
       os: linux
-      script:
-        - julia --project=docs/ -e 'using Pkg; Pkg.instantiate();
-                                    Pkg.develop(PackageSpec(path=pwd()))'
-        - julia --project=docs/ docs/make.jl
+      script: julia --project=docs -e '
+          using Pkg;
+          Pkg.develop(PackageSpec(path=pwd()));
+          Pkg.instantiate();
+          include("docs/make.jl");'
       after_success: skip
+
+notifications:
+  email: false
+
+after_success:
+  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(process_folder())'
