### 2020-09-03 ed7b108
Changing data structure
