+++ test/models/basicstructural.jl
@@ -1,9 +1,9 @@
 @testset "Basic Structural Model" begin
-    air_passengers = read_csv(StateSpaceModels.AIR_PASSENGERS)
+    air_passengers = CSV.read(StateSpaceModels.AIR_PASSENGERS, DataFrame)
     log_air_passengers = log.(air_passengers.passengers)
 
     @test has_fit_methods(BasicStructural)
-  
+
     model = BasicStructural(log_air_passengers, 12)
     StateSpaceModels.fit(model)
     @test loglike(model) ≈ 234.33641 atol = 1e-5 rtol = 1e-5
