+++ test/models/locallevel.jl
@@ -1,5 +1,5 @@
 @testset "LocalLevel" begin
-    nile = read_csv(StateSpaceModels.NILE)
+    nile = CSV.read(StateSpaceModels.NILE, DataFrame)
 
     @test has_fit_methods(LocalLevel)
 
@@ -24,7 +24,7 @@
 
     hyperparameters = get_hyperparameters(model)
     @test !isempty(get_minimizer_hyperparameter_position(hyperparameters))
-    
+
     @test loglike(model; filter = scalar_filter) ≈  -632.54421 atol = 1e-5 rtol = 1e-5
     @test get_constrained_value(model, "sigma2_η") ≈ 1469.1 atol = 1
 
@@ -34,7 +34,7 @@
     StateSpaceModels.fit(model)
     @test loglike(model) ≈ -632.53766f0 atol = 1e-5 rtol = 1e-5
 
-    # Missing values 
+    # Missing values
     nile.flow[[collect(21:40); collect(61:80)]] .= NaN
     model = LocalLevel(nile.flow)
     StateSpaceModels.fit(model)
