+++ test/models/locallevel.jl
@@ -12,7 +12,7 @@
     P1 = 1e7
     scalar_filter = ScalarKalmanFilter(a1, P1, 1)
     model = LocalLevel(nile.flow)
-    StateSpaceModels.fit(model; filter = scalar_filter)
+    StateSpaceModels.fit(model; filter=scalar_filter)
 
     # Without the concentrated filter and score calculation this is close enough
     @test get_constrained_value(model, "sigma2_ε") ≈ 15099 atol = 2
@@ -20,12 +20,12 @@
 
     # Fix some parameters
     fix_hyperparameters!(model, Dict("sigma2_ε" => 15099.0))
-    StateSpaceModels.fit(model; filter = scalar_filter)
+    StateSpaceModels.fit(model; filter=scalar_filter)
 
     hyperparameters = get_hyperparameters(model)
     @test !isempty(get_minimizer_hyperparameter_position(hyperparameters))
     
-    @test loglike(model; filter = scalar_filter) ≈  -632.54421 atol = 1e-5 rtol = 1e-5
+    @test loglike(model; filter=scalar_filter) ≈  -632.54421 atol = 1e-5 rtol = 1e-5
     @test get_constrained_value(model, "sigma2_η") ≈ 1469.1 atol = 1
 
     # Estimate with Float32
@@ -53,5 +53,6 @@
     @test monotone_forecast_variance(forec)
     # simulating
     scenarios = simulate_scenarios(model, 10, 10_000)
-    test_scenarios_adequacy_with_forecast(forec, scenarios, 5.0, 50.0)
+    test_scenarios_adequacy_with_forecast(forec, scenarios)
+
 end
