+++ test/models/basicstructural.jl
@@ -13,5 +13,5 @@
     @test monotone_forecast_variance(forec)
     # simualting
     scenarios = simulate_scenarios(model, 10, 10_000)
-    test_scenarios_adequacy_with_forecast(forec, scenarios, 0.2, 0.2)
+    test_scenarios_adequacy_with_forecast(forec, scenarios)
 end
