### 2020-09-05 f33e628
Add LocalLinearTrend with tests and example
