+++ docs/src/manual.md
@@ -7,10 +7,7 @@ The package provides a variaty of pre-defined models, fell free to contribute if
 ### Local Level
 ```@docs
 LocalLevel
-```
-
-### BasicStructural
-```@docs
+LocalLinearTrend
 BasicStructural
 ```
 
@@ -44,4 +41,6 @@ datasets using [DataFrames.jl](https://github.com/JuliaData/DataFrames.jl) and [
 StateSpaceModels.NILE
 StateSpaceModels.AIR_PASSENGERS
 StateSpaceModels.INTERNET
+StateSpaceModels.WHOLESALE_PRICE_INDEX
+StateSpaceModels.VEHICLE_FATALITIES
 ```

[…]
