### 2020-09-07 6479634
Fix has_fit_methods
