+++ test/models/basicstructural.jl
@@ -12,6 +12,6 @@
     forec = forecast(model, 10)
     @test monotone_forecast_variance(forec)
     # simualting
-    scenarios = simulate_scenarios(model, 10, 100_000)
+    scenarios = simulate_scenarios(model, 10, 10_000)
     test_scenarios_adequacy_with_forecast(forec, scenarios)
 end
