### 2020-09-08 fe44953
Fix remaining instances of update!
