### 2020-09-08 2f43f94
Fix has_fit_methods (#180)
* Fix has_fit_methods

* Update common.jl

* Fix remaining instances of update!
