### 2020-09-08 0e1238c
Fixed expectation bug for local level and structural model (#177)
* :bug: Fixed expectation bug for local level and structural model

* uncomment test

* update tests

* update tests
