+++ Project.toml
@@ -10,7 +10,7 @@ Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
-Optim = "0.20, 0.21, 0.22"
+Optim = "1.1"
 julia = "1"
 
 [extras]
