### 2020-09-11 d1e2cef
add fit! method
