+++ src/fit.jl
@@ -3,9 +3,7 @@ export fit!
 """
 TODO
 """
-function fit! end
-
-function fit!(model::StateSpaceModel, 
+function fit!(model::StateSpaceModel; 
              filter::KalmanFilter = default_filter(model),
              optimizer::Optimizer = Optimizer(Optim.LBFGS()))
     
@@ -18,8 +16,7 @@ function fit!(model::StateSpaceModel,
     opt_loglikelihood   = -opt.minimum
     opt_hyperparameters = opt.minimizer
     update_model_hyperparameters!(model, opt_hyperparameters)
-    # You must calculate the numerical hessian on the constrained values
-    numerical_hessian = hessian!(func, copy(get_unconstrained_values(model)))
+    numerical_hessian = Optim.hessian!(func, opt_hyperparameters)
     std_err = diag(inv(numerical_hessian))
     fill_results!(model, opt_loglikelihood, std_err)
     return nothing
@@ -31,11 +28,11 @@ function fill_results!(model::StateSpaceModel,
     n_obs = length(model.system.y)
     num_hyperparameters = number_free_hyperparameters(model)
     coef_table = build_coef_table(model, std_err)
-    info_crit = build_information_criterion(num_hyperparameters, n_obs, llk)
     # Fill results
     model.results.coef_table          = coef_table
-    model.results.info_criterion      = info_crit
     model.results.llk                 = llk
+    model.results.aic                 = AIC(num_hyperparameters, llk)
+    model.results.bic                 = BIC(n_obs, num_hyperparameters, llk)
     model.results.num_observations    = n_obs
     model.results.num_hyperparameters = num_hyperparameters
     model.results.fitted              = true
@@ -44,14 +41,6 @@ end
 
 AIC(n_free_hyperparameters::Int, llk::Fl) where Fl = Fl(2 * n_free_hyperparameters - 2 * llk)
 BIC(n::Int, n_free_hyperparameters::Int, llk::Fl) where Fl = Fl(log(n) * n_free_hyperparameters - 2 * llk)
-function build_information_criterion(n_free_hyperparameters::Int,
-                                     n_obs::Int,
-                                     llk::Fl) where Fl
-    return InformationCriterion{Fl}(
-                                AIC(n_free_hyperparameters, llk),
-                                BIC(n_obs, n_free_hyperparameters, llk)
-                            )
-end
 
 function build_coef_table(model::StateSpaceModel,
                           std_err::Vector{Fl}) where Fl
