+++ test/models/locallevel.jl
@@ -5,7 +5,6 @@
 
     model = StateSpaceModels.LocalLevel(nile.flow)
     StateSpaceModels.fit!(model)
-    model.results
     @test loglike(model) ≈ -632.5376 atol = 1e-5 rtol = 1e-5
 
     # Durbin Koopman 2012 section 2.2.5
@@ -13,7 +12,7 @@
     P1 = 1e7
     scalar_filter = ScalarKalmanFilter(a1, P1, 1)
     model = LocalLevel(nile.flow)
-    StateSpaceModels.fit(model; filter=scalar_filter)
+    StateSpaceModels.fit!(model; filter=scalar_filter)
 
     # Without the concentrated filter and score calculation this is close enough
     @test get_constrained_value(model, "sigma2_ε") ≈ 15099 atol = 2
@@ -21,7 +20,7 @@
 
     # Fix some parameters
     fix_hyperparameters!(model, Dict("sigma2_ε" => 15099.0))
-    StateSpaceModels.fit(model; filter=scalar_filter)
+    StateSpaceModels.fit!(model; filter=scalar_filter)
 
     hyperparameters = get_hyperparameters(model)
     @test !isempty(get_minimizer_hyperparameter_position(hyperparameters))
@@ -32,13 +31,13 @@
     # Estimate with Float32
     nile32 = Float32.(nile.flow)
     model = LocalLevel(nile32)
-    StateSpaceModels.fit(model)
+    StateSpaceModels.fit!(model)
     @test loglike(model) ≈ -632.53766f0 atol = 1e-5 rtol = 1e-5
 
     # Missing values 
     nile.flow[[collect(21:40); collect(61:80)]] .= NaN
     model = LocalLevel(nile.flow)
-    StateSpaceModels.fit(model)
+    StateSpaceModels.fit!(model)
     @test loglike(model) ≈ -379.9899 atol = 1e-5 rtol = 1e-5
 
     filter = kalman_filter(model)
