+++ test/models/basicstructural.jl
@@ -5,7 +5,7 @@
     @test has_fit_methods(BasicStructural)
   
     model = BasicStructural(log_air_passengers, 12)
-    StateSpaceModels.fit(model)
+    StateSpaceModels.fit!(model)
     # Runned on Python statsmodels
     @test loglike(model) ≈ 234.33641 atol = 1e-5 rtol = 1e-5
 
