### 2020-09-11 85c0ca1
Add fit! funtion (#184)
* add fit! method

* update

* rewrite fit

* updates
