### 2020-09-11 7bc1d2d
Add coefficients table
