+++ test/models/locallevel.jl
@@ -14,6 +14,10 @@
     Log-likelihood:               -632.54
     AIC:                          1269.08
     BIC:                          1274.29
+    --------------------------------------------------------
+    Parameter      Estimate   Std.Error     z stat   p-value
+    sigma2_ε       15108.33      164.21      92.01      0.00
+    sigma2_η        1463.16      279.88       5.23      0.00
     """
     @test sprint(show, results(model)) == result_print
     @test loglike(model) ≈ -632.5376 atol = 1e-5 rtol = 1e-5
@@ -32,6 +36,20 @@
     # Fix some parameters
     fix_hyperparameters!(model, Dict("sigma2_ε" => 15099.0))
     fit!(model; filter=scalar_filter)
+    result_print = 
+    """                         Results                        
+    ========================================================
+    Number of observations:       100
+    Number of unknown parameters: 1
+    Log-likelihood:               -632.54
+    AIC:                          1267.09
+    BIC:                          1269.69
+    --------------------------------------------------------
+    Parameter      Estimate   Std.Error     z stat   p-value
+    sigma2_ε       15099.00          -          -         - 
+    sigma2_η        1468.29      175.18       8.38      0.00
+    """
+    @test sprint(show, results(model)) == result_print
 
     hyperparameters = get_hyperparameters(model)
     @test !isempty(get_minimizer_hyperparameter_position(hyperparameters))
