+++ src/models/common.jl
@@ -22,17 +22,17 @@ function Results{Fl}() where Fl
 end
 results(model::StateSpaceModel) = model.results
 function Base.isempty(results::Results)
-    return isnan(results.llk) && 
-           isnan(results.aic) && 
+    return isnan(results.llk) &&
+           isnan(results.aic) &&
            isnan(results.bic) &&
-           iszero(results.num_observations) && 
+           iszero(results.num_observations) &&
            iszero(results.num_hyperparameters)
 end
 
 """
     has_fit_methods(model_type::Type{<:StateSpaceModel}) -> Bool
 
-Verify if a certain `StateSpaceModel` has the necessary methods to perform the fit!.
+Verify if a certain `StateSpaceModel` has the necessary methods to perform `fit!``.
 """
 function has_fit_methods(model_type::Type{<:StateSpaceModel})
     tuple_with_model_type = Tuple{model_type}
