+++ README.md
@@ -35,7 +35,7 @@ Current features include:
 
 ## Citing StateSpaceModels.jl
 
-If you use StateSpaceModels.jl in your work, we kindly ask you to cite the following paper ([pdf](https://arxiv.org/pdf/1908.01757.pdf)):
+If you use StateSpaceModels.jl in your work, we kindly ask you to cite the [following paper](https://arxiv.org/abs/1908.01757):
 
     @article{SaavedraBodinSouto2019,
       title={StateSpaceModels.jl: a Julia Package for Time-Series Analysis in a State-Space Framework},
