### 2020-09-15 6574189
Address some todo's
