### 2020-09-15 fd88e73
Improve performance of any
