### 2020-09-15 4ae2bbc
Add ARIMA to readme
