+++ README.md
@@ -30,6 +30,7 @@ Current features include:
   * Basic structural model (trend, slope, seasonal)
   * Linear trend model
   * Local level model
+  * ARIMA
 * Completion of missing values
 * Diagnostics for the residuals of fitted models
 
