### 2020-09-15 eae9a27
Address some TODO's  (#188)
* Update readme

* Add docstrings

* Update readme

* Update readme

* Address some todo's

* Improve performance of any

* Add ARIMA to readme
