+++ src/filters/scalar_kalman_filter.jl
@@ -1,7 +1,10 @@
 export ScalarKalmanFilter
 
-# This method is similar to the univariate kalman filter but
-# exploits the fact that the dimension of the state is 1
+"""
+    ScalarKalmanState{Fl <: AbstractFloat}
+
+The state of a `ScalarKalmanFilter`.
+"""
 mutable struct ScalarKalmanState{Fl <: AbstractFloat}
     v::Fl
     F::Fl
@@ -15,19 +18,19 @@ mutable struct ScalarKalmanState{Fl <: AbstractFloat}
     P_to_check_steady_state::Fl
     function ScalarKalmanState(a1::Fl, P1::Fl) where Fl
         P_to_check_steady_state = zero(Fl)
-        return new{Fl}(zero(Fl), zero(Fl), zero(Fl), 
+        return new{Fl}(zero(Fl), zero(Fl), zero(Fl),
                        a1, zero(Fl), P1, zero(Fl), INITIAL_STEADY_STATE,
                        P_to_check_steady_state)
     end
 end
 
-function save_a1_P1_in_filter_output!(filter_output::FilterOutput{Fl}, 
+function save_a1_P1_in_filter_output!(filter_output::FilterOutput{Fl},
                                       kalman_state::ScalarKalmanState{Fl}) where Fl
     filter_output.a[1] = copy(fill(kalman_state.a, 1))
     filter_output.P[1] = copy(fill(kalman_state.P, 1, 1))
     return
 end
-function save_kalman_state_in_filter_output!(filter_output::FilterOutput{Fl}, 
+function save_kalman_state_in_filter_output!(filter_output::FilterOutput{Fl},
                                              kalman_state::ScalarKalmanState{Fl},
                                              t::Int) where Fl
     filter_output.v[t] = copy(fill(kalman_state.v, 1))
@@ -41,8 +44,12 @@ function save_kalman_state_in_filter_output!(filter_output::FilterOutput{Fl},
     return
 end
 
-# Univariate Kalman filter with the recursions as described 
-# in Koopman's book TODO
+"""
+    ScalarKalmanFilter{Fl <: Real} <: KalmanFilter
+
+Similar to the univariate Kalman filter but exploits the fact that the dimension of the
+state is equal to 1.
+"""
 mutable struct ScalarKalmanFilter{Fl <: Real} <: KalmanFilter
     steadystate_tol::Fl
     a1::Fl
@@ -55,14 +62,14 @@ mutable struct ScalarKalmanFilter{Fl <: Real} <: KalmanFilter
                                 skip_llk_instants::Int = 1,
                                 steadystate_tol::Fl = Fl(1e-5)) where Fl
         kalman_state = ScalarKalmanState(a1, P1)
-        return new{Fl}(steadystate_tol, a1, P1, 
+        return new{Fl}(steadystate_tol, a1, P1,
                        skip_llk_instants, kalman_state)
     end
 end
 
 function reset_filter!(kf::ScalarKalmanFilter{Fl}) where Fl
     reset_filter!(kf.kalman_state, kf.a1, kf.P1)
-    return 
+    return
 end
 function reset_filter!(kalman_state::ScalarKalmanState, a1::Fl, P1::Fl) where Fl
     kalman_state.a = a1
@@ -96,7 +103,7 @@ function scalar_update_att!(kalman_state::ScalarKalmanState{Fl}, Z::Fl) where Fl
 end
 function scalar_update_a!(kalman_state::ScalarKalmanState{Fl}, T::Fl, c::Fl) where Fl
     kalman_state.a = T * kalman_state.att + c
-    return 
+    return
 end
[…]
