### 2020-09-15 edbf224
Fix docs Project.toml
