### 2020-09-15 cd82858
Fix docs Project.toml (#190)
* Fix docs Project.toml

* Add Plots
