### 2020-09-15 d5c8252
Add coefficients table (#187)
* Add coefficients table

* fix tests

* remove print test
