+++ src/models/common.jl
@@ -4,7 +4,41 @@ typeof_model_elements(model::StateSpaceModel) = eltype(model.system.y)
 num_states(model::StateSpaceModel) = num_states(system(model))
 system(model::StateSpaceModel) = model.system
 
+struct CoefficientTable{Fl <: AbstractFloat}
+    names::Vector{String}
+    coef::Vector{Fl}
+    std_err::Vector{Fl}
+    z::Vector{Fl}
+    p_value::Vector{Fl}
+
+    function CoefficientTable{Fl}(names::Vector{String}, 
+                                coef::Vector{Fl}, 
+                                std_err::Vector{Fl}, 
+                                z::Vector{Fl}, 
+                                p_value::Vector{Fl}) where Fl
+        @assert length(names) == length(coef) == length(std_err) == length(z) == length(p_value)
+        return new{Fl}(names, coef, std_err, z, p_value)
+    end
+end
+function CoefficientTable{Fl}() where Fl
+    return CoefficientTable{Fl}(
+                        String[],
+                        Fl[],
+                        Fl[],
+                        Fl[],
+                        Fl[])
+end
+Base.length(coef_table::CoefficientTable) = length(coef_table.names)
+function Base.isempty(coef_table::CoefficientTable)
+    return isempty(coef_table.names) &&
+        isempty(coef_table.coef) &&
+        isempty(coef_table.std_err) &&
+        isempty(coef_table.z) &&
+        isempty(coef_table.p_value)
+end
+
 mutable struct Results{Fl <: AbstractFloat}
+    coef_table::CoefficientTable{Fl}
     llk::Fl
     aic::Fl
     bic::Fl
@@ -13,6 +47,7 @@ mutable struct Results{Fl <: AbstractFloat}
 end
 function Results{Fl}() where Fl
     return Results{Fl}(
+        CoefficientTable{Fl}(),
         Fl(NaN),
         Fl(NaN),
         Fl(NaN),
@@ -22,8 +57,9 @@ function Results{Fl}() where Fl
 end
 results(model::StateSpaceModel) = model.results
 function Base.isempty(results::Results)
-    return isnan(results.llk) &&
-           isnan(results.aic) &&
+    return isempty(results.coef_table) &&
+           isnan(results.llk) && 
+           isnan(results.aic) && 
            isnan(results.bic) &&
            iszero(results.num_observations) &&
            iszero(results.num_hyperparameters)
