+++ test/models/locallineartrend.jl
@@ -17,5 +17,5 @@
     predicted_mean = [5.897660, 5.862062, 5.826465, 5.790867,
                       5.755270, 5.719672, 5.684074, 5.648477,
                       5.612879, 5.577282]
-    @test predicted_mean ≈ vcat(forec.expected_value...) atol=1e-3
+    @test maximum(abs.(predicted_mean .- vcat(forec.expected_value...))) < 1.5e-3
 end
