+++ src/filters/scalar_kalman_filter.jl
@@ -1,5 +1,3 @@
-export ScalarKalmanFilter
-
 """
     ScalarKalmanState{Fl <: AbstractFloat}
 
