### 2020-09-15 176958c
Remove get minimizer position
