+++ src/forecast.jl
@@ -1,7 +1,3 @@
-export forecast,
-    forecast_expected_value,
-    simulate_scenarios
-
 mutable struct Forecast{Fl}
     expected_value::Vector{Vector{Fl}}
     covariance::Vector{Matrix{Fl}}
