### 2020-09-15 f993dfe
Adjust functions
