### 2020-09-16 36cc8f1
Adjust remaining filter/smoother functions
