+++ src/hyperparameters.jl
@@ -173,6 +173,19 @@ function fix_hyperparameters!(hyperparameters::HyperParameters{Fl}) where Fl
     return
 end
 
+"""
+    isfitted(model::StateSpaceModel) -> Bool
+
+Verify if model is fitted, i.e., returns `true` if the `results` field is nonempty or if
+all the hyperparameters are fixed and `false` otherwise.
+"""
+function isfitted(model::StateSpaceModel)
+    if isempty(model.results) && number_fixed_hyperparameters(model) == number_hyperparameters(model)
+        return false
+    else
+        return true
+    end
+end
 
 # Some special constraint functions
 """
