+++ src/kalman_filter_and_smoother.jl
@@ -71,68 +71,96 @@ mutable struct FilterOutput{Fl<:Real}
 end
 
 """
+    get_innovations
+
 Returns the innovations `v` obtained with the Kalman filter.
 """
 function get_innovations end
 
 get_innovations(filter::FilterOutput) = permutedims(cat(filter.v...; dims=2))
 
-function get_innovations(model::StateSpaceModel)
-    isempty(model.results) && error("Model has not been estimated yet, please use `fit!`.")
-    return get_innovations(kalman_filter(model))
+function get_innovations(model::StateSpaceModel; filter=default_filter(model))
+    isfitted(model) && error("Model has not been estimated yet, please use `fit!`.")
+    return get_innovations(kalman_filter(model; filter=filter))
 end
 
 """
+    get_innovation_variance
+
 Returns the variance `F` of innovations obtained with the Kalman filter.
 """
 function get_innovation_variance end
 
 get_innovation_variance(filter_output::FilterOutput) = cat(filter_output.F...; dims=3)
 
-function get_innovation_variance(model::StateSpaceModel)
-    isempty(model.results) && error("Model has not been estimated yet, please use `fit!`.")
-    return get_innovation_variance(kalman_filter(model))
+function get_innovation_variance(model::StateSpaceModel; filter=default_filter(model))
+    isfitted(model) && error("Model has not been estimated yet, please use `fit!`.")
+    return get_innovation_variance(kalman_filter(model; filter=filter))
 end
 
 """
+    get_filtered_state
+
 Returns the filtered state `att` obtained with the Kalman filter.
 """
 function get_filtered_state end
 
 get_filtered_state(filter::FilterOutput) = permutedims(cat(filter.att...; dims=2))
 
-function get_filtered_state(model::StateSpaceModel)
-    isempty(model.results) && error("Model has not been estimated yet, please use `fit!`.")
-    return get_filtered_state(kalman_filter(model))
+function get_filtered_state(model::StateSpaceModel; filter=default_filter(model))
+    isfitted(model) && error("Model has not been estimated yet, please use `fit!`.")
+    return get_filtered_state(kalman_filter(model; filter=filter))
 end
 
 """
-Returns the filtered covariance `Ptt` of the states obtained with the Kalman filter.
+    get_filtered_variance
+
+Returns the variance `Ptt` of the filtered state obtained with the Kalman filter.
 """
-function get_filtered_covariance end
+function get_filtered_variance end
 
-get_filtered_covariance(filter::FilterOutput) = cat(filter.Ptt...; dims=3)
+get_filtered_variance(filter::FilterOutput) = cat(filter.Ptt...; dims=3)
 
-function get_filtered_covariance(model::StateSpaceModel)
-    isempty(model.results) && error("Model has not been estimated yet, please use `fit!`.")
-    return get_filtered_covariance(kalman_filter(model))
+function get_filtered_variance(model::StateSpaceModel; filter=default_filter(model))
+    isfitted(model) && error("Model has not been estimated yet, please use `fit!`.")
+    return get_filtered_variance(kalman_filter(model; filter=filter))
 end
 
 """
-TODO
+    get_predictive_state
+
+Returns the predictive state `a` obtained with the Kalman filter.
 """
-one_step_ahead_predictions(filter_output::FilterOutput) = permutedims(cat(filter_output.a...; dims = 2))
+function get_predictive_state end
+
+get_predictive_state(filter::FilterOutput) = permutedims(cat(filter.a...; dims=2))
+
+function get_predictive_state(model::StateSpaceModel; filter=default_filter(model))
[…]
