### 2020-09-16 d741096
Fix isfitted error logic
