+++ src/kalman_filter_and_smoother.jl
@@ -80,7 +80,7 @@ function get_innovations end
 get_innovations(filter::FilterOutput) = permutedims(cat(filter.v...; dims=2))
 
 function get_innovations(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) && error("Model has not been estimated yet, please use `fit!`.")
+    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
     return get_innovations(kalman_filter(model; filter=filter))
 end
 
@@ -94,7 +94,7 @@ function get_innovation_variance end
 get_innovation_variance(filter_output::FilterOutput) = cat(filter_output.F...; dims=3)
 
 function get_innovation_variance(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) && error("Model has not been estimated yet, please use `fit!`.")
+    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
     return get_innovation_variance(kalman_filter(model; filter=filter))
 end
 
@@ -108,7 +108,7 @@ function get_filtered_state end
 get_filtered_state(filter::FilterOutput) = permutedims(cat(filter.att...; dims=2))
 
 function get_filtered_state(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) && error("Model has not been estimated yet, please use `fit!`.")
+    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
     return get_filtered_state(kalman_filter(model; filter=filter))
 end
 
@@ -122,7 +122,7 @@ function get_filtered_variance end
 get_filtered_variance(filter::FilterOutput) = cat(filter.Ptt...; dims=3)
 
 function get_filtered_variance(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) && error("Model has not been estimated yet, please use `fit!`.")
+    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
     return get_filtered_variance(kalman_filter(model; filter=filter))
 end
 
@@ -136,7 +136,7 @@ function get_predictive_state end
 get_predictive_state(filter::FilterOutput) = permutedims(cat(filter.a...; dims=2))
 
 function get_predictive_state(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) && error("Model has not been estimated yet, please use `fit!`.")
+    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
     return get_predictive_state(kalman_filter(model); filter=filter)
 end
 
@@ -150,7 +150,7 @@ function get_predictive_variance end
 get_predictive_variance(filter::FilterOutput) = cat(filter.P...; dims=3)
 
 function get_predictive_variance(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) && error("Model has not been estimated yet, please use `fit!`.")
+    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
     return get_predictive_variance(kalman_filter(model; filter=filter))
 end
 
@@ -160,7 +160,7 @@ end
 Filter the data using the desired `filter` and the estimated hyperparameters in `model`.
 """
 function kalman_filter(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) && error("Model has not been estimated yet, please use `fit!`.")
+    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
     filter_output = FilterOutput(model)
     reset_filter!(filter)
     free_unconstrained_values = get_free_unconstrained_values(model)
@@ -212,7 +212,7 @@ function get_smoothed_state end
 get_smoothed_state(smoother::SmootherOutput) = permutedims(cat(smoother.alpha...; dims = 2))
 
[…]
