### 2020-09-16 f9b1493
Test getter functions
