### 2020-09-16 3ca85fb
Fix isfitted logic
