### 2020-09-16 12682e6
Changes to public API (#191)
* Move exports to main file

* Remove get minimizer position

* Adjust functions

* More improvements

* Adjust remaining filter/smoother functions

* Fix isfitted error logic

* Test getter functions

* Fix isfitted logic
