+++ src/models/basicstructural.jl
@@ -1,5 +1,3 @@
-export BasicStructural
-
 @doc raw"""
 The basic structural state-space model consists of a trend, a slope, and a seasonal
 components. It is defined by:

[…]
