### 2020-09-17 392fdf2
add forecast for univariate time variant
