+++ src/models/linear_regression.jl
@@ -5,11 +5,12 @@ struct RegressionHyperParametersAuxiliary
     end
 end
 
-mutable struct LinearRegression{Fl <: Real} <: StateSpaceModel
-    hyperparameters::HyperParameters{Fl}
+mutable struct LinearRegression <: StateSpaceModel
+    hyperparameters::HyperParameters
     hyperparameters_auxiliary::RegressionHyperParametersAuxiliary
-    system::LinearUnivariateTimeVariant{Fl}
+    system::LinearUnivariateTimeVariant
     results::Results
+    exogenous::Matrix
 
     function LinearRegression(X::Matrix{Fl}, y::Vector{Fl}) where Fl
 
@@ -34,7 +35,7 @@ mutable struct LinearRegression{Fl <: Real} <: StateSpaceModel
 
         hyperparameters_auxiliary = RegressionHyperParametersAuxiliary(num_states(system))
 
-        return new{Fl}(hyperparameters, hyperparameters_auxiliary, system, Results{Fl}())
+        return new(hyperparameters, hyperparameters_auxiliary, system, Results{Fl}(), X)
     end
 end
 
@@ -49,44 +50,50 @@ get_beta_name(model::LinearRegression, i::Int) = model.hyperparameters_auxiliary
 fill_H_in_time(model::LinearRegression, H::Fl) where Fl = fill_system_matrice_with_value_in_time(model.system.H, H)
 
 # Obligatory functions
-function default_filter(model::LinearRegression{Fl}) where Fl
+function default_filter(model::LinearRegression)
+    Fl = typeof_model_elements(model)
     a1 = zeros(Fl, num_states(model))
     return RegressionKalmanFilter(a1)
 end
-function initial_hyperparameters!(model::LinearRegression{Fl}) where Fl
-    initial_hyperparameters = Dict{String, Float64}(
+function initial_hyperparameters!(model::LinearRegression)
+    Fl = typeof_model_elements(model)
+    initial_hyperparameters = Dict{String, Fl}(
         "sigma2_ε" => var(model.system.y)
     )
     # The optimal regressors are the result of X \ y
-    betas = hcat(model.system.Z...)' \ model.system.y
+    betas = model.exogenous \ model.system.y
     for i in 1:num_states(model)
         initial_hyperparameters[get_beta_name(model, i)] = betas[i]
     end
     set_initial_hyperparameters!(model, initial_hyperparameters)
     return
 end
-function constrain_hyperparameters!(model::LinearRegression{Fl}) where {Fl}
+function constrain_hyperparameters!(model::LinearRegression)
     for i in 1:num_states(model)
         constrain_identity!(model, get_beta_name(model, i))
     end
     constrain_variance!(model, "sigma2_ε")
     return
 end
-function unconstrain_hyperparameters!(model::LinearRegression{Fl}) where Fl
+function unconstrain_hyperparameters!(model::LinearRegression)
     for i in 1:num_states(model)
         unconstrain_identity!(model, get_beta_name(model, i))
     end
     unconstrain_variance!(model, "sigma2_ε")
     return
 end
-function fill_model_system!(model::LinearRegression{Fl}) where Fl
+function fill_model_system!(model::LinearRegression)
     # Fill the same H for every timestamp
     H = get_constrained_value(model, "sigma2_ε")
     fill_H_in_time(model, H)
     return
 end
-function fill_model_filter!(filter::KalmanFilter, model::LinearRegression{Fl}) where Fl
+function fill_model_filter!(filter::KalmanFilter, model::LinearRegression)
     for i in axes(filter.kalman_state.a, 1)
[…]
