+++ test/models/linear_regression.jl
@@ -13,4 +13,23 @@
             @test get_constrained_value(model, "β_$i") .≈ β[i] atol = 1e-5
         end
     end
+    X = [
+        1 1
+        1 2
+        1 3
+        1 4
+        1 5.0
+    ]
+    y = [
+        6
+        7
+        8
+        9
+        10.0
+    ]
+    model = LinearRegression(X, y)
+    fit!(model)
+    forec = forecast(model, [1 6.0])
+    forec.expected_value[1] == [11.0]
+
 end
