+++ src/models/basicstructural.jl
@@ -87,3 +87,4 @@ end
 function reinstantiate(model::BasicStructural, y::Vector{Fl}) where Fl
     return BasicStructural(y, model.seasonality)
 end
+has_exogenous(::BasicStructural) = false
