+++ src/forecast.jl
@@ -24,7 +24,8 @@ function forecast(model::StateSpaceModel, steps_ahead::Int;
     # Query the type of model elements
     Fl = typeof_model_elements(model)
     # Observations to forecast
-    forecasting_y = [model.system.y; fill(NaN, steps_ahead)]
+    forecasting_y = isunivariate(model) ? [model.system.y; fill(NaN, steps_ahead)] :
+                                          [model.system.y; fill(NaN, steps_ahead, size(model.system.y, 2))]
     # Copy hyperparameters
     model_hyperparameters = deepcopy(model.hyperparameters)
     # Instantiate a new model
@@ -37,7 +38,11 @@ function forecast(model::StateSpaceModel, steps_ahead::Int;
     expected_value = Vector{Vector{Fl}}(undef, steps_ahead)
     covariance = Vector{Matrix{Fl}}(undef, steps_ahead)
     for i in 1:steps_ahead
-        expected_value[i] = [dot(model.system.Z, fo.a[end - steps_ahead + i]) + model.system.d]
+        if isunivariate(model)
+            expected_value[i] = [dot(model.system.Z, fo.a[end - steps_ahead + i]) + model.system.d]
+        else
+            expected_value[i] = model.system.Z * fo.a[end - steps_ahead + i] .+ model.system.d
+        end
         covariance[i] = fo.F[end - steps_ahead + i]
     end
     return Forecast{Fl}(expected_value, covariance)
@@ -50,7 +55,8 @@ function forecast(model::StateSpaceModel, new_exogenous::Matrix{Fl};
               "forecast(model::SSM, steps_ahead::Int; kwargs...) where SSM")
     end
     steps_ahead = size(new_exogenous, 1)
-    forecasting_y = [model.system.y; fill(NaN, steps_ahead)]
+    forecasting_y = isunivariate(model) ? [model.system.y; fill(NaN, steps_ahead)] :
+                                          [model.system.y; fill(NaN, steps_ahead, size(model.system.y, 2))]
     forecasting_X = [model.exogenous; new_exogenous]
     # Copy hyperparameters
     model_hyperparameters = deepcopy(model.hyperparameters)
@@ -64,8 +70,13 @@ function forecast(model::StateSpaceModel, new_exogenous::Matrix{Fl};
     expected_value = Vector{Vector{Fl}}(undef, steps_ahead)
     covariance = Vector{Matrix{Fl}}(undef, steps_ahead)
     for i in 1:steps_ahead
-        expected_value[i] = [dot(model.system.Z[end - steps_ahead + i], fo.a[end - steps_ahead + i]) + 
-                                 model.system.d[end - steps_ahead + i]]
+        if isunivariate(model)
+            expected_value[i] = [dot(model.system.Z[end - steps_ahead + i], fo.a[end - steps_ahead + i]) + 
+                                     model.system.d[end - steps_ahead + i]]
+        else
+            expected_value[i] = model.system.Z[end - steps_ahead + i] * fo.a[end - steps_ahead + i] + 
+                                model.system.d[end - steps_ahead + i]
+        end
         covariance[i] = fo.F[end - steps_ahead + i]
     end
     return Forecast{Fl}(expected_value, covariance)
