### 2020-09-19 3ee5574
Basic structural with multivariate option
