+++ src/models/basicstructural.jl
@@ -14,35 +14,74 @@ component. It is defined by:
 """
 mutable struct BasicStructural <: StateSpaceModel
     hyperparameters::HyperParameters
-    system::LinearUnivariateTimeInvariant
+    system::StateSpaceSystem
     seasonality::Int
     results::Results
 
-    function BasicStructural(y::Vector{Fl}, s::Int) where Fl
+    function BasicStructural(y::VecOrMat{Fl}, s::Int) where Fl
+        if isunivariate(y)
+            Z = [1; 0; 1; zeros(Fl, s - 2)]
+            T = [
+                1 1 zeros(Fl, 1, s - 1);
+                0 1 zeros(Fl, 1, s - 1);
+                0 0 -ones(Fl, 1, s - 1);
+                zeros(Fl, s - 2, 2) Matrix{Fl}(I, s - 2, s - 2) zeros(Fl, s - 2)
+            ]
+            R = [
+                Matrix{Fl}(I, 3, 3);
+                zeros(Fl, s - 2, 3)
+            ]
+            d = zero(Fl)
+            c = zeros(Fl, s + 1)
+            H = one(Fl)
+            Q = zeros(Fl, 3, 3)
+            
+            system = LinearUnivariateTimeInvariant{Fl}(y, Z, T, R, d, c, H, Q)
 
-        Z = [1; 0; 1; zeros(Fl, s - 2)]
-        T = [
-            1 1 zeros(Fl, 1, s - 1);
-            0 1 zeros(Fl, 1, s - 1);
-            0 0 -ones(Fl, 1, s - 1);
-            zeros(Fl, s - 2, 2) Matrix{Fl}(I, s - 2, s - 2) zeros(Fl, s - 2)
-        ]
-        R = [
-            Matrix{Fl}(I, 3, 3);
-            zeros(Fl, s - 2, 3)
-        ]
-        d = zero(Fl)
-        c = zeros(Fl, s + 1)
-        H = one(Fl)
-        Q = zeros(Fl, 3, 3)
+            names = ["sigma2_ε", "sigma2_ξ", "sigma2_ζ", "sigma2_ω"]
+            hyperparameters = HyperParameters{Fl}(names)
 
-        system = LinearUnivariateTimeInvariant{Fl}(y, Z, T, R, d, c, H, Q)
-
-        names = ["sigma2_ε", "sigma2_ξ", "sigma2_ζ", "sigma2_ω"]
-        hyperparameters = HyperParameters{Fl}(names)
-
-        return new(hyperparameters, system, s, Results{Fl}())
+            return new(hyperparameters, system, s, Results{Fl}())
+        else
+            p = size(y, 2)
+            Z = kron(Matrix{Fl}(I, p, p), [1 0 1 zeros(1, s - 2)])
+            T = kron(Matrix{Fl}(I, p, p),
+                    [1 1 zeros(Fl, 1, s - 1);
+                     0 1 zeros(Fl, 1, s - 1);
+                     0 0 -ones(Fl, 1, s - 1);
+                    zeros(Fl, s - 2, 2) Matrix{Fl}(I, s - 2, s - 2) zeros(Fl, s - 2)])
+                    
+            R = kron(Matrix{Fl}(I, p, p),
+                    [Matrix{Fl}(I, 3, 3);
+                    zeros(Fl, s - 2, 3)])
+    
+            d = zeros(Fl, p)
+            c = zeros(Fl, p * (s + 1))
+            H = kron(one(Fl), Matrix{Fl}(I, p, p))
+            Q = kron(ones(Fl, p, p), Matrix{Fl}(I, 3, 3))
+    
+            system = LinearMultivariateTimeInvariant{Fl}(y, Z, T, R, d, c, H, Q)
+    
[…]
