### 2020-09-21 9fad25c
add forecast for univariate time variant (#192)
* add forecast for univariate time variant

* change SSM to StateSpaceModel

* [ci skip] update docstring
