+++ src/forecast.jl
@@ -8,12 +8,19 @@ function forecast_expected_value(forec::Forecast)
 end
 
 """
-    forecast(model::SSM, steps_ahead::Int) where SSM
+    forecast(model::StateSpaceModel, steps_ahead::Int; kwargs...)
+    forecast(model::StateSpaceModel, exogenous::Matrix{Fl}; kwargs...) where {Fl}
 
 Forecast the mean and covariance for future observations from a StateSpaceModel (SSM).
 """
-function forecast(model::SSM, steps_ahead::Int;
-                  filter::KalmanFilter = default_filter(model)) where SSM
+function forecast end
+
+function forecast(model::StateSpaceModel, steps_ahead::Int;
+                  filter::KalmanFilter = default_filter(model))
+    if has_exogenous(model)
+        error("The model has exogenous variables, you should use the" *
+              "forecast(model::SSM, exogenous::Matrix{Fl}; kwargs...) method")
+    end
     # Query the type of model elements
     Fl = typeof_model_elements(model)
     # Observations to forecast
@@ -26,7 +33,6 @@ function forecast(model::SSM, steps_ahead::Int;
     forecasting_model.hyperparameters = model_hyperparameters
     # Perform the kalman filter
     fo = kalman_filter(forecasting_model)
-
     # fill forecast matrices
     expected_value = Vector{Vector{Fl}}(undef, steps_ahead)
     covariance = Vector{Matrix{Fl}}(undef, steps_ahead)
@@ -34,21 +40,48 @@ function forecast(model::SSM, steps_ahead::Int;
         expected_value[i] = [dot(model.system.Z, fo.a[end - steps_ahead + i]) + model.system.d]
         covariance[i] = fo.F[end - steps_ahead + i]
     end
+    return Forecast{Fl}(expected_value, covariance)
+end
 
+function forecast(model::StateSpaceModel, new_exogenous::Matrix{Fl};
+                  filter::KalmanFilter = default_filter(model)) where {Fl}
+    if !has_exogenous(model)
+        error("The model does not support exogenous variables, you should use the" *
+              "forecast(model::SSM, steps_ahead::Int; kwargs...) where SSM")
+    end
+    steps_ahead = size(new_exogenous, 1)
+    forecasting_y = [model.system.y; fill(NaN, steps_ahead)]
+    forecasting_X = [model.exogenous; new_exogenous]
+    # Copy hyperparameters
+    model_hyperparameters = deepcopy(model.hyperparameters)
+    # Instantiate a new model
+    forecasting_model = reinstantiate(model, forecasting_y, forecasting_X)
+    # Associate with the model hyperparameters
+    forecasting_model.hyperparameters = model_hyperparameters
+    # Perform the kalman filter
+    fo = kalman_filter(forecasting_model)
+    # fill forecast matrices
+    expected_value = Vector{Vector{Fl}}(undef, steps_ahead)
+    covariance = Vector{Matrix{Fl}}(undef, steps_ahead)
+    for i in 1:steps_ahead
+        expected_value[i] = [dot(model.system.Z[end - steps_ahead + i], fo.a[end - steps_ahead + i]) + 
+                                 model.system.d[end - steps_ahead + i]]
+        covariance[i] = fo.F[end - steps_ahead + i]
+    end
     return Forecast{Fl}(expected_value, covariance)
 end
 
 """
     simulate_scenarios(
-        model::StateSpaceModel, n_periods::Int, n_scenarios::Int;
+        model::StateSpaceModel, steps_ahead::Int, n_scenarios::Int;
         filter::KalmanFilter=default_filter(model)
     ) -> Array{<:AbstractFloat, 3}
 
[…]
