+++ src/models/basicstructural_multivariate.jl
@@ -1,0 +1,364 @@
+@doc raw"""
+An implementation of a non-homogeneous seemingly unrelated time series equations for basic structural 
+state-space model consists of trend (local linear trend) and seasonal components.
+It is defined by:
+```math
+\begin{gather*}
+    \begin{aligned}
+    y_{t} &=  \mu_{t} + \gamma_{t} + \varepsilon_{t} \quad &\varepsilon_{t} \sim \mathcal{N}(0, \Sigma_{\varepsilon})\\
+    \mu_{t+1} &= \mu_{t} + \nu_{t} + \xi_{t} \quad &\xi_{t} \sim \mathcal{N}(0, \Sigma_{\xi})\\
+    \nu_{t+1} &= \nu_{t} + \zeta_{t} \quad &\zeta_{t} \sim \mathcal{N}(0, \Sigma_{\zeta})\\
+    \gamma_{t+1} &= -\sum_{j=1}^{s-1} \gamma_{t+1-j} + \omega_{t} \quad & \omega_{t} \sim \mathcal{N}(0, \Sigma_{\omega})\\
+    \end{aligned}
+\end{gather*}
+```
+"""
+mutable struct MultivariateBasicStructural <: StateSpaceModel
+    hyperparameters::HyperParameters
+    system::LinearMultivariateTimeInvariant
+    seasonality::Int
+    results::Results
+
+    function MultivariateBasicStructural(y::Matrix{Fl}, s::Int) where Fl
+        p = size(y, 2)
+        Z = kron(Matrix{Fl}(I, p, p), [1 0 1 zeros(1, s - 2)])
+        T = kron(Matrix{Fl}(I, p, p),
+                [1 1 zeros(Fl, 1, s - 1);
+                 0 1 zeros(Fl, 1, s - 1);
+                 0 0 -ones(Fl, 1, s - 1);
+                zeros(Fl, s - 2, 2) Matrix{Fl}(I, s - 2, s - 2) zeros(Fl, s - 2)])
+                
+        R = kron(Matrix{Fl}(I, p, p),
+                [Matrix{Fl}(I, 3, 3);
+                zeros(Fl, s - 2, 3)])
+
+        d = zeros(Fl, p)
+        c = zeros(Fl, p * (s + 1))
+        H = kron(one(Fl), Matrix{Fl}(I, p, p))
+        Q = kron(ones(Fl, p, p), Matrix{Fl}(I, 3, 3))
+
+        system = LinearMultivariateTimeInvariant{Fl}(y, Z, T, R, d, c, H, Q)
+
+        names = handle_multivariate_names(p)
+        hyperparameters = HyperParameters{Fl}(names)
+        return new(hyperparameters, system, s, Results{Fl}())
+    end
+end
+
+function default_filter(model::MultivariateBasicStructural)
+    Fl = typeof_model_elements(model)
+    steadystate_tol = Fl(1e-5)
+    a1 = zeros(Fl, num_states(model))
+    P1 = 1e6 .* Matrix{Fl}(I, num_states(model), num_states(model))
+    return MultivariateKalmanFilter(size(model.system.y, 2), a1, P1, num_states(model), steadystate_tol)
+end
+function initial_hyperparameters!(model::MultivariateBasicStructural)
+    Fl = typeof_model_elements(model)
+    initial_hyperparameters = Dict{String, Fl}()
+    for variable in get_names(model)
+        if occursin("sigma2", variable)
+            initial_hyperparameters[variable] = one(Fl)
+        else
+            initial_hyperparameters[variable] = zero(Fl)
+        end
+    end
+    set_initial_hyperparameters!(model, initial_hyperparameters)
+    return
+end
[…]
