### 2020-09-22 a964cb6
Fix single precision default filter
