+++ src/models/locallineartrend.jl
@@ -50,7 +50,7 @@ function default_filter(model::LocalLinearTrend)
     Fl = typeof_model_elements(model)
     steadystate_tol = 1e-5
     a1 = zeros(Fl, 2)
-    P1 = 1e6 .* Matrix{Fl}(I, 2, 2)
+    P1 = Fl(1e6) .* Matrix{Fl}(I, 2, 2)
     return UnivariateKalmanFilter(a1, P1, 2, steadystate_tol)
 end
 function initial_hyperparameters!(model::LocalLinearTrend)
