+++ src/models/damped_lineartrend.jl
@@ -25,7 +25,7 @@ mutable struct DampedLinearTrend{Fl} <: StateSpaceModel
 end
 
 function default_filter(::DampedLinearTrend{Fl}) where Fl
-    steadystate_tol = 1e-5
+    steadystate_tol = Fl(1e-5)
     a1 = zeros(Fl, 2)
     P1 = Fl(1e6) .* Matrix{Fl}(I, 2, 2)
     return UnivariateKalmanFilter(a1, P1, 0, steadystate_tol)
