### 2020-09-30 84c3e74
Multivariate structural (#194)
* add forecast for univariate time variant

* change SSM to StateSpaceModel

* bivariate-structural

* Basic structural with multivariate option

* update multivariate

* update multivariate

* change function name
