+++ src/filters/multivariate_kalman_filter.jl
@@ -1,12 +1,11 @@
 """
-    UnivariateKalmanState{Fl <: AbstractFloat}
+    MultivariateKalmanState{Fl <: AbstractFloat}
 
-A Kalman filter that is tailored to univariate systems, exploiting the fact that the
-dimension of the observations at any time period is 1.
+A Kalman filter that is tailored to multivariate systems.
 """
-mutable struct UnivariateKalmanState{Fl <: AbstractFloat}
-    v::Fl
-    F::Fl
+mutable struct MultivariateKalmanState{Fl <: AbstractFloat}
+    v::Vector{Fl}
+    F::Matrix{Fl}
     att::Vector{Fl}
     a::Vector{Fl}
     Ptt::Matrix{Fl}
@@ -15,33 +14,30 @@ mutable struct UnivariateKalmanState{Fl <: AbstractFloat}
     steady_state::Bool
     # Auxiliary matrices
     P_to_check_steady_state::Matrix{Fl}
-    ZP::Vector{Fl}
     TPtt::Matrix{Fl}
-    function UnivariateKalmanState(a1::Vector{Fl}, P1::Matrix{Fl}) where Fl
+    function MultivariateKalmanState(num_series::Int, a1::Vector{Fl}, P1::Matrix{Fl}) where Fl
         m = length(a1)
         P_to_check_steady_state = zeros(Fl, m, m)
-        ZP = zeros(Fl, m)
         TPtt = zeros(Fl, m, m)
 
-        return new{Fl}(zero(Fl), zero(Fl), zeros(Fl, m),
+        return new{Fl}(zeros(Fl, num_series), zeros(Fl, num_series, num_series), zeros(Fl, m),
                        a1, zeros(Fl, m, m), P1, zero(Fl), false,
                        P_to_check_steady_state,
-                       ZP,
                        TPtt)
     end
 end
 
 function save_a1_P1_in_filter_output!(filter_output::FilterOutput{Fl},
-                                      kalman_state::UnivariateKalmanState{Fl}) where Fl
+                                      kalman_state::MultivariateKalmanState{Fl}) where Fl
     filter_output.a[1] = deepcopy(kalman_state.a)
     filter_output.P[1] = deepcopy(kalman_state.P)
     return
 end
 function save_kalman_state_in_filter_output!(filter_output::FilterOutput{Fl},
-                                             kalman_state::UnivariateKalmanState{Fl},
+                                             kalman_state::MultivariateKalmanState{Fl},
                                              t::Int) where Fl
-    filter_output.v[t] = copy(fill(kalman_state.v, 1))
-    filter_output.F[t] = copy(fill(kalman_state.F, 1, 1))
+    filter_output.v[t] = copy(kalman_state.v)
+    filter_output.F[t] = copy(kalman_state.F)
     filter_output.a[t + 1] = copy(kalman_state.a)
     filter_output.att[t] = copy(kalman_state.att)
     filter_output.P[t + 1] = copy(kalman_state.P)
@@ -51,41 +47,42 @@ function save_kalman_state_in_filter_output!(filter_output::FilterOutput{Fl},
     return
 end
 
-# Univariate Kalman filter with the recursions as described
-# in Koopman's book TODO
-mutable struct UnivariateKalmanFilter{Fl <: Real} <: KalmanFilter
+"""
+TODO
+"""
+mutable struct MultivariateKalmanFilter{Fl <: Real} <: KalmanFilter
     steadystate_tol::Fl
     a1::Vector{Fl}
     P1::Matrix{Fl}
     skip_llk_instants::Int
-    kalman_state::UnivariateKalmanState
[…]
