### 2020-10-01 085186b
Refactor forecast output
