### 2020-10-01 e08edef
Update src/forecast.jl
Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
