### 2020-10-01 019a303
Refactor forecast output (#200)
* Refactor forecast output

* Update src/forecast.jl

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* Bump minor release

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
