### 2020-10-01 6b1e2a5
Bump version to 0.5-DEV (#202)
