+++ src/utils.jl
@@ -1,233 +1,0 @@
-export size, ztr, copy, statespace_recursion
-
-"""
-    size(model::StateSpaceModel)
-
-Return the dimensions `n`, `p`, `m` and `r` of the `StateSpaceModel`
-"""
-function size(model::StateSpaceModel)
-    return model.dim.n, model.dim.p, model.dim.m, model.dim.r
-end
-
-"""
-    ztr(model::StateSpaceModel)
-
-Return the state space model arrays `Z`, `T` and `R` of the `StateSpaceModel`
-"""
-function ztr(model::StateSpaceModel)
-    return model.Z, model.T, model.R
-end
-
-"""
-    statespace_recursion(model::StateSpaceModel{Typ}, N::Int, initial_a::Matrix{Typ}) where Typ
-
-Runs the recursion of a state space model ignoring the y.
-"""
-function statespace_recursion(model::StateSpaceModel{Typ}, initial_a::Matrix{Typ}) where Typ
-
-    # If model has any unknows is not possible to perform the recursion
-    unknowns = Unknowns(model)
-    if unknowns.n_unknowns != 0
-        error("StateSpaceModel has unknown parameters.")
-    end
-
-    # Save the distributions H and Q
-    dist_H = MvNormal(model.H)
-    dist_Q = MvNormal(model.Q)
-
-    p, m, n = size(model.Z)
-
-    if size(initial_a, 2) != m
-        error("intial_a must be a 1 by $m matrix.")
-    end
-
-    y = Matrix{Typ}(undef, n, p)
-    α = Matrix{Typ}(undef, n, m)
-    α[1, :] = initial_a
-
-    for t = 1:n-1
-        y[t, :]   = model.Z[:, :, t]*α[t, :] + model.d[t, :] + rand(dist_H)
-        α[t+1, :] = model.T*α[t, :] + model.c[t, :] + model.R*rand(dist_Q)
-    end
-    y[n, :] = model.Z[:, :, n]*α[n, :] +  model.d[n, :] + rand(dist_H)
-
-    return y, α
-end
-
-
-# Linar Algebra wrappers
-function gram_in_time(mat::Array{T, 3}) where T
-    gram_in_time = similar(mat)
-    @inbounds @views for t = 1:size(gram_in_time, 3)
-        gram_in_time[:, :, t] = gram(mat[:, :, t])
-    end
-    return gram_in_time
-end
-
-function gram(mat::AbstractArray{T}) where T
-    if size(mat, 1) == 1
-        gram_mat = Matrix{T}(undef, 1, 1)
-        gram_mat[1, 1] = mat[1, 1]^2
-        return gram_mat
-    else
-        return LinearAlgebra.BLAS.gemm('N', 'T', mat, mat) # mat*mat'
-    end
-end
-
-"""
-    check_steady_state(P_t1::Matrix{T}, P_t::Matrix{T}, tol::T) where T
-
-Return `true` if steady state was attained with respect to tolerance `tol`, `false` otherwise.
-The steady state is checked by the following equation `maximum(abs.((P_t1 - P_t)./P_t1)) < tol`.
-"""
-function check_steady_state(P_t1::Matrix{T}, P_t::Matrix{T}, tol::T) where T
-    return maximum(abs.((P_t1 - P_t)./P_t1)) < tol ? true : false
-end
-
-function check_steady_state(P::AbstractArray{T}, t::Int, tol::T) where T
-    @inbounds for j in axes(P, 2), i in axes(P, 1)
-        if abs((P[i, j, t+1] - P[i, j, t])/P[i, j, t+1]) > tol
-            return false
-        end
-    end
-    return true
[…]
