### 2020-10-05 ae474bc
Add local level with cycle (#204)
* Add local level with cycle

* address comments

* [ci skip] Better comment
