+++ src/hyperparameters.jl
@@ -1,3 +1,6 @@
+"""
+TODO
+"""
 mutable struct HyperParameters{Fl <: AbstractFloat}
     num::Int
     names::Vector{String}
@@ -40,14 +43,17 @@ function register_unconstrained_values!(hyperparameters::HyperParameters,
     return nothing
 end
 
-#TODO same function but receiving hyperparameter
-function fill_minimizer_hyperparameter_position!(model::StateSpaceModel)
+function fill_minimizer_hyperparameter_position!(model::StateSpaceModel) 
+    fill_minimizer_hyperparameter_position!(model.hyperparameters)
+    return
+end
+function fill_minimizer_hyperparameter_position!(hyperparameters::HyperParameters)
     # Reset minimizer hyperparameter_position
-    model.hyperparameters.minimizer_hyperparameter_position = Dict{Int, String}()
+    hyperparameters.minimizer_hyperparameter_position = Dict{Int, String}()
     i = 1
-    for name in get_names(model.hyperparameters)
-        if is_free(name, model.hyperparameters)
-            model.hyperparameters.minimizer_hyperparameter_position[i] = name
+    for name in get_names(hyperparameters)
+        if is_free(name, hyperparameters)
+            hyperparameters.minimizer_hyperparameter_position[i] = name
             i += 1
         end
     end
@@ -68,8 +74,19 @@ function handle_optim_initial_hyperparameters(model::StateSpaceModel)
 end
 
 get_hyperparameters(model::StateSpaceModel) = model.hyperparameters
+
+"""
+    get_names(model::StateSpaceModel)
+
+Get the names of the hyperparameters registered on a `StateSpaceModel`.
+"""
 get_names(model::StateSpaceModel) = get_names(model.hyperparameters)
 
+"""
+    number_hyperparameters(model::StateSpaceModel)
+
+Get the number of hyperparameters registered on a `StateSpaceModel`.
+"""
 number_hyperparameters(model::StateSpaceModel) = number_hyperparameters(model.hyperparameters)
 number_hyperparameters(hyperparameters::HyperParameters) = hyperparameters.num
 
@@ -119,6 +136,15 @@ end
 
 
 """
+    set_initial_hyperparameters!(model::StateSpaceModel,
+                                 initial_hyperparameters::Dict{String, <:Real})
+
+Fill a model with user inputed initial points for hyperparameter optimzation.
+
+# Example
+```jldoctest
+julia> model = LocalLevel(rand(100));
+```
 """
 function set_initial_hyperparameters!(model::StateSpaceModel,
                                       initial_hyperparameters::Dict{String, <:Real})
@@ -144,21 +170,32 @@ function number_fixed_hyperparameters(hyperparameters::HyperParameters)
 end
 
 """
-    fix_hyperparameters!
+    fix_hyperparameters!(model::StateSpaceModel, fixed_hyperparameters::Dict)
 
 Fixes the desired hyperparameters so that they are not considered as decision variables in
 the model estimation.
+
+# Example
+```jldoctest
+julia> model = LocalLevel(rand(100));
+
+julia> get_names(model)
+2-element Array{String,1}:
+ "sigma2_ε"
+ "sigma2_η"
+
+julia> fix_hyperparameters!(model, Dict("sigma2_ε" => 15099.0))
+
+```
 """
 function fix_hyperparameters! end
 
 function fix_hyperparameters!(model::StateSpaceModel, fixed_hyperparameters::Dict)
[…]
