+++ src/StateSpaceModels.jl
@@ -25,8 +25,8 @@ include("smoothers/kalman_smoother.jl")
 
 include("models/common.jl")
 include("models/locallevel.jl")
+include("models/locallevelcycle.jl")
 include("models/locallineartrend.jl")
-include("models/damped_lineartrend.jl")
 include("models/basicstructural.jl")
 include("models/basicstructural_multivariate.jl")
 include("models/arima.jl")
@@ -46,6 +46,7 @@ export LinearRegression
 export LinearUnivariateTimeInvariant
 export LinearUnivariateTimeVariant
 export LocalLevel
+export LocalLevelCycle
 export LocalLinearTrend
 export MultivariateBasicStructural
 export Optimizer
@@ -53,6 +54,9 @@ export ScalarKalmanFilter
 export UnivariateKalmanFilter
 
 # Exported functions
+export constrain_box!
+export constrain_identity!
+export constrain_variance!
 export fit!
 export fix_hyperparameters!
 export forecast
@@ -63,6 +67,7 @@ export get_filtered_variance
 export get_hyperparameters
 export get_innovations
 export get_innovation_variance
+export get_names
 export get_predictive_state
 export get_predictive_variance
 export get_smoothed_state
@@ -77,5 +82,8 @@ export results
 export set_initial_hyperparameters!
 export simulate
 export simulate_scenarios
+export unconstrain_box!
+export unconstrain_identity!
+export unconstrain_variance!
 
 end

[…]
