+++ src/models/basicstructural_multivariate.jl
@@ -1,5 +1,5 @@
 @doc raw"""
-An implementation of a non-homogeneous seemingly unrelated time series equations for basic structural 
+An implementation of a non-homogeneous seemingly unrelated time series equations for basic structural
 state-space model consists of trend (local linear trend) and seasonal components.
 It is defined by:
 ```math
@@ -22,15 +22,23 @@ mutable struct MultivariateBasicStructural <: StateSpaceModel
     function MultivariateBasicStructural(y::Matrix{Fl}, s::Int) where Fl
         p = size(y, 2)
         Z = kron(Matrix{Fl}(I, p, p), [1 0 1 zeros(Fl, 1, s - 2)])
-        T = kron(Matrix{Fl}(I, p, p),
-                [1 1 zeros(Fl, 1, s - 1);
-                 0 1 zeros(Fl, 1, s - 1);
-                 0 0 -ones(Fl, 1, s - 1);
-                zeros(Fl, s - 2, 2) Matrix{Fl}(I, s - 2, s - 2) zeros(Fl, s - 2)])
-                
-        R = kron(Matrix{Fl}(I, p, p),
-                [Matrix{Fl}(I, 3, 3);
-                zeros(Fl, s - 2, 3)])
+        T = kron(
+            Matrix{Fl}(I, p, p),
+            [
+                1 1 zeros(Fl, 1, s - 1)
+                0 1 zeros(Fl, 1, s - 1)
+                0 0 -ones(Fl, 1, s - 1)
+                zeros(Fl, s - 2, 2) Matrix{Fl}(I, s - 2, s - 2) zeros(Fl, s - 2)
+            ],
+        )
+
+        R = kron(
+            Matrix{Fl}(I, p, p),
+            [
+                Matrix{Fl}(I, 3, 3)
+                zeros(Fl, s - 2, 3)
+            ],
+        )
 
         d = zeros(Fl, p)
         c = zeros(Fl, p * (s + 1))
@@ -65,11 +73,14 @@ function default_filter(model::MultivariateBasicStructural)
     steadystate_tol = Fl(1e-5)
     a1 = zeros(Fl, num_states(model))
     P1 = Fl(1e6) .* Matrix{Fl}(I, num_states(model), num_states(model))
-    return MultivariateKalmanFilter(size(model.system.y, 2), a1, P1, num_states(model), steadystate_tol)
+    return MultivariateKalmanFilter(
+        size(model.system.y, 2), a1, P1, num_states(model), steadystate_tol
+    )
 end
+
 function initial_hyperparameters!(model::MultivariateBasicStructural)
     Fl = typeof_model_elements(model)
-    initial_hyperparameters = Dict{String, Fl}()
+    initial_hyperparameters = Dict{String,Fl}()
     for variable in get_names(model)
         if occursin("sigma2", variable)
             initial_hyperparameters[variable] = one(Fl)
@@ -78,8 +89,9 @@ function initial_hyperparameters!(model::MultivariateBasicStructural)
         end
     end
     set_initial_hyperparameters!(model, initial_hyperparameters)
-    return
+    return nothing
 end
+
 function constrain_hyperparameters!(model::MultivariateBasicStructural)
     Fl = typeof_model_elements(model)
     p = size(model.system.y, 2)
@@ -87,7 +99,9 @@ function constrain_hyperparameters!(model::MultivariateBasicStructural)
     H_unconstrained = zeros(Fl, p, p)
     current_name = 1
     for i in 1:p, j in 1:i
-        H_unconstrained[i, j] = get_unconstrained_value(model, get_names(model)[current_name])
+        H_unconstrained[i, j] = get_unconstrained_value(
[…]
