+++ src/hyperparameters.jl
@@ -1,55 +1,60 @@
 """
 TODO
 """
-mutable struct HyperParameters{Fl <: AbstractFloat}
+mutable struct HyperParameters{Fl<:AbstractFloat}
     num::Int
     names::Vector{String}
     # Poistion of the hyperparameter on the constrained and uncontrained values
-    position::Dict{String, Int}
+    position::Dict{String,Int}
     # Minimizer vector corresponds to which hyperparameter
-    minimizer_hyperparameter_position::Dict{Int, String}
+    minimizer_hyperparameter_position::Dict{Int,String}
     constrained_values::Vector{Fl}
     unconstrained_values::Vector{Fl}
-    fixed_constrained_values::Dict{String, Fl}
+    fixed_constrained_values::Dict{String,Fl}
 
-    function HyperParameters{Fl}(names::Vector{String}) where {Fl}
-        num                               = length(names)
-        positions                         = Dict(names .=> collect(1:num))
-        minimizer_hyperparameter_position = Dict{Int, String}()
-        constrained_values                = fill(Fl(NaN), num)
-        unconstrained_values              = fill(Fl(NaN), num)
-        fixed_constrained_values          = Dict{String, Fl}()
-        return new{Fl}(num,
-                       names,
-                       positions,
-                       minimizer_hyperparameter_position,
-                       constrained_values,
-                       unconstrained_values,
-                       fixed_constrained_values)
+    function HyperParameters{Fl}(names::Vector{String}) where Fl
+        num = length(names)
+        positions = Dict(names .=> collect(1:num))
+        minimizer_hyperparameter_position = Dict{Int,String}()
+        constrained_values = fill(Fl(NaN), num)
+        unconstrained_values = fill(Fl(NaN), num)
+        fixed_constrained_values = Dict{String,Fl}()
+        return new{Fl}(
+            num,
+            names,
+            positions,
+            minimizer_hyperparameter_position,
+            constrained_values,
+            unconstrained_values,
+            fixed_constrained_values,
+        )
     end
 end
 
-function register_unconstrained_values!(model::StateSpaceModel,
-                                        unconstrained_values::Vector{Fl}) where Fl
+function register_unconstrained_values!(
+    model::StateSpaceModel, unconstrained_values::Vector{Fl}
+) where Fl
     return register_unconstrained_values!(model.hyperparameters, unconstrained_values)
 end
-function register_unconstrained_values!(hyperparameters::HyperParameters,
-                                        unconstrained_values::Vector{Fl}) where Fl
+function register_unconstrained_values!(
+    hyperparameters::HyperParameters, unconstrained_values::Vector{Fl}
+) where Fl
     for (i, val) in enumerate(unconstrained_values)
-        update_unconstrained_value!(hyperparameters,
-                                    hyperparameters.minimizer_hyperparameter_position[i],
-                                    val)
+        update_unconstrained_value!(
+            hyperparameters, hyperparameters.minimizer_hyperparameter_position[i], val
+        )
     end
-    return nothing
+    return hyperparameters
 end
 
-function fill_minimizer_hyperparameter_position!(model::StateSpaceModel) 
+function fill_minimizer_hyperparameter_position!(model::StateSpaceModel)
     fill_minimizer_hyperparameter_position!(model.hyperparameters)
-    return
+    return model
 end
+
[…]
