+++ test/models/arima.jl
@@ -11,14 +11,23 @@
     forec = forecast(model, 10)
     @test monotone_forecast_variance(forec)
     # Prediction from Pyhton statsmodels
-    predicted_mean = [-0.72809028,  -0.47352952,  -0.30797034,  -0.20029528,
-                      -0.13026644,  -0.08472164,  -0.05510058,  -0.03583588,
-                      -0.02330665,  -0.01515799]
-    @test predicted_mean ≈ vcat(forec.expected_value...) atol=1e-3
+    predicted_mean = [
+        -0.72809028,
+        -0.47352952,
+        -0.30797034,
+        -0.20029528,
+        -0.13026644,
+        -0.08472164,
+        -0.05510058,
+        -0.03583588,
+        -0.02330665,
+        -0.01515799,
+    ]
+    @test predicted_mean ≈ vcat(forec.expected_value...) atol = 1e-3
     # simualting
     scenarios = simulate_scenarios(model, 10, 100_000)
     # Values are very close to 0.0 so we test with absolute tolerance
-    # It attains 1e-3 when we make 10M simulations, which is too much 
+    # It attains 1e-3 when we make 10M simulations, which is too much
     # computation for a rather simple test.
     test_scenarios_adequacy_with_forecast(forec, scenarios; atol=1e-1)
 
