### 2020-10-07 12454ab
Fix doctests
