+++ src/hyperparameters.jl
@@ -182,6 +182,7 @@ function has_hyperparameter(str::Vector{String}, hyperparameters::HyperParameter
     return true
 end
 
+# TODO incomplete doctest?
 """
     set_initial_hyperparameters!(model::StateSpaceModel,
                                  initial_hyperparameters::Dict{String, <:Real})
@@ -242,8 +243,8 @@ julia> get_names(model)
  "sigma2_ε"
  "sigma2_η"
 
-julia> fix_hyperparameters!(model, Dict("sigma2_ε" => 15099.0))
-
+julia> fix_hyperparameters!(model, Dict("sigma2_ε" => 100.0))
+StateSpaceModels.HyperParameters{Float64}(2, ["sigma2_ε", "sigma2_η"], Dict("sigma2_η" => 2,"sigma2_ε" => 1), Dict{Int64,String}(), [NaN, NaN], [NaN, NaN], Dict("sigma2_ε" => 100.0))
 ```
 """
 function fix_hyperparameters! end
