### 2020-10-08 67341f1
Adjust code to Blue Style (#205)
* Adjust to Blue Style

* Fix doctests

* Make doctests run only on one job

* Fix @examples
