+++ src/filters/univariate_kalman_filter.jl
@@ -4,7 +4,7 @@
 A Kalman filter that is tailored to univariate systems, exploiting the fact that the
 dimension of the observations at any time period is 1.
 """
-mutable struct UnivariateKalmanState{Fl <: AbstractFloat}
+mutable struct UnivariateKalmanState{Fl<:AbstractFloat}
     v::Fl
     F::Fl
     att::Vector{Fl}
@@ -23,23 +23,33 @@ mutable struct UnivariateKalmanState{Fl <: AbstractFloat}
         ZP = zeros(Fl, m)
         TPtt = zeros(Fl, m, m)
 
-        return new{Fl}(zero(Fl), zero(Fl), zeros(Fl, m),
-                       a1, zeros(Fl, m, m), P1, zero(Fl), false,
-                       P_to_check_steady_state,
-                       ZP,
-                       TPtt)
+        return new{Fl}(
+            zero(Fl),
+            zero(Fl),
+            zeros(Fl, m),
+            a1,
+            zeros(Fl, m, m),
+            P1,
+            zero(Fl),
+            false,
+            P_to_check_steady_state,
+            ZP,
+            TPtt,
+        )
     end
 end
 
-function save_a1_P1_in_filter_output!(filter_output::FilterOutput{Fl},
-                                      kalman_state::UnivariateKalmanState{Fl}) where Fl
+function save_a1_P1_in_filter_output!(
+    filter_output::FilterOutput{Fl}, kalman_state::UnivariateKalmanState{Fl}
+) where Fl
     filter_output.a[1] = deepcopy(kalman_state.a)
     filter_output.P[1] = deepcopy(kalman_state.P)
-    return
+    return filter_output
 end
-function save_kalman_state_in_filter_output!(filter_output::FilterOutput{Fl},
-                                             kalman_state::UnivariateKalmanState{Fl},
-                                             t::Int) where Fl
+
+function save_kalman_state_in_filter_output!(
+    filter_output::FilterOutput{Fl}, kalman_state::UnivariateKalmanState{Fl}, t::Int
+) where Fl
     filter_output.v[t] = copy(fill(kalman_state.v, 1))
     filter_output.F[t] = copy(fill(kalman_state.F, 1, 1))
     filter_output.a[t + 1] = copy(kalman_state.a)
@@ -48,25 +58,26 @@ function save_kalman_state_in_filter_output!(filter_output::FilterOutput{Fl},
     filter_output.Ptt[t] = copy(kalman_state.Ptt)
     # There is no diffuse part in UnivariateKalmanState
     filter_output.Pinf[t] = copy(fill(zero(Fl), size(kalman_state.Ptt)))
-    return
+    return filter_output
 end
 
 # Univariate Kalman filter with the recursions as described
 # in Koopman's book TODO
-mutable struct UnivariateKalmanFilter{Fl <: Real} <: KalmanFilter
+mutable struct UnivariateKalmanFilter{Fl<:Real} <: KalmanFilter
     steadystate_tol::Fl
     a1::Vector{Fl}
     P1::Matrix{Fl}
     skip_llk_instants::Int
     kalman_state::UnivariateKalmanState
 
-    function UnivariateKalmanFilter(a1::Vector{Fl},
-                                    P1::Matrix{Fl},
-                                    skip_llk_instants::Int = length(a1),
-                                    steadystate_tol::Fl = Fl(1e-5)) where Fl
+    function UnivariateKalmanFilter(
+        a1::Vector{Fl},
+        P1::Matrix{Fl},
+        skip_llk_instants::Int=length(a1),
[…]
