+++ src/filters/multivariate_kalman_filter.jl
@@ -3,7 +3,7 @@
 
 A Kalman filter that is tailored to multivariate systems.
 """
-mutable struct MultivariateKalmanState{Fl <: AbstractFloat}
+mutable struct MultivariateKalmanState{Fl<:AbstractFloat}
     v::Vector{Fl}
     F::Matrix{Fl}
     att::Vector{Fl}
@@ -15,27 +15,39 @@ mutable struct MultivariateKalmanState{Fl <: AbstractFloat}
     # Auxiliary matrices
     P_to_check_steady_state::Matrix{Fl}
     TPtt::Matrix{Fl}
-    function MultivariateKalmanState(num_series::Int, a1::Vector{Fl}, P1::Matrix{Fl}) where Fl
+    function MultivariateKalmanState(
+        num_series::Int, a1::Vector{Fl}, P1::Matrix{Fl}
+    ) where Fl
         m = length(a1)
         P_to_check_steady_state = zeros(Fl, m, m)
         TPtt = zeros(Fl, m, m)
 
-        return new{Fl}(zeros(Fl, num_series), zeros(Fl, num_series, num_series), zeros(Fl, m),
-                       a1, zeros(Fl, m, m), P1, zero(Fl), false,
-                       P_to_check_steady_state,
-                       TPtt)
+        return new{Fl}(
+            zeros(Fl, num_series),
+            zeros(Fl, num_series, num_series),
+            zeros(Fl, m),
+            a1,
+            zeros(Fl, m, m),
+            P1,
+            zero(Fl),
+            false,
+            P_to_check_steady_state,
+            TPtt,
+        )
     end
 end
 
-function save_a1_P1_in_filter_output!(filter_output::FilterOutput{Fl},
-                                      kalman_state::MultivariateKalmanState{Fl}) where Fl
+function save_a1_P1_in_filter_output!(
+    filter_output::FilterOutput{Fl}, kalman_state::MultivariateKalmanState{Fl}
+) where Fl
     filter_output.a[1] = deepcopy(kalman_state.a)
     filter_output.P[1] = deepcopy(kalman_state.P)
-    return
+    return filter_output
 end
-function save_kalman_state_in_filter_output!(filter_output::FilterOutput{Fl},
-                                             kalman_state::MultivariateKalmanState{Fl},
-                                             t::Int) where Fl
+
+function save_kalman_state_in_filter_output!(
+    filter_output::FilterOutput{Fl}, kalman_state::MultivariateKalmanState{Fl}, t::Int
+) where Fl
     filter_output.v[t] = copy(kalman_state.v)
     filter_output.F[t] = copy(kalman_state.F)
     filter_output.a[t + 1] = copy(kalman_state.a)
@@ -44,27 +56,28 @@ function save_kalman_state_in_filter_output!(filter_output::FilterOutput{Fl},
     filter_output.Ptt[t] = copy(kalman_state.Ptt)
     # There is no diffuse part in UnivariateKalmanState
     filter_output.Pinf[t] = copy(fill(zero(Fl), size(kalman_state.Ptt)))
-    return
+    return filter_output
 end
 
 """
 TODO
 """
-mutable struct MultivariateKalmanFilter{Fl <: Real} <: KalmanFilter
+mutable struct MultivariateKalmanFilter{Fl<:Real} <: KalmanFilter
     steadystate_tol::Fl
     a1::Vector{Fl}
     P1::Matrix{Fl}
     skip_llk_instants::Int
     kalman_state::MultivariateKalmanState
 
-    function MultivariateKalmanFilter(num_series::Int,
-                                    a1::Vector{Fl},
-                                    P1::Matrix{Fl},
[…]
