+++ Project.toml
@@ -12,6 +12,7 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
 Optim = "0.20, 0.21, 0.22"
+Distributions = "0.23, 0.24"
 julia = "1"
 
 [extras]
