### 2020-10-16 b4794e6
Update manual.md
