### 2020-10-16 97c3713
Update examples.md
