### 2020-10-16 3be09e0
Some doc improvements (#208)
* Update index.md

* Update index.md

* Update manual.md

* Update examples.md

* Add mathjax
