+++ docs/make.jl
@@ -10,7 +10,7 @@ makedocs(;
     modules=[StateSpaceModels],
     doctest=true,
     clean=true,
-    format=Documenter.HTML(),
+    format=Documenter.HTML(mathengine=Documenter.MathJax()),
     sitename="StateSpaceModels.jl",
     authors="Raphael Saavedra, Guilherme Bodin, and Mario Souto",
     pages=[
