+++ src/models/arima.jl
@@ -323,4 +323,5 @@ function reinstantiate(model::ARIMA, y::Vector{Fl}) where Fl
     order = (model.order.p, model.order.d, model.order.q)
     return ARIMA(y, order)
 end
+
 has_exogenous(::ARIMA) = false
