### 2020-10-16 5bf6ebd
Remove damped linear trend
