### 2020-10-19 eaaf94e
Update hyperparameters.jl
