+++ src/models/linear_regression.jl
@@ -5,6 +5,19 @@ struct RegressionHyperParametersAuxiliary
     end
 end
 
+@doc raw"""
+The linear regression state-space model is defined by:
+```math
+\begin{gather*}
+    \begin{aligned}
+        y_{t} &=  \X_{1,t} \cdot \beta_{1,t} + \dots + \X_{n,t} \cdot \beta_{1,t} + \varepsilon_{t} \quad &\varepsilon_{t} \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\
+        \beta_{1,t+1} &= \beta_{1,t}\\
+        \dots &= \dots\\
+        \beta_{n,t+1} &= \beta_{n,t}\\
+    \end{aligned}
+\end{gather*}
+```
+"""
 mutable struct LinearRegression <: StateSpaceModel
     hyperparameters::HyperParameters
     hyperparameters_auxiliary::RegressionHyperParametersAuxiliary
