### 2020-10-20 64ad732
Update src/models/linear_regression.jl
Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
