### 2020-10-20 1fe1ada
fix docs (#210)
* fix docs

* fix README.md

* Update src/models/linear_regression.jl

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
