### 2020-11-30 e50523c
Added LocalRegression Model
