### 2020-12-01 ea25db0
Add local level with explanatory variables
