+++ test/utils.jl
@@ -1,6 +1,6 @@
 function monotone_forecast_variance(forec::StateSpaceModels.Forecast)
     for i in 2:length(forec.covariance)
-        if all(diag(forec.covariance[i]) .<= diag(forec.covariance[i - 1]))
+        if all(diag(forec.covariance[i]) .< diag(forec.covariance[i - 1]))
             return false
         end
     end
