+++ test/runtests.jl
@@ -12,6 +12,7 @@ include("utils.jl")
     include("models/locallevel.jl")
     include("models/locallineartrend.jl")
     include("models/locallevelcycle.jl")
+    include("models/locallevelexplanatory.jl")
     include("models/basicstructural.jl")
     include("models/basicstructural_multivariate.jl")
     include("models/arima.jl")
