+++ test/models/locallevelexplanatory.jl
@@ -11,6 +11,6 @@
     @test monotone_forecast_variance(forec)
     kf = kalman_filter(model)
     a = get_predictive_state(kf)
-    @test a[1, 2] ≈ a[end, 2]
-    @test a[1, 3] ≈ a[end, 3]
+    @test a[1, 2] ≈ a[end, 2] atol=1e-3
+    @test a[1, 3] ≈ a[end, 3] atol=1e-3
 end
