+++ .travis.yml
@@ -1,31 +1,0 @@
-language: julia
-
-os:
-  - linux
-  - osx
-
-julia:
-  - 1.0 # Latest LTS release
-  - 1.4
-  - 1   # Latest stable release
-  - nightly
-
-jobs:
-  allow_failures:
-    - julia: nightly
-  include:
-    - stage: "Documentation"
-      julia: 1
-      os: linux
-      script: julia --project=docs -e '
-          using Pkg;
-          Pkg.develop(PackageSpec(path=pwd()));
-          Pkg.instantiate();
-          include("docs/make.jl");'
-      after_success: skip
-
-notifications:
-  email: false
-
-after_success:
-  - julia -e 'using Pkg; Pkg.add("Coverage"); using Coverage; Codecov.submit(process_folder())'
