+++ .github/workflows/ci.yml
@@ -12,12 +12,15 @@ jobs:
       fail-fast: false
       matrix:
         include:
-          version:
-          - '1.0'
-          os:
-            - ubuntu-latest
-          arch:
-            - x64
+          - version: '1'
+            os: ubuntu-latest
+            arch: x64
+          - version: '1.0'
+            os: ubuntu-latest
+            arch: x64
+          - version: '1.0'
+            os: ubuntu-latest
+            arch: x86
     steps:
       - uses: actions/checkout@v2
       - uses: julia-actions/setup-julia@v1
