### 2020-12-01 119a748
Update test/models/locallevelexplanatory.jl
Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
