### 2020-12-01 e442870
Add local level with explanatory variables (#215)
* Add local level with explanatory variables

* update tolerance

* Update to github ci

* Update ci

* Updaaate TagBot action

* Update src/models/locallevelexplanatory.jl

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* Update test/models/locallevelexplanatory.jl

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* fix test example

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
