### 2020-12-02 3ac9df8
Fix broken arima test and some issues more
