+++ test/models/basicstructural.jl
@@ -18,8 +18,8 @@
     test_scenarios_adequacy_with_forecast(forec, scenarios)
 
     log_ap32 = Float32.(log_air_passengers)
-    model = BasicStructural(log_ap32, 12)
+    model = StateSpaceModels.BasicStructural(log_ap32, 12)
     @test model.system.Z == Float32[1.0; 0.0; 1.0; zeros(10)]
-    @test_broken fit!(model)
+    @test_broken StateSpaceModels.fit!(model)
     @test_broken loglike(model) ≈ 234.33641f0 atol = 1e-5 rtol = 1e-5
 end
