### 2020-12-04 bfca979
Add simplification
