### 2020-12-06 12f887b
Fixed ARIMA tests
