### 2020-12-06 c21ed86
Add compat for ShiftedArraays
