### 2020-12-06 20c9333
Fix arima (#218)
* Fix broken arima test and some issues more

* back to the original file

* Add simplification

* Fixed ARIMA tests

* Add compat for ShiftedArraays
