### 2020-12-07 e55bbb5
Add mean spec to ARIMA
