+++ Project.toml
@@ -14,8 +14,8 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 [compat]
 Distributions = "0.23, 0.24"
 Optim = "0.20, 0.21, 0.22"
-julia = "1"
 ShiftedArrays = "1.0"
+julia = "1"
 
 [extras]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
