### 2020-12-08 4d4130a
Assert ARIMA order
