+++ src/datasets.jl
@@ -93,9 +93,14 @@ TODO
 const RJ_TEMPERATURE = joinpath(dirname(@__DIR__()), "datasets", "rj_temperature.csv")
 
 @doc raw"""
+    US_CHANGE
+
 Percentage changes in quarterly personal consumption expenditure, personal disposable income,
 production, savings and the unemployment rate for the US, 1960 to 2016.
 
 Federal Reserve Bank of St Louis.
+
+# References
+ * Hyndman, Rob J., Athanasopoulos, George. "Forecasting: Principles and Practice"
 """
 const US_CHANGE = joinpath(dirname(@__DIR__()), "datasets", "uschange.csv")
