### 2020-12-10 609db03
Add Vehicle Tracking model
