+++ src/kalman_filter_and_smoother.jl
@@ -71,6 +71,13 @@ mutable struct FilterOutput{Fl<:Real}
     end
 end
 
+function assert_possible_to_filter(model::StateSpaceModel)
+    if !has_hyperparameter(model) || isfitted(model)
+        return true
+    end
+    error("Model has not been estimated yet, please use `fit!`.")
+end
+
 """
     get_innovations
 
@@ -81,7 +88,7 @@ function get_innovations end
 get_innovations(filter::FilterOutput) = permutedims(cat(filter.v...; dims=2))
 
 function get_innovations(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
+    assert_possible_to_filter(model)
     return get_innovations(kalman_filter(model; filter=filter))
 end
 
@@ -95,7 +102,7 @@ function get_innovation_variance end
 get_innovation_variance(filter_output::FilterOutput) = cat(filter_output.F...; dims=3)
 
 function get_innovation_variance(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
+    assert_possible_to_filter(model)
     return get_innovation_variance(kalman_filter(model; filter=filter))
 end
 
@@ -109,7 +116,7 @@ function get_filtered_state end
 get_filtered_state(filter::FilterOutput) = permutedims(cat(filter.att...; dims=2))
 
 function get_filtered_state(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
+    assert_possible_to_filter(model)
     return get_filtered_state(kalman_filter(model; filter=filter))
 end
 
@@ -123,7 +130,7 @@ function get_filtered_variance end
 get_filtered_variance(filter::FilterOutput) = cat(filter.Ptt...; dims=3)
 
 function get_filtered_variance(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
+    assert_possible_to_filter(model)
     return get_filtered_variance(kalman_filter(model; filter=filter))
 end
 
@@ -137,7 +144,7 @@ function get_predictive_state end
 get_predictive_state(filter::FilterOutput) = permutedims(cat(filter.a...; dims=2))
 
 function get_predictive_state(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
+    assert_possible_to_filter(model)
     return get_predictive_state(kalman_filter(model; filter=filter))
 end
 
@@ -151,7 +158,7 @@ function get_predictive_variance end
 get_predictive_variance(filter::FilterOutput) = cat(filter.P...; dims=3)
 
 function get_predictive_variance(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
+    assert_possible_to_filter(model)
     return get_predictive_variance(kalman_filter(model; filter=filter))
 end
 
@@ -161,12 +168,14 @@ end
 Filter the data using the desired `filter` and the estimated hyperparameters in `model`.
 """
 function kalman_filter(model::StateSpaceModel; filter=default_filter(model))
-    isfitted(model) || error("Model has not been estimated yet, please use `fit!`.")
+    assert_possible_to_filter(model)
     filter_output = FilterOutput(model)
     reset_filter!(filter)
-    free_unconstrained_values = get_free_unconstrained_values(model)
-    update_model_hyperparameters!(model, free_unconstrained_values)
[…]

[…]
