### 2020-12-10 4f510fe
Add a mean term in ARIMA (#219)
* Add mean spec to ARIMA

* Include mean in ARIMA

* Assert ARIMA order

* [ci skip] Change README.md

* [ci skip] Add contributing on README.md

* Add reference to US_CHANGE
