+++ test/models/arima.jl
@@ -43,4 +43,13 @@
     model = ARIMA(wpi, (1, 1, 1))
     fit!(model)
     @test loglike(model) ≈ -137.246818 atol = 1e-5 rtol = 1e-5
+
+    uschange_consumption = CSV.read(StateSpaceModels.US_CHANGE, DataFrame).Consumption
+    model = ARIMA(uschange_consumption, (1, 0, 3); include_mean = true)
+    fit!(model)
+    @test loglike(model) ≈ -164.8 atol = 1e-1 rtol = 1e-1
+
+    model = ARIMA(uschange_consumption, (3, 0, 0); include_mean = true)
+    fit!(model)
+    @test loglike(model) ≈ -165.2 atol = 1e-1 rtol = 1e-1
 end
