+++ src/models/arima.jl
@@ -2,6 +2,12 @@ struct ARIMAOrder
     p::Int
     d::Int
     q::Int
+    function ARIMAOrder(p::Int, d::Int, q::Int)
+        @assert p >= 0
+        @assert d >= 0
+        @assert q >= 0
+        return new(p, d, q)
+    end
 end
 
 # This is used to some optimizations such as
@@ -35,8 +41,9 @@ mutable struct ARIMA <: StateSpaceModel
     hyperparameters::HyperParameters
     system::LinearUnivariateTimeInvariant
     results::Results
+    include_mean::Bool
 
-    function ARIMA(y::Vector{Fl}, order::Tuple{Int,Int,Int}) where Fl
+    function ARIMA(y::Vector{Fl}, order::Tuple{Int,Int,Int}; include_mean::Bool = false) where Fl
         or = ARIMAOrder(order[1], order[2], order[3])
         hyperparameters_auxiliary = ARIMAHyperParametersAuxiliary(or)
 
@@ -51,6 +58,9 @@ mutable struct ARIMA <: StateSpaceModel
         system = LinearUnivariateTimeInvariant{Fl}(y, Z, T, R, d_ss, c, H, Q_ss)
 
         num_hyperparameters = or.p + or.q + 1
+        if include_mean 
+            num_hyperparameters += 1
+        end
         names = Vector{String}(undef, num_hyperparameters)
         for i in 1:(or.p)
             names[i] = "ar_L$i"
@@ -58,11 +68,14 @@ mutable struct ARIMA <: StateSpaceModel
         for j in 1:(or.q)
             names[or.p + j] = "ma_L$j"
         end
+        if include_mean
+            names[end-1] = "mean"
+        end
         names[end] = "sigma2_η"
 
         hyperparameters = HyperParameters{Fl}(names)
 
-        return new(or, hyperparameters_auxiliary, hyperparameters, system, Results{Fl}())
+        return new(or, hyperparameters_auxiliary, hyperparameters, system, Results{Fl}(), include_mean)
     end
 end
 
@@ -252,27 +265,22 @@ function concatenate_on_bottom(X1::Matrix{Fl}, X2::Matrix{Fl}) where Fl
 end
 
 function diff_arma(y::Vector{Fl}, d::Int) where Fl
-    if d > 0
-        return Vector{Fl}((y - lag(y, d))[d+1:end])
-    else
-        return y
+    for _ in 1:d
+        y = diff(y)
     end
+    return y
 end
 
 function initial_hyperparameters!(model::ARIMA)
     Fl = typeof_model_elements(model)
-    # TODO find out how to better estimate initial parameters
     initial_hyperparameters = Dict{String,Fl}()
     # Heuristic inspired in statsmodels from python
     # TODO find a reference to this heuristic
     y = filter(!isnan, model.system.y)
     y_diff = diff_arma(y, model.order.d)
     k = model.order.p + model.order.q
-    residuals = nothing
     X = lagmat(y_diff, k)
-    if model.order.q > 0
-        residuals = y_diff[k+1:end] - X * (X \ y_diff[k+1:end])
-    end
+    residuals = y_diff[k+1:end] - X * (X \ y_diff[k+1:end])
     X_ar = lagmat(y_diff, model.order.p)
     X_ma = lagmat(residuals, model.order.q)
     X = concatenate_on_bottom(X_ar, X_ma)
@@ -287,8 +295,11 @@ function initial_hyperparameters!(model::ARIMA)
         initial_hyperparameters[get_ma_name(model, j)] = initial_params[offset]
         offset += 1
     end
-    initial_hyperparameters["sigma2_η"] = var(y_diff)
+    if model.include_mean 
[…]
