### 2020-12-11 98be82e
arima model working
