+++ src/models/arima.jl
@@ -1,12 +1,26 @@
-struct ARIMAOrder
+struct SARIMAOrder
     p::Int
     d::Int
     q::Int
-    function ARIMAOrder(p::Int, d::Int, q::Int)
+    P::Int
+    D::Int
+    Q::Int
+    s::Int
+    r::Int # Model order
+    n_states_diff::Int
+    n_states::Int
+    function SARIMAOrder(p::Int, d::Int, q::Int, P::Int, D::Int, Q::Int, s::Int)
         @assert p >= 0
         @assert d >= 0
         @assert q >= 0
-        return new(p, d, q)
+        @assert P >= 0
+        @assert D >= 0
+        @assert Q >= 0
+        @assert s >= 0
+        r = max(p + s * P, q + s * Q + 1)
+        n_states_diff = d + D * s
+        n_states = d + D * s + r
+        return new(p, d, q, P, D, Q, s, r, n_states_diff, n_states)
     end
 end
 
@@ -14,38 +28,50 @@ end
 # defining the position of each parameter before hand
 # and avoiding string interpolations such as "ar_L$i"
 # for "ar_L1"
-struct ARIMAHyperParametersAuxiliary
+struct SARIMAHyperParametersAuxiliary
     ar_pos::Vector{Int}
     ar_names::Vector{String}
+    seasonal_ar_pos::Vector{Int}
+    seasonal_ar_names::Vector{String}
     ma_pos::Vector{Int}
     ma_names::Vector{String}
-    function ARIMAHyperParametersAuxiliary(order::ARIMAOrder)
+    seasonal_ma_pos::Vector{Int}
+    seasonal_ma_names::Vector{String}
+    function SARIMAHyperParametersAuxiliary(order::SARIMAOrder)
         return new(
             ar_position_T_matrix(order),
-            create_ar_names(order),
+            create_names("ar_L", order.p),
+            seasonal_ar_position_T_matrix(order),
+            create_names("s_ar_L", order.P),
             ma_position_R_matrix(order),
-            create_ma_names(order),
+            create_names("ma_L", order.q),
+            seasonal_ma_position_R_matrix(order),
+            create_names("s_ma_L", order.Q),
         )
     end
 end
 
 """
-    ARIMA
+    SARIMA
 
-An ARIMA model (autoregressive integrated moving average) implemented within the state-space
+A SARIMA model (seasonal autoregressive integrated moving average) implemented within the state-space
 framework.
 """
-mutable struct ARIMA <: StateSpaceModel
-    order::ARIMAOrder
-    hyperparameters_auxiliary::ARIMAHyperParametersAuxiliary
+mutable struct SARIMA <: StateSpaceModel
+    order::SARIMAOrder
+    hyperparameters_auxiliary::SARIMAHyperParametersAuxiliary
     hyperparameters::HyperParameters
     system::LinearUnivariateTimeInvariant
     results::Results
     include_mean::Bool
 
-    function ARIMA(y::Vector{Fl}, order::Tuple{Int,Int,Int}; include_mean::Bool = false) where Fl
-        or = ARIMAOrder(order[1], order[2], order[3])
-        hyperparameters_auxiliary = ARIMAHyperParametersAuxiliary(or)
+    function SARIMA(y::Vector{Fl}; 
+                    order::Tuple{Int,Int,Int} = (1, 0, 0), 
+                    seasonal_order::Tuple{Int, Int, Int, Int} = (0, 0, 0, 0),
+                    include_mean::Bool = false) where Fl
+        or = SARIMAOrder(order[1], order[2], order[3], 
[…]
