### 2020-12-11 a343557
saarima working with polynomials
