+++ src/models/arima.jl
@@ -1,553 +1,0 @@
-struct SARIMAOrder
-    p::Int
-    d::Int
-    q::Int
-    P::Int
-    D::Int
-    Q::Int
-    s::Int
-    r::Int # Model order
-    n_states_diff::Int
-    n_states::Int
-    function SARIMAOrder(p::Int, d::Int, q::Int, P::Int, D::Int, Q::Int, s::Int)
-        @assert p >= 0
-        @assert d >= 0
-        @assert q >= 0
-        @assert P >= 0
-        @assert D >= 0
-        @assert Q >= 0
-        @assert s >= 0
-        r = max(p + s * P, q + s * Q + 1)
-        n_states_diff = d + D * s
-        n_states = d + D * s + r
-        return new(p, d, q, P, D, Q, s, r, n_states_diff, n_states)
-    end
-end
-
-# This is used to some optimizations such as
-# defining the position of each parameter before hand
-# and avoiding string interpolations such as "ar_L$i"
-# for "ar_L1"
-struct SARIMAHyperParametersAuxiliary
-    ar_pos::Vector{Int}
-    ar_names::Vector{String}
-    seasonal_ar_pos::Vector{Int}
-    seasonal_ar_names::Vector{String}
-    ma_pos::Vector{Int}
-    ma_names::Vector{String}
-    seasonal_ma_pos::Vector{Int}
-    seasonal_ma_names::Vector{String}
-    function SARIMAHyperParametersAuxiliary(order::SARIMAOrder)
-        return new(
-            ar_position_T_matrix(order),
-            create_names("ar_L", order.p),
-            seasonal_ar_position_T_matrix(order),
-            create_names("s_ar_L", order.P),
-            ma_position_R_matrix(order),
-            create_names("ma_L", order.q),
-            seasonal_ma_position_R_matrix(order),
-            create_names("s_ma_L", order.Q),
-        )
-    end
-end
-
-"""
-    SARIMA
-
-A SARIMA model (seasonal autoregressive integrated moving average) implemented within the state-space
-framework.
-"""
-mutable struct SARIMA <: StateSpaceModel
-    order::SARIMAOrder
-    hyperparameters_auxiliary::SARIMAHyperParametersAuxiliary
-    hyperparameters::HyperParameters
-    system::LinearUnivariateTimeInvariant
-    results::Results
-    include_mean::Bool
-
-    function SARIMA(y::Vector{Fl}; 
-                    order::Tuple{Int,Int,Int} = (1, 0, 0), 
-                    seasonal_order::Tuple{Int, Int, Int, Int} = (0, 0, 0, 0),
-                    include_mean::Bool = false) where Fl
-        or = SARIMAOrder(order[1], order[2], order[3], 
-                        seasonal_order[1], seasonal_order[2], seasonal_order[3], seasonal_order[4])
-        hyperparameters_auxiliary = SARIMAHyperParametersAuxiliary(or)
-
-        Z = build_Z(or, Fl)
-        T = build_T(or, Fl)
-        R = build_R(or, Fl)
-        d_ss = zero(Fl)
-        c = build_c(or, Fl)
-        H = zero(Fl)
-        Q_ss = zeros(Fl, 1, 1)
-
-        system = LinearUnivariateTimeInvariant{Fl}(y, Z, T, R, d_ss, c, H, Q_ss)
-
-        num_hyperparameters = or.p + or.q + or.P + or.Q + 1
-        if include_mean 
-            num_hyperparameters += 1
-        end
-        names = Vector{String}(undef, num_hyperparameters)
-        for i in 1:(or.p)
-            names[i] = "ar_L$i"
-        end
-        for j in 1:(or.q)
[…]
