+++ test/runtests.jl
@@ -18,5 +18,5 @@ include("models/locallevelcycle.jl")
 include("models/locallevelexplanatory.jl")
 include("models/basicstructural.jl")
 include("models/basicstructural_multivariate.jl")
-include("models/arima.jl")
+include("models/sarima.jl")
 include("models/linear_regression.jl")
