### 2020-12-13 ae011b2
working for some sarima models
