+++ src/models/sarima.jl
@@ -38,15 +38,19 @@ mutable struct SARIMAHyperParametersAuxiliary{Fl <: AbstractFloat}
     seasonal_ma_poly::Vector{Fl}
     seasonal_ma_names::Vector{String}
     function SARIMAHyperParametersAuxiliary{Fl}(order::SARIMAOrder) where Fl
+        ar_poly = create_poly(order.p, Fl)
+        seasonal_ar_poly = create_seasonal_poly(order.P, order.s, Fl)
+        ma_poly = create_poly(order.q, Fl)
+        seasonal_ma_poly = create_seasonal_poly(order.Q, order.s, Fl)
         return new{Fl}(
-            create_poly(order.p, Fl),
+            ar_poly,
             create_names("ar_L", order.p),
-            create_seasonal_poly(order.P, order.s, Fl),
+            seasonal_ar_poly,
             create_names("s_ar_L", order.P, order.s),
-            create_poly(order.q, Fl),
+            ma_poly,
             create_names("ma_L", order.q),
-            create_seasonal_poly(order.Q, order.s, Fl),
-            create_names("s_ma_L", order.Q, order.s),
+            seasonal_ma_poly,
+            create_names("s_ma_L", order.Q, order.s)
         )
     end
 end
@@ -183,7 +187,20 @@ function create_seasonal_poly(n::Int, s::Int, Fl::DataType)
     end
     return seasonal_poly
 end
-
+function poly_mul(p1::Vector{Fl}, p2::Vector{Fl}) where Fl
+    n, m = length(p1), length(p2)
+    if n > 1 && m > 1
+        c = zeros(Fl, m + n - 1)
+        for i in 1:n, j in 1:m
+            c[i + j - 1] += p1[i] * p2[j]
+        end
+        return c
+    elseif n <= 1
+        return p2
+    else
+        return p1
+    end
+end
 get_ar_name(model::SARIMA, i::Int) = model.hyperparameters_auxiliary.ar_names[i]
 get_ma_name(model::SARIMA, i::Int) = model.hyperparameters_auxiliary.ma_names[i]
 get_seasonal_ar_name(model::SARIMA, i::Int) = model.hyperparameters_auxiliary.seasonal_ar_names[i]
@@ -192,21 +209,25 @@ get_seasonal_ma_name(model::SARIMA, i::Int) = model.hyperparameters_auxiliary.se
 function update_T_terms!(model::SARIMA)
     update_ar_poly!(model)
     update_seasonal_ar_poly!(model)
-    p3 = Polynomial(model.hyperparameters_auxiliary.ar_poly) * Polynomial(model.hyperparameters_auxiliary.seasonal_ar_poly)
+    # TODO increaase performance
+    # We could preallocate this polynomial p3
+    p3 = poly_mul(model.hyperparameters_auxiliary.ar_poly, model.hyperparameters_auxiliary.seasonal_ar_poly)
     start_row = model.order.n_states_diff + 1
     end_row = start_row + model.order.p + model.order.s * model.order.P - 1
     col = model.order.n_states_diff + 1
-    model.system.T[start_row:end_row, col] = p3.coeffs[2:end]
+    model.system.T[start_row:end_row, col] = p3[2:end]
     return nothing
 end
 
 function update_R_terms!(model::SARIMA)
     update_ma_poly!(model)
     update_seasonal_ma_poly!(model)
-    p3 = Polynomial(model.hyperparameters_auxiliary.ma_poly) * Polynomial(model.hyperparameters_auxiliary.seasonal_ma_poly)
+    # TODO increaase performance
+    # We could preallocate this polynomial p3
+    p3 = poly_mul(model.hyperparameters_auxiliary.ma_poly, model.hyperparameters_auxiliary.seasonal_ma_poly)
[…]
