### 2020-12-13 a77598a
Added a forecasting recipe
