### 2020-12-14 16f3b20
SARIMA working and more stable
