+++ test/models/sarima.jl
@@ -64,7 +64,7 @@
     @test loglike(model) ≈ 129.732 atol = 1e-3 rtol = 1e-3
 
     model = SARIMA(log_air_passengers; order = (2, 1, 0), seasonal_order = (1, 1, 0, 4))
-    @test_broken fit!(model)
+    fit!(model)
     @test_broken loglike(model) ≈ 69.931 atol = 1e-3 rtol = 1e-3
 
     model = SARIMA(log_air_passengers; order = (0, 1, 1), seasonal_order = (0, 1, 1, 12))
