+++ Project.toml
@@ -21,8 +21,8 @@ julia = "1"
 [extras]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["CSV", "DataFrames", "Random", "Test"]
+test = ["CSV", "DataFrames", "Plots", "Test"]
