+++ Project.toml
@@ -16,6 +16,7 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Distributions = "0.23, 0.24"
 Optim = "0.20, 0.21, 0.22"
 ShiftedArrays = "1.0"
+RecipesBase = "1"
 julia = "1"
 
 [extras]
