### 2020-12-14 60f64a7
Added a forecasting recipe (#223)
* Added a forecasting recipe

* update Recipe test

* Correctly testing Recipes
