### 2020-12-14 eef44c2
Add SARIMA (#221)
* arima model working

* saarima working with polynomials

* Not stable SARIMA

* working for some sarima models

* SARIMA working and more stable

* [ci skip] remove airlines gitignore
