+++ src/optimizers.jl
@@ -23,3 +23,9 @@ function Optimizer(
 )
     return Optimizer(method, options)
 end
+
+# General to every StateSpaceModel, some of them haave trouble to converge 
+# or have numerical errors with LBFGS
+function default_optimizer(::StateSpaceModel)
+    return Optimizer(Optim.LBFGS())
+end
