### 2020-12-15 e7dc25b
Add SARIMA docs
