+++ Project.toml
@@ -19,7 +19,6 @@ Distributions = "0.23, 0.24"
 MatrixEquations = "1"
 Optim = "0.20, 0.21, 0.22"
 Polynomials = "1"
-ShiftedArrays = "1"
 RecipesBase = "1"
 ShiftedArrays = "1"
 julia = "1"
