### 2020-12-16 d657475
SARIMA docs (#225)
* Update T aand Z for performance

* Add SARIMA docs

* fix conflict
