+++ src/forecast.jl
@@ -11,7 +11,7 @@ end
     forecast(model::StateSpaceModel, steps_ahead::Int; kwargs...)
     forecast(model::StateSpaceModel, exogenous::Matrix{Fl}; kwargs...) where Fl
 
-Forecast the mean and covariance for future observations from a StateSpaceModel (SSM).
+Forecast the mean and covariance for future observations from a StateSpaceModel.
 """
 function forecast end
 
