+++ Project.toml
@@ -19,8 +19,8 @@ Distributions = "0.23, 0.24"
 MatrixEquations = "1"
 Optim = "0.20, 0.21, 0.22"
 Polynomials = "1"
-ShiftedArrays = "1"
 RecipesBase = "1"
+ShiftedArrays = "1"
 julia = "1"
 
 [extras]

[…]
