+++ src/fit.jl
@@ -8,25 +8,29 @@
 Estimate the state-space model parameters via maximum likelihood. The resulting optimal
 hyperparameters and the corresponding log-likelihood are stored within the model. You can
 choose the desired filter method (`UnivariateKalmanFilter`, `ScalarKalmanFilter`, etc.) and
-the `Optim.jl` optimization algortihm. 
+the [Optim.jl](https://github.com/JuliaNLSolvers/Optim.jl) optimization algortihm. 
 
 # Example
 ```jldoctest
 julia> model = LocalLevel(rand(100))
 LocalLevel model
+
 julia> fit!(model)
 LocalLevel model
+
 julia> model = LocalLinearTrend(LinRange(1, 100, 100) + rand(100))
 LocalLinearTrend model
+
 julia> fit!(model; optimizer = Optimizer(StateSpaceModels.Optim.NelderMead()))
 LocalLinearTrend model
 ```
 """
 function fit!(
-    model::StateSpaceModel;
+    model::SS;
     filter::KalmanFilter=default_filter(model),
     optimizer::Optimizer=default_optimizer(model),
-)
+) where SS <: StateSpaceModel
+    @assert has_fit_methods(SS)
     initial_unconstrained_hyperparameter = handle_optim_initial_hyperparameters(model)
     # TODO Should there be a try catch?
     func = TwiceDifferentiable(
@@ -47,6 +51,86 @@ function fit!(
     std_err = diag(inv(numerical_hessian))
     fill_results!(model, opt_loglikelihood, std_err)
     return model
+end
+
+"""
+    has_fit_methods(model_type::Type{<:StateSpaceModel}) -> Bool
+
+Verify if a certain `StateSpaceModel` has the necessary methods to perform `fit!``.
+"""
+function has_fit_methods(model_type::Type{<:StateSpaceModel})
+    tuple_with_model_type = Tuple{model_type}
+    m1 = hasmethod(default_filter, tuple_with_model_type)
+    m2 = hasmethod(initial_hyperparameters!, tuple_with_model_type)
+    m3 = hasmethod(constrain_hyperparameters!, tuple_with_model_type)
+    m4 = hasmethod(unconstrain_hyperparameters!, tuple_with_model_type)
+    m5 = hasmethod(fill_model_system!, tuple_with_model_type)
+    return m1 && m2 && m3 && m4 && m5
+end
+
+struct CoefficientTable{Fl<:AbstractFloat}
+    names::Vector{String}
+    coef::Vector{Fl}
+    std_err::Vector{Fl}
+    z::Vector{Fl}
+    p_value::Vector{Fl}
+
+    function CoefficientTable{Fl}(
+        names::Vector{String},
+        coef::Vector{Fl},
+        std_err::Vector{Fl},
+        z::Vector{Fl},
+        p_value::Vector{Fl},
+    ) where Fl
+        @assert length(names) ==
+                length(coef) ==
+                length(std_err) ==
+                length(z) ==
+                length(p_value)
+        return new{Fl}(names, coef, std_err, z, p_value)
+    end
+end
+
+function CoefficientTable{Fl}() where Fl
+    return CoefficientTable{Fl}(String[], Fl[], Fl[], Fl[], Fl[])
+end
+
+Base.length(coef_table::CoefficientTable) = length(coef_table.names)
+
+function Base.isempty(coef_table::CoefficientTable)
+    return isempty(coef_table.names) &&
+           isempty(coef_table.coef) &&
+           isempty(coef_table.std_err) &&
+           isempty(coef_table.z) &&
+           isempty(coef_table.p_value)
+end
+
+mutable struct Results{Fl<:AbstractFloat}
[…]
