### 2020-12-16 5b565e1
Update docs/src/manual.md
Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
