### 2020-12-16 c2b89fa
update fitted
