### 2020-12-17 65e240c
Updaaate incomplete doctest
