### 2020-12-17 2a37268
Update ARIMA example
