### 2020-12-17 476052a
Update documentation everywhere (#226)
* Update documentation everywhere

* Update docs/src/manual.md

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>

* update fitted

* fix method syntax

* Updaaate incomplete doctest

* Update ARIMA example

Co-authored-by: Raphael Saavedra <raphael.saavedra93@gmail.com>
