+++ test/runtests.jl
@@ -11,6 +11,7 @@ include("utils.jl")
 
 # Core functionality
 include("systems.jl")
+include("prints.jl")
 
 # Models
 include("models/locallevel.jl")
