### 2020-12-21 ed221ac
Add Optim v1.2 to compat in Project.toml. (#227)
