+++ Project.toml
@@ -17,7 +17,7 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 [compat]
 Distributions = "0.23, 0.24"
 MatrixEquations = "1"
-Optim = "0.20, 0.21, 0.22"
+Optim = "0.20, 0.21, 0.22, 1.2"
 Polynomials = "1"
 RecipesBase = "1"
 ShiftedArrays = "1"
