+++ src/models/unobserved_components.jl
@@ -1,0 +1,295 @@
+# In total we can have trend, slope, seasonal, cycle and exogenous
+# Maybe autoregressive
+function validate_trend(trend::String)
+    @assert trend in [
+        "local level",
+        "random walk",
+        # "deterministic trend",
+        # "linear deterministic trend",
+        # "linear trend",
+        "local linear trend",
+        # "damped local lineaar trend",
+        # "random walk with drift"
+        "smooth trend"
+    ]
+    return true
+end
+function parse_trend(trend::String)
+    validate_trend(trend)
+    (has_irregular, has_trend, stochastic_trend, 
+    has_slope, stochastic_slope) = (false, false, false, false, false)
+    if trend == "local level"
+        (has_irregular, has_trend, stochastic_trend, 
+        has_slope, stochastic_slope) = (true, true, true, false, false)
+    elseif trend == "random walk"
+        (has_irregular, has_trend, stochastic_trend, 
+        has_slope, stochastic_slope) = (false, true, true, false, false)
+    elseif trend == "local linear trend"
+        (has_irregular, has_trend, stochastic_trend, 
+        has_slope, stochastic_slope) = (true, true, true, true, true)
+    elseif trend == "smooth trend"
+        (has_irregular, has_trend, stochastic_trend, 
+        has_slope, stochastic_slope) = (true, true, false, true, true)
+    end
+    # Validate trend booleans
+    if !has_trend && stochastic_trend
+        error("Invalid trend specification.")
+    end
+    if !has_trend && has_slope
+        error("Invalid trend specification.")
+    end
+    if !has_slope && stochastic_slope
+        error("Invalid trend specification.")
+    end
+    return (has_irregular, has_trend, stochastic_trend, has_slope, stochastic_slope)
+end
+function validate_seasonal(seasonal::String)
+    spl = split(seasonal)
+    # TODO better error messaage
+    # Maybe a no string can be also valid
+    @assert length(spl) == 2
+    @assert spl[1] in ["deterministic", "stochastic"]
+    return true
+end
+function parse_seasonal(seasonal::String)
+    validate_seasonal(seasonal)
+    spl = split(seasonal)
+    stochastic_seasonal = spl[1] == "stochastic"
+    seasonal_freq = parse(Int, spl[2])
+    has_seasonal = seasonal_freq == 0 ? false : true
+    return has_seasonal, stochastic_seasonal, seasonal_freq
+end
+
+@doc raw"""
+# TODO
+- TREND
+
+- SEASONAL
+
+
+# References
+ * Durbin, James, & Siem Jan Koopman. (2012). "Time Series Analysis by State Space Methods: Second Edition." Oxford University Press.
+"""
+mutable struct UnobservedComponents <: StateSpaceModel
+    hyperparameters::HyperParameters
+    trend::String
+    seasonal::String
+    has_irregular::Bool
+    has_trend::Bool
+    stochastic_trend::Bool
+    has_slope::Bool
+    stochastic_slope::Bool
+    # damped_slope::Bool TODO
+    has_seasonal::Bool
+    stochastic_seasonal::Bool
+    # TODO
+    # has_cycle
+    # stochastic_cycle
+    # damped_cycle
+    seasonal_freq::Int
+    system::LinearUnivariateTimeInvariant
+    results::Results
+
+    function UnobservedComponents(y::Vector{Fl}; 
+                                  trend::String = "local level", # Multiple predefined options
+                                  seasonal::String = "deterministic 0" # Stochastic or deterministic and period
[…]
