+++ src/models/unobserved_components.jl
@@ -44,20 +44,67 @@ function parse_trend(trend::String)
     return (has_irregular, has_trend, stochastic_trend, has_slope, stochastic_slope)
 end
 function validate_seasonal(seasonal::String)
-    spl = split(seasonal)
-    # TODO better error messaage
-    # Maybe a no string can be also valid
-    @assert length(spl) == 2
-    @assert spl[1] in ["deterministic", "stochastic"]
-    return true
+    if seasonal == "no"
+        return true
+    else
+        spl = split(seasonal)
+        # TODO better error messaage
+        # Maybe a no string can be also valid
+        @assert length(spl) == 2
+        @assert spl[1] in ["deterministic", "stochastic"]
+        return true
+    end
 end
 function parse_seasonal(seasonal::String)
     validate_seasonal(seasonal)
-    spl = split(seasonal)
-    stochastic_seasonal = spl[1] == "stochastic"
-    seasonal_freq = parse(Int, spl[2])
-    has_seasonal = seasonal_freq == 0 ? false : true
-    return has_seasonal, stochastic_seasonal, seasonal_freq
+    if seasonal == "no"
+        has_seasonal, stochastic_seasonal, seasonal_freq = (false, false, 0)
+        return has_seasonal, stochastic_seasonal, seasonal_freq
+    else
+        spl = split(seasonal)
+        stochastic_seasonal = spl[1] == "stochastic"
+        seasonal_freq = parse(Int, spl[2])
+        has_seasonal = seasonal_freq == 0 ? false : true
+        return has_seasonal, stochastic_seasonal, seasonal_freq
+    end
+end
+function validate_cycle(cycle::String)
+    if cycle == "no"
+        return true
+    else
+        spl = split(cycle)
+        # TODO better error messaage
+        # Maybe a no string can be also valid
+        if length(spl) == 1
+            @assert spl[1] in ["deterministic", "stochastic"]
+            return true
+        elseif length(spl) == 2
+            @assert spl[1] in ["deterministic", "stochastic"]
+            @assert spl[2] == "damped"
+            return true
+        end
+    end
+    return false
+end
+function parse_cycle(cycle::String)
+    validate_cycle(cycle)
+    if cycle == "no"
+        has_cycle, stochastic_cycle, damped_cycle = (false, false, false)
+        return has_cycle, stochastic_cycle, damped_cycle
+    else
+        spl = split(cycle)
+        has_cycle = true
+        # TODO better error messaage
+        # Maybe a no string can be also valid
+        if length(spl) == 1
+            stochastic_cycle = spl[1] == "stochastic"
+            return has_cycle, stochastic_cycle, false
+        elseif length(spl) == 2
+            stochastic_cycle = spl[1] == "stochastic"
+            damped_cycle = spl[2] == "damped"
+            return has_cycle, stochastic_cycle, damped_cycle
+        end
+    end
 end
 
 @doc raw"""
@@ -66,6 +113,7 @@ end
 
 - SEASONAL
 
+- CYCLE
 
 # References
  * Durbin, James, & Siem Jan Koopman. (2012). "Time Series Analysis by State Space Methods: Second Edition." Oxford University Press.
@@ -74,6 +122,7 @@ mutable struct UnobservedComponents <: StateSpaceModel
     hyperparameters::HyperParameters
     trend::String
     seasonal::String
+    cycle::String
     has_irregular::Bool
     has_trend::Bool
     stochastic_trend::Bool
@@ -82,28 +131,28 @@ mutable struct UnobservedComponents <: StateSpaceModel
     # damped_slope::Bool TODO
     has_seasonal::Bool
     stochastic_seasonal::Bool
-    # TODO
[…]
