### 2020-12-23 55a6908
remove unecessary mutable in some structs
