### 2020-12-23 1ff14db
Add UnobservedComponents (#228)
* Add UnobservedComponents

* Added cycles

* remove unecessary mutable in some structs
