+++ src/models/unobserved_components.jl
@@ -1,0 +1,428 @@
+# In total we can have trend, slope, seasonal, cycle and exogenous
+# Maybe autoregressive
+function validate_trend(trend::String)
+    @assert trend in [
+        "local level",
+        "random walk",
+        # "deterministic trend",
+        # "linear deterministic trend",
+        # "linear trend",
+        "local linear trend",
+        # "damped local lineaar trend",
+        # "random walk with drift"
+        "smooth trend"
+    ]
+    return true
+end
+function parse_trend(trend::String)
+    validate_trend(trend)
+    (has_irregular, has_trend, stochastic_trend, 
+    has_slope, stochastic_slope) = (false, false, false, false, false)
+    if trend == "local level"
+        (has_irregular, has_trend, stochastic_trend, 
+        has_slope, stochastic_slope) = (true, true, true, false, false)
+    elseif trend == "random walk"
+        (has_irregular, has_trend, stochastic_trend, 
+        has_slope, stochastic_slope) = (false, true, true, false, false)
+    elseif trend == "local linear trend"
+        (has_irregular, has_trend, stochastic_trend, 
+        has_slope, stochastic_slope) = (true, true, true, true, true)
+    elseif trend == "smooth trend"
+        (has_irregular, has_trend, stochastic_trend, 
+        has_slope, stochastic_slope) = (true, true, false, true, true)
+    end
+    # Validate trend booleans
+    if !has_trend && stochastic_trend
+        error("Invalid trend specification.")
+    end
+    if !has_trend && has_slope
+        error("Invalid trend specification.")
+    end
+    if !has_slope && stochastic_slope
+        error("Invalid trend specification.")
+    end
+    return (has_irregular, has_trend, stochastic_trend, has_slope, stochastic_slope)
+end
+function validate_seasonal(seasonal::String)
+    if seasonal == "no"
+        return true
+    else
+        spl = split(seasonal)
+        # TODO better error messaage
+        # Maybe a no string can be also valid
+        @assert length(spl) == 2
+        @assert spl[1] in ["deterministic", "stochastic"]
+        return true
+    end
+end
+function parse_seasonal(seasonal::String)
+    validate_seasonal(seasonal)
+    if seasonal == "no"
+        has_seasonal, stochastic_seasonal, seasonal_freq = (false, false, 0)
+        return has_seasonal, stochastic_seasonal, seasonal_freq
+    else
+        spl = split(seasonal)
+        stochastic_seasonal = spl[1] == "stochastic"
+        seasonal_freq = parse(Int, spl[2])
+        has_seasonal = seasonal_freq == 0 ? false : true
+        return has_seasonal, stochastic_seasonal, seasonal_freq
+    end
+end
+function validate_cycle(cycle::String)
+    if cycle == "no"
+        return true
+    else
+        spl = split(cycle)
+        # TODO better error messaage
+        # Maybe a no string can be also valid
+        if length(spl) == 1
+            @assert spl[1] in ["deterministic", "stochastic"]
+            return true
+        elseif length(spl) == 2
+            @assert spl[1] in ["deterministic", "stochastic"]
+            @assert spl[2] == "damped"
+            return true
+        end
+    end
+    return false
+end
+function parse_cycle(cycle::String)
+    validate_cycle(cycle)
+    if cycle == "no"
+        has_cycle, stochastic_cycle, damped_cycle = (false, false, false)
[…]
