### 2020-12-24 f7e6103
small bug fixes
