### 2020-12-24 4e68afe
small bug fixes (#229)
