### 2021-01-14 #28 @raphaelsaavedra
Hi, since the submission, we have made a significant refactor of the package. As a result, we removed the paper from the master, as it had some outdated API documentation. I could perhaps open a branch using an old version and point whedon to it?
