### 2021-01-14 #28 @matbesancon
Do you have the paper on a special branch? What I would suggest is:
1. get your paper on the repo and compiling
2. if you did some changes, we will run it by reviewers
3. publish
4. profit
