### 2021-01-14 #28 @matbesancon
yes that would work. Also, update the paper to reflect the current package state :)
