### 2021-01-14 #28 @mschauer
Hi, thank you for pointing out the new version, I can see that the package has been redesigned and matured substantially. Now this is a package I can see having a niche in the ecosystem. I am in particular happy about https://github.com/LAMPSPUC/StateSpaceModels.jl/commit/38685eeb9a714dca52a23a6968ad6f4511804597, the introduction of https://github.com/LAMPSPUC/StateSpaceModels.jl/blob/master/src/systems.jl and https://github.com/LAMPSPUC/StateSpaceModels.jl/pull/174, very much in line with our first discussions about your design. I support writing about this modern StateSpaceModels.jl, reflecting the clean new design (no need to preserve the older approaches and the older paper) and then wrapping this up
