### 2021-02-02 68b08a1
Added some docs for unobserved components
