+++ test/models/locallevelcycle.jl
@@ -5,7 +5,7 @@
     model = LocalLevelCycle(rj_temp)
     fit!(model)
     # TODO check with other software maybe statsmodels
-    @test loglike(model) ≈ -617.92385 atol = 1e-5 rtol = 1e-5
+    @test loglike(model) ≈ -619.8679932 atol = 1e-5 rtol = 1e-5
 
     # forecasting
     forec = forecast(model, 52)
