### 2021-02-02 1b1255f
Added some docs for unobserved components (#230)
* Added some docs for unobserved components

* fix imprecision on cycle component and docs
