+++ src/models/unobserved_components.jl
@@ -74,7 +74,6 @@ function validate_cycle(cycle::String)
     else
         spl = split(cycle)
         # TODO better error messaage
-        # Maybe a no string can be also valid
         if length(spl) == 1
             @assert spl[1] in ["deterministic", "stochastic"]
             return true
@@ -108,12 +107,121 @@ function parse_cycle(cycle::String)
 end
 
 @doc raw"""
-# TODO
-- TREND
+    UnobservedComponents(
+        y::Vector{Fl}; 
+        trend::String = "local level",
+        seasonal::String = "no"
+        cycle::String = "no"
+    ) where Fl
 
-- SEASONAL
+An unobserved components model that can have trend/level, seasonal and cycle components. 
+Each component should be specified by strings, if the component is not desired in the model 
+a string with "no" can be passed as keyword argument. 
+    
+These models take the general form 
 
-- CYCLE
+```math
+\begin{gather*}
+    \begin{aligned}
+    y_t = \mu_t + \gamma_t + c_t + \varepsilon_t
+    \end{aligned}
+\end{gather*}
+```
+where ``y_t`` refers to the observation vector at time ``t``,
+``\mu_t`` refers to the trend component, ``\gamma_t`` refers to the
+seasonal component, ``c_t`` refers to the cycle, and
+``\varepsilon_t`` is the irregular. The modeling details of these
+components are given below.
+
+### **Trend**
+
+The trend component can be modeled in a lot of different ways, usually it is called level when
+there is no slope component. The modelling options can be expressed as in the example `trend = "local level"`.
+
+* Local Level
+string: `"local level"`
+```math
+\begin{gather*}
+    \begin{aligned}
+        y_{t} &=  \mu_{t} + \varepsilon_{t} \quad \varepsilon_{t} \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\
+        \mu_{t+1} &= \mu_{t} + \eta_{t} \quad \eta_{t} \sim \mathcal{N}(0, \sigma^2_{\eta})\\
+    \end{aligned}
+\end{gather*}
+```
+
+* Random Walk
+string: `"random walk"`
+```math
+\begin{gather*}
+    \begin{aligned}
+        y_{t} &=  \mu_{t}\\
+        \mu_{t+1} &= \mu_{t} + \eta_{t} \quad \eta_{t} \sim \mathcal{N}(0, \sigma^2_{\eta})\\
+    \end{aligned}
+\end{gather*}
+```
+
+* Local Linear Trend
+string: `"local linear trend"`
+```math
+\begin{gather*}
+    \begin{aligned}
+        y_{t} &=  \mu_{t} + \gamma_{t} + \varepsilon_{t} \quad &\varepsilon_{t} \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\
+        \mu_{t+1} &= \mu_{t} + \nu_{t} + \xi_{t} \quad &\xi_{t} \sim \mathcal{N}(0, \sigma^2_{\xi})\\
+        \nu_{t+1} &= \nu_{t} + \zeta_{t} \quad &\zeta_{t} \sim \mathcal{N}(0, \sigma^2_{\zeta})\\
+    \end{aligned}
+\end{gather*}
+```
+
+* Smooth Trend
+string: `"smooth trend"`
+```math
+\begin{gather*}
+    \begin{aligned}
+        y_{t} &=  \mu_{t} + \gamma_{t} + \varepsilon_{t} \quad &\varepsilon_{t} \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\
[…]
