### 2021-02-03 c337af7
Added vehicle tracking example
