### 2021-02-04 7a24d1e
Add scenarios plots (#236)
* Add scenarios plots

* Fix idx names
