+++ test/visualization/forecast.jl
@@ -7,4 +7,8 @@
     forec = forecast(model, 12)
     rec = RecipesBase.apply_recipe(Dict{Symbol, Any}(), model, forec)
     @test length(rec) == 4
+    # simulating
+    scen = simulate_scenarios(model, 12, 100)
+    rec = RecipesBase.apply_recipe(Dict{Symbol, Any}(), model, scen)
+    @test length(rec) == 2
 end
