### 2021-02-09 3930ba7
Add UC plot recipe
