### 2021-02-09 757d55e
Add UC plot recipe (#237)
