+++ test/visualization/forecast.jl
@@ -1,4 +1,4 @@
-@testset "Visualization" begin
+@testset "Visualization Forecast" begin
     air_passengers = CSV.read(StateSpaceModels.AIR_PASSENGERS, DataFrame)
     log_air_passengers = log.(air_passengers.passengers)
     model = BasicStructural(log_air_passengers, 12)
