+++ test/models/unobserved_components.jl
@@ -1,41 +1,60 @@
 @testset "UnobservedComponents" begin
-    @test has_fit_methods(UnobservedComponents)
+    @test StateSpaceModels.has_fit_methods(StateSpaceModels.UnobservedComponents)
 
     nile = CSV.read(StateSpaceModels.NILE, DataFrame)
-    model = UnobservedComponents(nile.flow)
-    fit!(model)
-    @test loglike(model) ≈ -632.5376 atol = 1e-5 rtol = 1e-5
-    filt = kalman_filter(model)
-    smoother = kalman_smoother(model)
-    forec = forecast(model, 10)
+    model = StateSpaceModels.UnobservedComponents(nile.flow)
+    @show "fit começa $(now())"
+    StateSpaceModels.fit!(model)
+    @show "fit termina $(now())"
+    @test StateSpaceModels.loglike(model) ≈ -632.5376 atol = 1e-5 rtol = 1e-5
+    filt = StateSpaceModels.kalman_filter(model)
+    smoother = StateSpaceModels.kalman_smoother(model)
+    forec = StateSpaceModels.forecast(model, 10)
+    @test monotone_forecast_variance(forec)
+
+    nile = CSV.read(StateSpaceModels.NILE, DataFrame)
+    model = StateSpaceModels.UnobservedComponents(nile.flow)
+    @show "fit começa $(now())"
+    StateSpaceModels.fit!(model)
+    @show "fit termina $(now())"
+    @test StateSpaceModels.loglike(model) ≈ -632.5376 atol = 1e-5 rtol = 1e-5
+    filt = StateSpaceModels.kalman_filter(model)
+    smoother = StateSpaceModels.kalman_smoother(model)
+    forec = StateSpaceModels.forecast(model, 10)
     @test monotone_forecast_variance(forec)
 
     finland_fatalities = CSV.read(StateSpaceModels.VEHICLE_FATALITIES, DataFrame)
     log_finland_fatalities = log.(finland_fatalities.ff)
-    model = UnobservedComponents(log_finland_fatalities; trend = "local linear trend")
-    fit!(model)
-    @test loglike(model) ≈ 26.740 atol = 1e-5 rtol = 1e-5
-    forec = forecast(model, 10)
+    model = StateSpaceModels.UnobservedComponents(log_finland_fatalities; trend = "local linear trend")
+    @show "fit começa $(now())"
+    StateSpaceModels.fit!(model)
+    @show "fit termina $(now())"
+    @test StateSpaceModels.loglike(model) ≈ 26.740 atol = 1e-5 rtol = 1e-5
+    forec = StateSpaceModels.forecast(model, 10)
     @test monotone_forecast_variance(forec)
 
     air_passengers = CSV.read(StateSpaceModels.AIR_PASSENGERS, DataFrame)
     log_air_passengers = log.(air_passengers.passengers)
-    model = UnobservedComponents(log_air_passengers; trend = "local linear trend", seasonal = "stochastic 12")
-    fit!(model)
-    @test loglike(model) ≈ 234.33641 atol = 1e-5 rtol = 1e-5
-    forec = forecast(model, 10)
+    model = StateSpaceModels.UnobservedComponents(log_air_passengers; trend = "local linear trend", seasonal = "stochastic 12")
+    @show "fit começa $(now())"
+    StateSpaceModels.fit!(model)
+    @show "fit termina $(now())"
+    @test StateSpaceModels.loglike(model) ≈ 234.33641 atol = 1e-5 rtol = 1e-5
+    forec = StateSpaceModels.forecast(model, 10)
     @test monotone_forecast_variance(forec)
 
[…]
