+++ src/filters/robust_kalman_filter.jl
@@ -1,0 +1,432 @@
+mutable struct RobustKalmanState{Fl<:AbstractFloat}
+    v::Fl
+    F::Fl
+    att::Vector{Fl}
+    a::Vector{Fl}
+    Ptt::Matrix{Fl}
+    P::Matrix{Fl}
+    llk::Fl
+    steady_state::Bool
+    # Auxiliary matrices
+    P_to_check_steady_state::Matrix{Fl}
+    ZP::Vector{Fl}
+    TPtt::Matrix{Fl}
+    function RobustKalmanState(a1::Vector{Fl}, P1::Matrix{Fl}) where Fl
+        m = length(a1)
+        P_to_check_steady_state = zeros(Fl, m, m)
+        ZP = zeros(Fl, m)
+        TPtt = zeros(Fl, m, m)
+
+        return new{Fl}(
+            zero(Fl),
+            zero(Fl),
+            zeros(Fl, m),
+            a1,
+            zeros(Fl, m, m),
+            P1,
+            zero(Fl),
+            false,
+            P_to_check_steady_state,
+            ZP,
+            TPtt,
+        )
+    end
+end
+
+function save_a1_P1_in_filter_output!(
+    filter_output::FilterOutput{Fl}, kalman_state::RobustKalmanState{Fl}
+) where Fl
+    filter_output.a[1] = deepcopy(kalman_state.a)
+    filter_output.P[1] = deepcopy(kalman_state.P)
+    return filter_output
+end
+
+function save_kalman_state_in_filter_output!(
+    filter_output::FilterOutput{Fl}, kalman_state::RobustKalmanState{Fl}, t::Int
+) where Fl
+    filter_output.v[t] = copy(fill(kalman_state.v, 1))
+    filter_output.F[t] = copy(fill(kalman_state.F, 1, 1))
+    filter_output.a[t + 1] = copy(kalman_state.a)
+    filter_output.att[t] = copy(kalman_state.att)
+    filter_output.P[t + 1] = copy(kalman_state.P)
+    filter_output.Ptt[t] = copy(kalman_state.Ptt)
+    # There is no diffuse part in RobustKalmanState
+    filter_output.Pinf[t] = copy(fill(zero(Fl), size(kalman_state.Ptt)))
+    return filter_output
+end
+
+# Univariate Kalman filter with the recursions as described
+# in Koopman's book TODO
+"""
+    RobustKalmanFilter{Fl <: AbstractFloat}
+
+A Kalman filter that is tailored to univariate systems, exploiting the fact that the
+dimension of the observations at any time period is 1.
+
+# TODO equations and descriptions of a1 and P1
+"""
+mutable struct RobustKalmanFilter{Fl<:AbstractFloat} <: KalmanFilter
+    steadystate_tol::Fl
+    a1::Vector{Fl}
+    P1::Matrix{Fl}
+    skip_llk_instants::Int
+    kalman_state::RobustKalmanState
+
+    function RobustKalmanFilter(
+        a1::Vector{Fl},
+        P1::Matrix{Fl},
+        skip_llk_instants::Int=length(a1),
+        steadystate_tol::Fl=Fl(1e-5),
+    ) where Fl
+        kalman_state = RobustKalmanState(copy(a1), copy(P1))
+        return new{Fl}(steadystate_tol, a1, P1, skip_llk_instants, kalman_state)
+    end
+end
+
+function reset_filter!(kf::RobustKalmanFilter{Fl}) where Fl
+    copyto!(kf.kalman_state.a, kf.a1)
+    copyto!(kf.kalman_state.P, kf.P1)
+    set_state_llk_to_zero!(kf.kalman_state)
+    fill!(kf.kalman_state.P_to_check_steady_state, zero(Fl))
+    fill!(kf.kalman_state.ZP, zero(Fl))
[…]
