+++ src/models/unobserved_components.jl
@@ -448,7 +448,7 @@ function default_filter(model::UnobservedComponents)
     steadystate_tol = Fl(1e-5)
     a1 = zeros(Fl, num_states(model))
     P1 = Fl(1e6) .* Matrix{Fl}(I, num_states(model), num_states(model))
-    return UnivariateKalmanFilter(a1, P1, num_states(model), steadystate_tol)
+    return RobustKalmanFilter(a1, P1, num_states(model), steadystate_tol)
 end
 
 function initial_hyperparameters!(model::UnobservedComponents)
