+++ test/runtests.jl
@@ -27,3 +27,4 @@ include("models/linear_regression.jl")
 # Visualization
 include("visualization/forecast.jl")
 include("visualization/unobserved_components.jl")
+include("visualization/backtest.jl")
