### 2021-03-03 7458106
Add backtest docs and forecasting docs on manual
