+++ docs/src/examples.md
@@ -210,6 +210,6 @@ using CSV, DataFrames
 air_passengers = CSV.read(StateSpaceModels.AIR_PASSENGERS, DataFrame)
 log_air_passengers = log.(air_passengers.passengers)
 model = BasicStructural(log_air_passengers, 12)
-b = backtest(model, log_air_passengers, 24, 50)
+b = backtest(model, 24, 50)
 plot(b, "Basic structural model")
 ```
