### 2021-03-04 492c098
Implement a backtest procedure (#244)
* Add backtest

* Add backtest docs and forecasting docs on manual

* remove y from backtest
