+++ src/forecast.jl
@@ -46,13 +46,13 @@ function forecast(
     for i in 1:steps_ahead
         if isunivariate(model)
             expected_value[i] = [
-                dot(model.system.Z, fo.a[end - steps_ahead + i]) + model.system.d
+                dot(model.system.Z, fo.a[end - steps_ahead + i - 1]) + model.system.d
             ]
         else
             expected_value[i] =
-                model.system.Z * fo.a[end - steps_ahead + i] .+ model.system.d
+                model.system.Z * fo.a[end - steps_ahead + i - 1] .+ model.system.d
         end
-        covariance[i] = fo.F[end - steps_ahead + i]
+        covariance[i] = fo.F[end - steps_ahead + i - 1]
     end
     return Forecast{Fl}(expected_value, covariance)
 end
@@ -89,15 +89,15 @@ function forecast(
     for i in 1:steps_ahead
         if isunivariate(model)
             expected_value[i] = [
-                dot(model.system.Z[end - steps_ahead + i], fo.a[end - steps_ahead + i]) +
-                model.system.d[end - steps_ahead + i],
+                dot(model.system.Z[end - steps_ahead + i - 1], fo.a[end - steps_ahead + i - 1]) +
+                model.system.d[end - steps_ahead + i - 1],
             ]
         else
             expected_value[i] =
-                model.system.Z[end - steps_ahead + i] * fo.a[end - steps_ahead + i] +
-                model.system.d[end - steps_ahead + i]
+                model.system.Z[end - steps_ahead + i - 1] * fo.a[end - steps_ahead + i - 1] +
+                model.system.d[end - steps_ahead + i - 1]
         end
-        covariance[i] = fo.F[end - steps_ahead + i]
+        covariance[i] = fo.F[end - steps_ahead + i - 1]
     end
     return Forecast{Fl}(expected_value, covariance)
 end
@@ -120,7 +120,7 @@ function simulate_scenarios(
     # Query the type of model elements
     Fl = typeof_model_elements(model)
     fo = kalman_filter(model)
-    last_state = fo.a[end]
+    last_state = fo.a[end - 1]
     num_series = size(model.system.y, 2)
 
     scenarios = Array{Fl,3}(undef, steps_ahead, num_series, n_scenarios)
