+++ test/models/sarima.jl
@@ -12,6 +12,7 @@
     @test monotone_forecast_variance(forec)
     # Prediction from Pyhton statsmodels
     predicted_mean = [
+        -1.11949821,
         -0.72809028,
         -0.47352952,
         -0.30797034,
@@ -20,16 +21,15 @@
         -0.08472164,
         -0.05510058,
         -0.03583588,
-        -0.02330665,
-        -0.01515799,
+        -0.02330665
     ]
     @test predicted_mean ≈ vcat(forec.expected_value...) atol = 1e-3
     # simualting
-    scenarios = simulate_scenarios(model, 10, 100_000)
+    @time scenarios = simulate_scenarios(model, 10, 30_000)
     # Values are very close to 0.0 so we test with absolute tolerance
     # It attains 1e-3 when we make 10M simulations, which is too much
     # computation for a rather simple test.
-    test_scenarios_adequacy_with_forecast(forec, scenarios; atol=1e-1)
+    test_scenarios_adequacy_with_forecast(forec, scenarios; atol=5e-2)
 
     missing_obs = [6, 16, 26, 36, 46, 56, 66, 72, 73, 74, 75, 76, 86, 96]
     missing_dinternet = copy(dinternet)
