### 2021-03-10 e5f6a53
Fix idx forecast (#246)
* Fix index in forecasting

* bump version

* Fix idx of forecasting

* remove StateSpaceModels  prefix

* fix atol

* update number of scenarios
