+++ src/models/unobserved_components.jl
@@ -448,7 +448,9 @@ function default_filter(model::UnobservedComponents)
     steadystate_tol = Fl(1e-5)
     a1 = zeros(Fl, num_states(model))
     P1 = Fl(1e6) .* Matrix{Fl}(I, num_states(model), num_states(model))
-    return RobustKalmanFilter(a1, P1, num_states(model), steadystate_tol)
+    lambda_o = zero(Fl)
+    lambda_m = zero(Fl)
+    return RobustKalmanFilter(a1, P1, num_states(model), steadystate_tol, lambda_o, lambda_m)
 end
 
 function initial_hyperparameters!(model::UnobservedComponents)
