### 2021-03-11 #28 @matbesancon
Thanks for your answer. As highlighted before, the package is now in a much more interesting and feature-complete state (no pun intended). These changes add value to the proceeding and should be reflected there.

There is no blocking point to finish this within a month, I just wanted to know whether to wait or not, there is substantial work that went in this from different sides and it would be sad to throw that away now that the bulk of the work is done.

I will let the author action required to reflect the changes in the proceeding
