+++ script.jl
@@ -17,18 +17,27 @@ plot!(att)
 include("/Users/guilhermebodin/Documents/StateSpaceModels.jl/src/StateSpaceModels.jl")
 nile = CSV.File(StateSpaceModels.NILE) |> DataFrame
 nile.flow[50] += 2000
-model = StateSpaceModels.UnobservedComponents(nile.flow)
-Fl = Float64
-steadystate_tol = Fl(1e-5)
-n_states = StateSpaceModels.num_states(model)
-a1 = zeros(Fl, n_states)
-P1 = Fl(1e6) .* Matrix{Fl}(I, n_states, n_states)
-lambda_o = 1.0
-lambda_m = zero(Fl)
-filt =  StateSpaceModels.RobustKalmanFilter(a1, P1, n_states, steadystate_tol, lambda_o, lambda_m)
-StateSpaceModels.fit!(model; filter = filt, optimizer = StateSpaceModels.Optimizer(Optim.LBFGS(), Optim.Options(;show_trace = true)))
-StateSpaceModels.loglike(model; filter = filt)
-kf = StateSpaceModels.kalman_filter(model; filter = filt)
+lambda_os = [0.0001, 0.00001, 0.000001, 10.0]
+llks = []
+models = []
+for lambda_o in lambda_os
+    @show "$(length(llks)+1) of $(length(lambda_os))"
+    model = StateSpaceModels.UnobservedComponents(nile.flow)
+    Fl = Float64
+    steadystate_tol = Fl(1e-5)
+    n_states = StateSpaceModels.num_states(model)
+    a1 = zeros(Fl, n_states)
+    P1 = Fl(1e6) .* Matrix{Fl}(I, n_states, n_states)
+    lambda_m = zero(Fl)
+    filt =  StateSpaceModels.RobustKalmanFilter(a1, P1, n_states, steadystate_tol, lambda_o, lambda_m)
+    StateSpaceModels.fit!(model; filter = filt, optimizer = StateSpaceModels.Optimizer(Optim.LBFGS(), Optim.Options(; g_tol = 1e-6, show_trace = true)))
+    push!(llks, StateSpaceModels.loglike(model; filter = filt))
+    push!(models, model)
+end
+models
+llks
+m = models[1]
+kf = StateSpaceModels.kalman_filter(m)
 att = StateSpaceModels.get_filtered_state(kf)
 plot(nile.flow)
 plot!(att)
