### 2021-03-30 f35ac04
Add plotdiagnostics
