+++ test/visualization/diagnostics.jl
@@ -1,0 +1,9 @@
+@testset "Visualization Diagnostics" begin
+    # air_passengers = CSV.read(StateSpaceModels.AIR_PASSENGERS, DataFrame)
+    # log_air_passengers = log.(air_passengers.passengers)
+    # model = BasicStructural(log_air_passengers, 12)
+    # fit!(model)
+    # kf = kalman_filter(model)
+    # rec = RecipesBase.apply_recipe(Dict{Symbol, Any}(), kf)
+    # @test length(rec) == 9
+end
