+++ test/visualization/forecast.jl
@@ -1,5 +1,5 @@
 @testset "Visualization Forecast" begin
-    air_passengers = CSV.read(StateSpaceModels.AIR_PASSENGERS, DataFrame)
+    air_passengers = CSV.File(StateSpaceModels.AIR_PASSENGERS) |> DataFrame
     log_air_passengers = log.(air_passengers.passengers)
     model = BasicStructural(log_air_passengers, 12)
     fit!(model)
