+++ Project.toml
@@ -22,7 +22,7 @@ Optim = "0.20, 0.21, 0.22, 1.2"
 Polynomials = "1"
 RecipesBase = "1"
 ShiftedArrays = "1"
-StatsBase = "^0"
+StatsBase = "0.33"
 julia = "1"
 
 [extras]
