### 2021-04-02 42f6b6f
Improve table print and add model name
