+++ src/prints.jl
@@ -1,12 +1,13 @@
 function Base.show(io::IO, results::Results{Fl}) where Fl
     if !isempty(results)
-        println(io, "                         Results                        ")
-        println(io, "========================================================")
+        println(io, "                             Results                           ")
+        println(io, "===============================================================" * adjust_charachters("=", results.coef_table))
+        println(io, "Model:                        ", results.model_name)
         println(io, "Number of observations:       ", results.num_observations)
         println(io, "Number of unknown parameters: ", results.num_hyperparameters)
-        println(io, "Log-likelihood:               ", @sprintf("%.3f", results.llk))
-        println(io, "AIC:                          ", @sprintf("%.3f", results.aic))
-        println(io, "BIC:                          ", @sprintf("%.3f", results.bic))
+        println(io, "Log-likelihood:               ", pretty_number(results.llk))
+        println(io, "AIC:                          ", pretty_number(results.aic))
+        println(io, "BIC:                          ", pretty_number(results.bic))
         print_coef_table(io, results.coef_table)
     else
         # Up to discussion of what should be the behaviour
@@ -16,48 +17,57 @@ function Base.show(io::IO, results::Results{Fl}) where Fl
 end
 
 function Base.show(io::IO, model::StateSpaceModel)
-    print(io, typeof(model), " model")
+    print(io, model_name(model), " model")
     return nothing
 end
 
 function print_coef_table(io::IO, coef_table::CoefficientTable{Fl}) where Fl
-    println(io, "--------------------------------------------------------")
-    println(io, "Parameter      Estimate   Std.Error     z stat   p-value")
+    # The repeats adjust the space for large names
+    println(io, "---------------------------------------------------------------" * adjust_charachters("-", coef_table))
+    println(io, "Parameter" * adjust_charachters(" ", coef_table) * "      Estimate      Std.Error      z stat      p-value")
     offset = 1
     for i in 1:length(coef_table)
         p_c, p_std, p_z_stat, p_p_val = print_coef(coef_table, offset)
-        println(io, build_print(p_c, p_std, p_z_stat, p_p_val, coef_table.names[offset]))
+        println(io, build_print(p_c, p_std, p_z_stat, p_p_val, offset, coef_table))
         offset += 1
     end
     return nothing
 end
 
 function print_coef(coef_table::CoefficientTable{Fl}, offset::Int) where Fl
-    p_c = @sprintf("%.4f", coef_table.coef[offset])
-    p_std = if isnan(coef_table.std_err[offset])
-        " - "
-    else
-        @sprintf("%.3f", coef_table.std_err[offset])
-    end
-    p_z_stat = isnan(coef_table.z[offset]) ? " - " : @sprintf("%.3f", coef_table.z[offset])
-    p_p_val = if isnan(coef_table.p_value[offset])
-        " - "
-    else
-        @sprintf("%.3f", coef_table.p_value[offset])
-    end
+    p_c = pretty_number(coef_table.coef[offset])
+    p_std = pretty_number(coef_table.std_err[offset])
+    p_z_stat = pretty_number(coef_table.z[offset])
+    p_p_val = pretty_number(coef_table.p_value[offset])
     return p_c, p_std, p_z_stat, p_p_val
 end
 
-function build_print(p_c, p_std, p_z_stat, p_p_val, param)
+function build_print(p_c, p_std, p_z_stat, p_p_val, offset, coef_table)
+    max_len_name = maximum(length.(coef_table.names))
     p = ""
-    p *= param
[…]
