### 2021-04-06 5ca7641
Add auto_ets function
