### 2021-04-06 0786dde
fix auto_ets
