### 2021-04-15 b5c4e17
separe unit_root function into assert_stationarity and assert_invertibility
