### 2021-04-16 d4ea568
Add dynamic autoregressive model
