+++ src/models/common.jl
@@ -3,3 +3,5 @@ num_states(model::StateSpaceModel) = num_states(system(model))
 system(model::StateSpaceModel) = model.system
 isunivariate(model::StateSpaceModel) = isa(model.system.y, Vector)
 model_name(model::StateSpaceModel) = "$(typeof(model))"
+length_observations(model::StateSpaceModel) = length(model.system.y)
+observations(model::StateSpaceModel) = model.system.y
