+++ src/StateSpaceModels.jl
@@ -46,6 +46,7 @@ include("models/sarima.jl")
 include("models/linear_regression.jl")
 include("models/unobserved_components.jl")
 include("models/exponential_smoothing.jl")
+include("models/naive_models.jl")
 
 include("visualization/forecast.jl")
 include("visualization/components.jl")
