+++ src/models/naive_models.jl
@@ -1,0 +1,149 @@
+abstract type NaiveModel end
+
+function isfitted(model::NaiveModel)
+    return model.fitted
+end
+
+isunivariate(model::NaiveModel) = true
+length_observations(model::NaiveModel) = length(model.y)
+observations(model::NaiveModel) = model.y
+get_standard_residuals(model::NaiveModel) = model.residuals ./ sqrt(model.sigma2)
+typeof_model_elements(model::NaiveModel) = eltype(model.y)
+
+function assert_zero_missing_values(model::NaiveModel)
+    for i in 1:length(model.y)
+        if isnan(y[i])
+            return error("model $(typeof(model)) does not support missing values.)")
+        end
+    end
+    return nothing
+end
+
+@doc raw"""
+    Naive(y::Vector{<:Real})
+
+# TODO
+
+# References
+ * Hyndman, Rob J., Athanasopoulos, George. "Forecasting: Principles and Practice"
+"""
+mutable struct Naive <: NaiveModel
+    y::Vector{<:Real}
+    residuals::Vector{<:Real}
+    sigma2::Real
+    fitted::Bool
+
+    function Naive(y::Vector{<:Real})
+        Fl = eltype(y)
+        n = length(y)
+        return new(y, zeros(Fl, n), zero(Fl), false)
+    end
+end
+
+function fit!(model::Naive)
+    residuals = model.y[2:end] - model.y[1:end-1]
+    model.residuals = residuals
+    model.sigma2 = var(residuals)
+    model.fitted = true
+    return model
+end
+
+function forecast(model::Naive, steps_ahead::Int)
+    @assert isfitted(model)
+    Fl = eltype(model.y)
+    expected_value = Vector{Vector{Fl}}(undef, steps_ahead)
+    covariance = Vector{Matrix{Fl}}(undef, steps_ahead)
+    for i in 1:steps_ahead
+        expected_value[i] = [model.y[end]]
+        covariance[i] = [model.sigma2 * i][:, :]
+    end
+    return Forecast{Fl}(expected_value, covariance)
+end
+
+function simulate_scenarios(model::Naive, steps_ahead::Int, n_scenarios::Int)
+    Fl = typeof_model_elements(model)
+    scenarios = Array{Fl, 3}(undef, steps_ahead, num_series, n_scenarios)
+    for s in 1:n_scenarios, i in 1:steps_ahead
+        if i == 1
+            scenarios[i, 1, s] = model.y[end] + rand(model.residuals)
+        else
+            scenarios[i, 1, s] = scenarios[i-1, 1, s] + rand(model.residuals)
+        end
+    end
+    return scenarios
+end
+
+@doc raw"""
+    SeasonalNaive(y::Vector{<:Real}, seasoanl::Int)
+
+# TODO
+
+# References
+ * Hyndman, Rob J., Athanasopoulos, George. "Forecasting: Principles and Practice"
+"""
+mutable struct SeasonalNaive <: NaiveModel
+    y::Vector{<:Real}
+    residuals::Vector{<:Real}
+    seasonal::Int
+    sigma2::Real
+    fitted::Bool
+
+    function SeasonalNaive(y::Vector{<:Real}, seasonal::Int)
+        Fl = eltype(y)
+        n = length(y)
+        return new(y, zeros(Fl, n), seasonal, zero(Fl), false)
+    end
+end
+
+function fit!(model::SeasonalNaive)
+    residuals = model.y[seasonal+1:end] - model.y[1:end-seasonal]
+    model.residuals = residuals
+    model.sigma2 = var(residuals)
+    model.fitted = true
+    return model
+end
+
[…]
