+++ Project.toml
@@ -1,7 +1,7 @@
 name = "StateSpaceModels"
 uuid = "99342f36-827c-5390-97c9-d7f9ee765c78"
 authors = ["raphaelsaavedra <raphael.saavedra93@gmail.com>, guilhermebodin <guilherme.b.moraes@gmail.com>, mariohsouto"]
-version = "0.5.9"
+version = "0.5.10"
 
 [deps]
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
@@ -17,14 +17,14 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [compat]
-Distributions = "0.23, 0.24"
+Distributions = "0.16-0.24"
 MatrixEquations = "1"
-Optim = "0.20, 0.21, 0.22, 1.2"
+Optim = "0.17-0.22, 1.2"
 OrderedCollections = "1"
 Polynomials = "1"
 RecipesBase = "1"
 ShiftedArrays = "1"
-StatsBase = "0.33"
+StatsBase = "0.27-0.33"
 julia = "1"
 
 [extras]
