+++ Project.toml
@@ -17,14 +17,14 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
 
 [compat]
-Distributions = "0.16-0.24"
+Distributions = "0.16, 0.17, 0.18, 0.19, 0.20, 0.21, 0.22, 0.23, 0.24"
 MatrixEquations = "1"
-Optim = "0.17-0.22, 1.2"
+Optim = "0.17, 0.18, 0.19, 0.20, 0.21, 0.22, 1.2"
 OrderedCollections = "1"
 Polynomials = "1"
 RecipesBase = "1"
 ShiftedArrays = "1"
-StatsBase = "0.27-0.33"
+StatsBase = "0.27, 0.28, 0.29, 0.30, 0.31, 0.32, 0.33"
 julia = "1"
 
 [extras]
