+++ src/models/naive_models.jl
@@ -48,29 +48,37 @@ function fit!(model::Naive)
     return model
 end
 
-function forecast(model::Naive, steps_ahead::Int)
+function forecast(model::Naive, steps_ahead::Int; bootstrap::Bool = false)
     @assert isfitted(model)
     Fl = eltype(model.y)
     expected_value = Vector{Vector{Fl}}(undef, steps_ahead)
     covariance = Vector{Matrix{Fl}}(undef, steps_ahead)
-    for i in 1:steps_ahead
-        expected_value[i] = [model.y[end]]
-        covariance[i] = [model.sigma2 * i][:, :]
+    if bootstrap
+        scenarios = simulate_scenarios(model, steps_ahead, 10_000)
+        for i in 1:steps_ahead
+            expected_value[i] = [mean(scenarios[i, 1, :])]
+            covariance[i] = [var(scenarios[i, 1, :])][:, :]
+        end
+    else
+        for i in 1:steps_ahead
+            expected_value[i] = [model.y[end]]
+            covariance[i] = [model.sigma2 * i][:, :]
+        end
     end
     return Forecast{Fl}(expected_value, covariance)
 end
 
 function simulate_scenarios(model::Naive, steps_ahead::Int, n_scenarios::Int)
     Fl = typeof_model_elements(model)
-    scenarios = Array{Fl, 3}(undef, steps_ahead, num_series, n_scenarios)
+    scenarios = Array{Fl, 3}(undef, steps_ahead, 1, n_scenarios)
     for s in 1:n_scenarios, i in 1:steps_ahead
-        if i == 1
-            scenarios[i, 1, s] = model.y[end] + rand(model.residuals)
-        else
-            scenarios[i, 1, s] = scenarios[i-1, 1, s] + rand(model.residuals)
-        end
+        scenarios[i, 1, s] = model.y[end] + rand(model.residuals)
     end
     return scenarios
+end
+
+function reinstantiate(::Naive, y::Vector{<:Real})
+    return Naive(y)
 end
 
 @doc raw"""
@@ -96,44 +104,62 @@ mutable struct SeasonalNaive <: NaiveModel
 end
 
 function fit!(model::SeasonalNaive)
-    residuals = model.y[seasonal+1:end] - model.y[1:end-seasonal]
+    residuals = model.y[model.seasonal+1:end] - model.y[1:end-model.seasonal]
     model.residuals = residuals
     model.sigma2 = var(residuals)
     model.fitted = true
     return model
 end
 
-function forecast(model::SeasonalNaive, steps_ahead::Int)
+function forecast(model::SeasonalNaive, steps_ahead::Int; bootstrap::Bool = false)
     @assert isfitted(model)
     Fl = eltype(model.y)
     expected_value = Vector{Vector{Fl}}(undef, steps_ahead)
     covariance = Vector{Matrix{Fl}}(undef, steps_ahead)
-    for i in 1:steps_ahead
-        expected_value[i] = [model.y[end - seasonal + i]]
-        covariance[i] = [model.sigma2 * (floor(Int, (i - 1)/model.seasonal) + 1)][:, :]
+    if bootstrap
+        scenarios = simulate_scenarios(model, steps_ahead, 10_000)
+        for i in 1:steps_ahead
+            expected_value[i] = [mean(scenarios[i, 1, :])]
+            covariance[i] = [var(scenarios[i, 1, :])][:, :]
+        end
+    else
+        for i in 1:steps_ahead
+            if i in 1:model.seasonal
+                expected_value[i] = [model.y[end - model.seasonal + i]]
+            else
+                expected_value[i] = copy(expected_value[i - model.seasonal])
+            end
[…]
