+++ test/models/locallevelexplanatory.jl
@@ -42,6 +42,7 @@
     forec = forecast(model, ones(5, 2))
     @test monotone_forecast_variance(forec)
     kf = kalman_filter(model)
+    ks = kalman_smoother(model)
     a = get_predictive_state(kf)
     @test a[1, 2] ≈ a[end, 2] atol=1e-3
     @test a[1, 3] ≈ a[end, 3] atol=1e-3
