### 2021-04-23 42143c5
Add simulate function
