+++ src/models/basicstructural_explanatory.jl
@@ -1,0 +1,203 @@
+@doc raw"""
+    BasicStructuralExplanatory(y::Vector{Fl}, s::Int, X::Matrix{Fl}) where Fl
+
+It is defined by:
+```math
+\begin{gather*}
+    \begin{aligned}
+        y_{t} &=  \mu_{t} + \gamma_{t} + \beta_{t, i}X_{t, i} \varepsilon_{t} \quad &\varepsilon_{t} \sim \mathcal{N}(0, \sigma^2_{\varepsilon})\\
+        \mu_{t+1} &= \mu_{t} + \nu_{t} + \xi_{t} \quad &\xi_{t} \sim \mathcal{N}(0, \sigma^2_{\xi})\\
+        \nu_{t+1} &= \nu_{t} + \zeta_{t} \quad &\zeta_{t} \sim \mathcal{N}(0, \sigma^2_{\zeta})\\
+        \gamma_{t+1} &= -\sum_{j=1}^{s-1} \gamma_{t+1-j} + \omega_{t} \quad & \omega_{t} \sim \mathcal{N}(0, \sigma^2_{\omega})\\
+        \beta_{t+1} &= \beta_{t}
+    \end{aligned}
+\end{gather*}
+```
+
+# Example
+```jldoctest
+julia> model = BasicStructuralExplanatory(rand(100), 12, rand(100, 2))
+BasicStructuralExplanatory model
+```
+
+# References
+ * Durbin, James, & Siem Jan Koopman. (2012). "Time Series Analysis by State Space Methods: Second Edition." Oxford University Press.
+"""
+mutable struct BasicStructuralExplanatory <: StateSpaceModel
+    hyperparameters::HyperParameters
+    system::LinearUnivariateTimeVariant
+    seasonality::Int
+    results::Results
+    exogenous::Matrix
+
+    function BasicStructuralExplanatory(y::Vector{Fl}, s::Int, X::Matrix{Fl}) where Fl
+
+        @assert length(y) == size(X, 1)
+
+        num_observations = size(X, 1)
+        num_exogenous = size(X, 2)
+
+        Z = [vcat([1; 0; 1; zeros(Fl, s - 2)], X[t, :]) for t in 1:num_observations]
+        T = [[
+            1 1 zeros(Fl, 1, s - 1) zeros(Fl, 1, num_exogenous)
+            0 1 zeros(Fl, 1, s - 1) zeros(Fl, 1, num_exogenous)
+            0 0 -ones(Fl, 1, s - 1) zeros(Fl, 1, num_exogenous)
+            zeros(Fl, s - 2, 2) Matrix{Fl}(I, s - 2, s - 2) zeros(Fl, s - 2) zeros(Fl, 10, num_exogenous)
+            zeros(Fl, num_exogenous, 13) Matrix{Fl}(I, num_exogenous, num_exogenous)
+        ] for _ in 1:num_observations]
+        R = [[
+            Matrix{Fl}(I, 3, 3)
+            zeros(Fl, s - 2, 3)
+            zeros(num_exogenous, 3)
+        ] for _ in 1:num_observations]
+        d = [zero(Fl) for _ in 1:num_observations]
+        c = [zeros(Fl, size(T[1], 1)) for _ in 1:num_observations]
+        H = [one(Fl) for _ in 1:num_observations]
+        Q = [zeros(Fl, 3, 3) for _ in 1:num_observations]
+
+        system = LinearUnivariateTimeVariant{Fl}(y, Z, T, R, d, c, H, Q)
+
+        names = [["sigma2_ε", "sigma2_ξ", "sigma2_ζ", "sigma2_ω"]; ["β_$i" for i in 1:num_exogenous]]
+
[…]
