### 2021-04-26 cddda49
Add scenarios exogenous in bse
