### 2021-04-26 9a43a36
autoarima first skeleton
Co-authored-by: Marina Dietze <marinadietze@users.noreply.github.com>
