+++ src/StateSpaceModels.jl
@@ -22,6 +22,8 @@ include("hyperparameters.jl")
 include("systems.jl")
 include("kalman_filter_and_smoother.jl")
 
+include("kpss.jl")
+
 include("filters/univariate_kalman_filter.jl")
 include("filters/multivariate_kalman_filter.jl")
 include("filters/scalar_kalman_filter.jl")
