+++ src/models/sarima.jl
@@ -678,14 +678,166 @@ function model_name(model::SARIMA)
 end
 
 # autoarima functions
+abstract type UnitRootTests end
+
+struct KPSS <: UnitRootTests end
+struct ADF <: UnitRootTests end
+
+
+function assert_ic(ic::Symbol)
+    @assert ic in [:AIC, :AICc, :BIC]
+    return 
+end
+
+function assert_test(test::Symbol)
+    @assert test in [:KPSS, :ADF]
+    return
+end
+
+function assert_non_negativity(order::Int)
+    @assert order >= 0
+    return
+end
+
+function assert_frequency(frequency, seasonal::Bool)
+    if seasonal 
+        @assert frequency > 0
+    end
+    return
+end
+
+function assert_parameters(frequency,
+                            d,
+                            D,
+                            max_p::Int,
+                            max_q::Int,
+                            max_P::Int,
+                            max_Q::Int,
+                            max_d::Int,
+                            max_D::Int,
+                            start_p::Int,
+                            start_q::Int,
+                            start_P::Int,
+                            start_Q::Int,
+                            seasonal::Bool,
+                            ic::Symbol,
+                            test::Symbol,
+                            )
+    assert_frequency(frequency, seasonal)
+    isnan(d) ? nothing : assert_non_negativity(d)
+    isnan(D) ? nothing : assert_non_negativity(D)
+    assert_non_negativity(max_p)
+    assert_non_negativity(max_q)
+    assert_non_negativity(max_P)
+    assert_non_negativity(max_Q)
+    assert_non_negativity(max_d)
+    assert_non_negativity(max_D)
+    assert_non_negativity(start_p)
+    assert_non_negativity(start_q)
+    assert_non_negativity(start_P)
+    assert_non_negativity(start_Q)
+    assert_test(test)
+    assert_ic(ic)
+
+    return 
+end
+
+function handle_seasonal_integration_order(y::Vector{Float64}, max_D::Int, D)
+    
+    return !isnan(D) ? D : handle_seasonal_integration_order(y, max_D)
+end
+
+function handle_seasonal_integration_order(y::Vector{Float64}, max_D::Int)
+    #todo 
+    
+    return 0
+end
+
+function handle_seasonal_diff(y::Vector{Float64}, frequency::Int)
+    
+    return y[frequency + 1:end] .- y[1:end - frequency]
+end
+
+function handle_test_args(test_args::Dict)
+    if isempty(test_args)
+        test_args[:null] = :level
+        test_args[:lags] = true
+    end
+
+    return test_args
+end
+
+# function handle_test_args(test_args::Dict, y::Vector{Float64})
+#     if isempty(test_args)
+#         test_args[:null] = :constant
+#         test_args[:lags] = trunc(Int64, (length(y) - 1)^(1/3))
+#     end
+#     return test_args
+# end
+
+function handle_integration_order(y::Vector{Float64}, max_d::Int, D::Int, frequency, test_args::Dict, ::Type{KPSS})
+    test_parms = handle_test_args(test_args)
+    seasonal_diff_y = D > 0 ? handle_seasonal_diff(y, frequency) : y
+    p_value = KPSSTest(seasonal_diff_y; null = test_parms[:null], lags = test_parms[:lags])
+    d = 0
+    
+    while p_value <= 0.01 && d < max_d
+       seasonal_diff_y = diff(seasonal_diff_y)
+       p_value = KPSSTest(seasonal_diff_y; null = test_parms[:null], lags = test_parms[:lags])   
+       d += 1
+    end
+    
+    return d
+end
+
+function handle_integration_order(y::Vector{Float64}, max_d::Int, D::Int, frequency, test_args::Dict, test::Symbol)
+    unit_root_type = handle_unit_root_type(test)
[…]
