+++ test/models/basicstructural_explanatory.jl
@@ -10,11 +10,8 @@
     # For a fixed forecasting explanatory the variance must not decrease
     forec = forecast(model, ones(10, 2))
     @test monotone_forecast_variance(forec)
-    kf = kalman_filter(model);
-    ks = kalman_smoother(model);
-    a = get_predictive_state(kf)
-    @test a[1, 14] ≈ a[end, 14] atol=1e-3
-    @test a[1, 15] ≈ a[end, 15] atol=1e-3
+    kf = kalman_filter(model)
+    ks = kalman_smoother(model)
     @test_throws AssertionError simulate_scenarios(model, 10, 1000, ones(5, 2))
     scenarios = simulate_scenarios(model, 10, 1000, ones(10, 2))
     test_scenarios_adequacy_with_forecast(forec, scenarios)
