+++ test/runtests.jl
@@ -24,6 +24,7 @@ include("models/sarima.jl")
 include("models/unobserved_components.jl")
 include("models/linear_regression.jl")
 include("models/exponential_smoothing.jl")
+include("models/naive_models.jl")
 
 # Visualization
 include("visualization/forecast.jl")
