+++ src/cross_validation.jl
@@ -1,9 +1,9 @@
-struct Backtest{Fl <: AbstractFloat}
+struct CrossValidation{Fl <: AbstractFloat}
     abs_errors::Matrix{Fl}
     mae::Vector{Fl}
     crps_scores::Matrix{Fl}
     mean_crps::Vector{Fl}
-    function Backtest{Fl}(n::Int, steps_ahead::Int) where Fl
+    function CrossValidation{Fl}(n::Int, steps_ahead::Int) where Fl
         abs_errors = Matrix{Fl}(undef, steps_ahead, n)
         crps_scores = Matrix{Fl}(undef, steps_ahead, n)
         mae = Vector{Fl}(undef, steps_ahead)
@@ -34,27 +34,27 @@ function evaluate_crps(y::Vector{Fl}, scenarios::Matrix{Fl}) where {Fl}
 end
 
 """
-    backtest(model::StateSpaceModel, steps_ahead::Int, start_idx::Int;
+    cross_validation(model::StateSpaceModel, steps_ahead::Int, start_idx::Int;
              n_scenarios::Int = 10_000,
              filter::KalmanFilter=default_filter(model),
              optimizer::Optimizer=default_optimizer(model)) where Fl
 
 Makes rolling window estimating and forecasting to benchmark the forecasting skill of the model
 in for different time periods and different lead times. The function returns a struct with the MAE
-and mean CRPS per lead time. See more on [Backtest the forecasts of a model](@ref)
+and mean CRPS per lead time. See more on [CrossValidation the forecasts of a model](@ref)
 
 # References
  * DTU course "31761 - Renewables in electricity markets" available on youtube https://www.youtube.com/watch?v=Ffo8XilZAZw&t=556s
 """
-function backtest(model::StateSpaceModel, steps_ahead::Int, start_idx::Int;
+function cross_validation(model::StateSpaceModel, steps_ahead::Int, start_idx::Int;
                   n_scenarios::Int = 10_000,
                   filter::KalmanFilter=default_filter(model),
                   optimizer::Optimizer=default_optimizer(model))
     Fl = typeof_model_elements(model)
     num_mle = length(model.system.y) - start_idx - steps_ahead
-    b = Backtest{Fl}(num_mle, steps_ahead)
+    cv = CrossValidation{Fl}(num_mle, steps_ahead)
     for i in 1:num_mle
-        println("Backtest: step $i of $num_mle")
+        println("CrossValidation: step $i of $num_mle")
         y_to_fit = model.system.y[1:start_idx - 1 + i]
         y_to_verify = model.system.y[start_idx + i:start_idx - 1 + i + steps_ahead]
         model_to_fit = reinstantiate(model, y_to_fit)
@@ -64,12 +64,12 @@ function backtest(model::StateSpaceModel, steps_ahead::Int, start_idx::Int;
         expected_value_vector = forecast_expected_value(forec)[:]
         abs_errors = evaluate_abs_error(y_to_verify, expected_value_vector)
         crps_scores = evaluate_crps(y_to_verify, scenarios[:, 1, :])
-        b.abs_errors[:, i] = abs_errors
-        b.crps_scores[:, i] = crps_scores
+        cv.abs_errors[:, i] = abs_errors
+        cv.crps_scores[:, i] = crps_scores
     end
     for i in 1:steps_ahead
-        b.mae[i] = mean(b.abs_errors[i, :])
-        b.mean_crps[i] = mean(b.crps_scores[i, :])
+        cv.mae[i] = mean(cv.abs_errors[i, :])
+        cv.mean_crps[i] = mean(cv.crps_scores[i, :])
     end
-    return b
+    return cv
 end
