### 2021-05-14 92a322e
Fix bug in Naive assertion and increase coverage
