+++ test/visualization/components.jl
@@ -20,4 +20,24 @@
     @test length(rec) == 3
     rec = RecipesBase.apply_recipe(Dict{Symbol, Any}(), model, ks)
     @test length(rec) == 3
+
+    model = ExponentialSmoothing(log_air_passengers; trend = true, seasonal = 12)
+    fit!(model)
+    kf = kalman_filter(model)
+    ks = kalman_smoother(model)
+
+    rec = RecipesBase.apply_recipe(Dict{Symbol, Any}(), model, kf)
+    @test length(rec) == 4
+    rec = RecipesBase.apply_recipe(Dict{Symbol, Any}(), model, ks)
+    @test length(rec) == 4
+
+    finland_fatalities = CSV.File(StateSpaceModels.VEHICLE_FATALITIES) |> DataFrame
+    log_finland_fatalities = log.(finland_fatalities.ff)
+    model = ExponentialSmoothing(log_finland_fatalities; trend = true)
+    fit!(model)
+
+    rec = RecipesBase.apply_recipe(Dict{Symbol, Any}(), model, kf)
+    @test length(rec) == 3
+    rec = RecipesBase.apply_recipe(Dict{Symbol, Any}(), model, ks)
+    @test length(rec) == 3
 end
