+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"examples/#Examples","page":"Examples","title":"Examples","text":"","category":"section"},{"location":"examples/","page":"Examples","title":"Examples","text":"In this section we show examples of applications and use cases of the package.","category":"page"},{"location":"examples/#Nile-river-annual-flow","page":"Examples","title":"Nile river annual flow","text":"","category":"section"},{"location":"examples/","page":"Examples","title":"Examples","text":"Here we will follow an example from Durbin & Koopman's book. We will use the LocalLevel model applied to the annual flow of the Nile river at the city of Aswan between 1871 and 1970.","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"using StateSpaceModels\nusing Plots\nusing Dates","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"First, we load the data:","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"using CSV, DataFrames\nnile = CSV.File(StateSpaceModels.NILE) |> DataFrame\nplt = plot(nile.year, nile.flow, label = \"Nile river annual flow\")","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"Next, we fit a LocalLevel model:","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"model = LocalLevel(nile.flow)\nfit!(model)","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"We can analyze the filtered estimates for the level of the annual flow:","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"filter_output = kalman_filter(model)\nplot!(plt, nile.year, get_filtered_state(filter_output), label = \"Filtered level\")","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"We can do the same for the smoothed estimates for the level of the annual flow:","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"smoother_output = kalman_smoother(model)\nplot!(plt, nile.year, get_smoothed_state(smoother_output), label = \"Smoothed level\")","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"StateSpaceModels.jl can also be used to obtain forecasts. Here we forecast 10 steps ahead:","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"steps_ahead = 10\ndates = collect(nile.year[end] + Year(1):Year(1):nile.year[end] + Year(10))\nforec = forecast(model, 10)\nexpected_value = forecast_expected_value(forec)\nplot!(plt, dates, expected_value, label = \"Forecast\")","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"We can also simulate multiple scenarios for the forecasting horizon based on the estimated predictive distributions.","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"scenarios = simulate_scenarios(model, 10, 100)\nplot!(plt, dates, scenarios[:, 1, :], label = \"\", color = \"grey\", width = 0.2)","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"The package also handles missing values automatically. To that end, the package considers that any NaN entries in the observations are missing values.","category":"page"},{"location":"examples/","page":"Examples","title":"Examples","text":"nile.flow[[collect(21:40); collect(61:80)]] .= NaN\np
[…]
