### 2021-06-07 c34b167
Added quick examples
