### 2021-06-08 eba9daa
Draft of auto_arima
