### 2021-06-09 2e9f829
remove model from print
