+++ src/models/sarima.jl
@@ -60,7 +60,8 @@ end
         y::Vector{Fl}; 
         order::Tuple{Int,Int,Int} = (1, 0, 0), 
         seasonal_order::Tuple{Int, Int, Int, Int} = (0, 0, 0, 0),
-        include_mean::Bool = false
+        include_mean::Bool = false,
+        suppress_warns::Bool = false
     ) where Fl
 
 A SARIMA model (Seasonal AutoRegressive Integrated Moving Average) implemented within the state-space
@@ -105,11 +106,13 @@ mutable struct SARIMA <: StateSpaceModel
     system::LinearUnivariateTimeInvariant
     results::Results
     include_mean::Bool
+    suppress_warns::Bool
 
     function SARIMA(y::Vector{Fl}; 
                     order::Tuple{Int,Int,Int} = (1, 0, 0), 
                     seasonal_order::Tuple{Int, Int, Int, Int} = (0, 0, 0, 0),
-                    include_mean::Bool = false) where Fl
+                    include_mean::Bool = false,
+                    suppress_warns::Bool = false) where Fl
         or = SARIMAOrder(order[1], order[2], order[3], 
                         seasonal_order[1], seasonal_order[2], seasonal_order[3], seasonal_order[4])
         hyperparameters_auxiliary = SARIMAHyperParametersAuxiliary{Fl}(or)
@@ -148,7 +151,7 @@ mutable struct SARIMA <: StateSpaceModel
 
         hyperparameters = HyperParameters{Fl}(names)
 
-        return new(or, hyperparameters_auxiliary, hyperparameters, system, Results{Fl}(), include_mean)
+        return new(or, hyperparameters_auxiliary, hyperparameters, system, Results{Fl}(), include_mean, suppress_warns)
     end
 end
 
@@ -581,22 +584,22 @@ function initial_hyperparameters!(model::SARIMA)
     # Non-seasonal ARMA
     (initial_ar, initial_ma) = conditional_sum_of_squares(y_diff, model.order.p, model.order.q)
     if !assert_stationarity(initial_ar)
-        @warn("Conditional sum of squares estimated initial_ar out of the unit circle, using zero as starting params")
+        model.suppress_warns || @warn("Conditional sum of squares estimated initial_ar out of the unit circle, using zero as starting params")
         initial_ar .= zero(Fl)
     end
     if !assert_invertibility(initial_ma)
-        @warn("Conditional sum of squares estimated initial_ma out of the unit circle, using zero as starting params")
+        model.suppress_warns || @warn("Conditional sum of squares estimated initial_ma out of the unit circle, using zero as starting params")
         initial_ma .= zero(Fl)
     end
     # Seasonal ARMA
     (initial_seasonal_ar, 
     initial_seasonal_ma) = conditional_sum_of_squares(y_diff, model.order.P, model.order.Q)
     if !assert_stationarity(initial_seasonal_ar)
-        @warn("Conditional sum of squares estimated initial_seasonal_ar out of the unit circle, using zero as starting params")
+        model.suppress_warns || @warn("Conditional sum of squares estimated initial_seasonal_ar out of the unit circle, using zero as starting params")
         initial_seasonal_ar .= zero(Fl)
     end
     if !assert_invertibility(initial_seasonal_ma)
-        @warn("Conditional sum of squares estimated initial_seasonal_ma out of the unit circle, using zero as starting params")
+        model.suppress_warns || @warn("Conditional sum of squares estimated initial_seasonal_ma out of the unit circle, using zero as starting params")
         initial_seasonal_ma .= zero(Fl)
     end
     for i in 1:(model.order.p)
[…]
