+++ test/models/sarima.jl
@@ -95,7 +95,22 @@
     fit!(model)
     @test loglike(model) ≈ 228.502 atol = 1e-3 rtol = 1e-3
 
-    # Automatic ARIMA
-    # test auto_arima
-    # with_logger(NullLogger()) do; StateSpaceModels.auto_arima(y) end
+    # auto arima tests
+    model = auto_arima(dinternet)
+    @test model.order.p == 1
+    @test model.order.q == 1
+    @test model.include_mean == false
+
+    nile = CSV.File(StateSpaceModels.NILE) |> DataFrame
+    model = auto_arima(nile.flow)
+    @test model.order.p == 1
+    @test model.order.d == 1
+    @test model.order.q == 1
+    @test model.include_mean == false
+
+    model = auto_arima(uschange_consumption.Consumption)
+    @test model.order.p == 1
+    @test model.order.d == 0
+    @test model.order.q == 3
+    @test model.include_mean == true
 end
