+++ src/models/sarima.jl
@@ -689,7 +689,7 @@ function repeated_kpss_test(y::Vector{Fl}, max_d::Int) where Fl
     d = 0
     while p_value <= 0.01 && d < max_d
        seasonal_diff_y = diff(seasonal_diff_y)
-       p_value = kpss_test(seasonal_diff_y; H0 = :trend) 
+       p_value = kpss_test(seasonal_diff_y) 
        d += 1
     end
     return d
