### 2021-06-10 69acc73
Minor improvment in canova test
