### 2021-06-10 bcebf4c
only run canova is seasonal != 0
