+++ src/canova_hensen.jl
@@ -9,22 +9,26 @@
 # Bruce E. HANSE
 # Journal of Business & Economic Statistics, July 1995, Vol. 13, No. 3
 
-λ(j, s) = (2*pi*j)/s
+λ(j::Int, s::Int) = (2*pi*j)/s
 
-CRIT_VALS_05 = [0.470, 0.749, 1.01, 1.24, 1.47, 1.68, 1.9, 2.11, 2.32, 2.54, 2.75, 2.96]
+const CRIT_VALS_05 = [0.470, 0.749, 1.01, 1.24, 1.47, 1.68, 1.9, 2.11, 2.32, 2.54, 2.75, 2.96]
 
-function crit_val_05_generalized_von_mises_distribution(s)
-    return s > 12 ? 0.269*s*0.928 : CRIT_VALS_05[s]
+function crit_val_05_generalized_von_mises_distribution(s::Int)
+    if s > 12
+        return 0.269*s*0.928
+    else
+        return CRIT_VALS_05[s]
+    end
 end
 
-function π_statistic(ε::Vector{Fl}, σ::Float64, T::Int64) where Fl
+function π_statistic(ε::Vector{Fl}, σ::Fl, T::Int) where Fl
     return (T^(-2)*σ^(-2))*
             sum(
                 sum(ε[i]*(-1)^(i) for i = 1:t) for t = 1:T
             )
 end
 
-function std_statistic(ε::Vector{Fl}, σ::Float64, T::Int64, s::Int64, j::Int64) where Fl
+function std_statistic(ε::Vector{Fl}, σ::Fl, T::Int, s::Int, j::Int) where Fl
     return (2*T^(-2)*σ^(-2))*
             sum(
                 sum(ε[i]*cos(λ(j, s)*i) for i = 1:t)^2 
@@ -33,7 +37,7 @@ function std_statistic(ε::Vector{Fl}, σ::Float64, T::Int64, s::Int64, j::Int64
             )
 end
 
-function test_statistic(ε::Vector{Fl}, σ::Float64, T::Int64, s::Int64) where Fl
+function test_statistic(ε::Vector{Fl}, σ::Fl, T::Int, s::Int) where Fl
     ω = Float64[]
     for j = 1:Int(floor(s/2)) #check
         ω_j = j == s/2 ? π_statistic(ε, σ, T) : std_statistic(ε, σ, T, s, j)
@@ -42,17 +46,17 @@ function test_statistic(ε::Vector{Fl}, σ::Float64, T::Int64, s::Int64) where F
     return sum(ω)
 end
 
-function build_std(T::Int64, s::Int64, j::Int64)
+function build_std(T::Int, s::Int, j::Int)
     t = collect(1:T)
     return [cos.(λ(j, s).*t) sin.(λ(j, s).*t)]
 end
 
-function build_π(T::Int64, s::Int64, j::Int64)
+function build_π(T::Int, s::Int, j::Int)
     t = collect(1:T)
     return [cos.(λ(j, s)*t) ones(T)]
 end
 
-function build_X(s::Int64, T::Int64)
+function build_X(s::Int, T::Int)
     X = Array{Float64, 2}(undef, T, s)
     for j = 1:Int(floor(s/2))
         X[:,[2*j-1, 2*j]] = j == s/2 ? build_π(T, s, j) : build_std(T, s, j)
@@ -60,14 +64,12 @@ function build_X(s::Int64, T::Int64)
     return X
 end
 
-function seasonal_regression(y::Vector{Fl}, s::Int64, T::Int64) where Fl
+function seasonal_regression(y::Vector{Fl}, s::Int, T::Int) where Fl
     X = build_X(s, T)
[…]
