+++ src/canova_hensen.jl
@@ -5,8 +5,8 @@
 
 # Are Seasonal Patterns Constant Over Time?
 # A Test for Seasonal Stability 
-# Fabio CANO
-# Bruce E. HANSE
+# Fabio CANOVA
+# Bruce E. HANSEN
 # Journal of Business & Economic Statistics, July 1995, Vol. 13, No. 3
 
 λ(j::Int, s::Int) = (2*pi*j)/s
@@ -15,7 +15,7 @@ const CRIT_VALS_05 = [0.470, 0.749, 1.01, 1.24, 1.47, 1.68, 1.9, 2.11, 2.32, 2.5
 
 function crit_val_05_generalized_von_mises_distribution(s::Int)
     if s > 12
-        return 0.269*s*0.928
+        return 0.269*s^0.928
     else
         return CRIT_VALS_05[s]
     end
@@ -38,8 +38,9 @@ function std_statistic(ε::Vector{Fl}, σ::Fl, T::Int, s::Int, j::Int) where Fl
 end
 
 function test_statistic(ε::Vector{Fl}, σ::Fl, T::Int, s::Int) where Fl
-    ω = Float64[]
-    for j = 1:Int(floor(s/2)) #check
+    # TODO we know this size a priori
+    ω = Fl[]
+    for j = 1:floor(Int, s/2) #check
         ω_j = j == s/2 ? π_statistic(ε, σ, T) : std_statistic(ε, σ, T, s, j)
         push!(ω, ω_j)
     end
@@ -47,26 +48,24 @@ function test_statistic(ε::Vector{Fl}, σ::Fl, T::Int, s::Int) where Fl
 end
 
 function build_std(T::Int, s::Int, j::Int)
-    t = collect(1:T)
-    return [cos.(λ(j, s).*t) sin.(λ(j, s).*t)]
+    return hcat(cos.(λ(j, s)*(1:T)), sin.(λ(j, s)*(1:T)))
 end
 
 function build_π(T::Int, s::Int, j::Int)
-    t = collect(1:T)
-    return [cos.(λ(j, s)*t) ones(T)]
+    return hcat(cos.(λ(j, s)*(1:T)), ones(T))
 end
 
-function build_X(s::Int, T::Int)
-    X = Array{Float64, 2}(undef, T, s)
-    for j = 1:Int(floor(s/2))
+function build_X(s::Int, T::Int, Fl::Type)
+    X = Array{Fl, 2}(undef, T, s)
+    for j = 1:floor(Int, s/2)
         X[:,[2*j-1, 2*j]] = j == s/2 ? build_π(T, s, j) : build_std(T, s, j)
     end
     return X
 end
 
 function seasonal_regression(y::Vector{Fl}, s::Int, T::Int) where Fl
-    X = build_X(s, T)
-    return y - X \ y * y
+    X = build_X(s, T, Fl)
+    return y - X * (X \ y)
 end
 
 function seasonal_stationarity_test(y::Vector{Fl}, s::Int) where Fl
