### 2021-06-11 26ea213
update P and Q only is seasonal is given
