+++ versions.js
@@ -5,3 +5,5 @@ var DOC_VERSIONS = [
   "v0.2",
   "dev",
 ];
+var DOCUMENTER_NEWEST = "v0.5.15";
+var DOCUMENTER_STABLE = "stable";
