### 2021-06-11 4439a54
path to the SeasonalTrendLoess module
