### 2021-06-11 0ca8c74
Add seasonal_strength test and modify auto_arima function
