+++ src/StateSpaceModels.jl
@@ -13,7 +13,7 @@ using Optim
 using OrderedCollections
 using RecipesBase
 using StatsBase
-# using SeasonalTrendLoess
+using SeasonalTrendLoess
 
 abstract type StateSpaceModel end
 
@@ -24,8 +24,6 @@ include("systems.jl")
 include("kalman_filter_and_smoother.jl")
 include("kpss.jl")
 include("canova_hensen.jl")
-include("../../SeasonalTrendLoess.jl/src/SeasonalTrendLoess.jl") # path to stl module
-
 include("filters/univariate_kalman_filter.jl")
 include("filters/multivariate_kalman_filter.jl")
 include("filters/scalar_kalman_filter.jl")
