### 2021-06-11 523ec6c
Change test and seasonal_test names
