+++ src/models/sarima.jl
@@ -889,10 +889,20 @@ end
                max_order::Int = 5,
                information_criteria::String = "aicc",
                allow_mean::Bool = true,
-               show_trace::Bool = false
+               show_trace::Bool = false,
+               integration_test::String = "kpss",
+               seasonal_integration_test::String = "seas"
                ) where Fl
 
-TODO
+
+Automatically fits the best [`SARIMA`](@ref) model according to the best information criteria 
+
+# TODO write example
+
+# References
+ * Hyndman, RJ and Khandakar. 
+ Automatic time series forecasting: The forecast package for R.
+Journal of Statistical Software, 26(3), 2008.
 """
 function auto_arima(y::Vector{Fl};
                     seasonal::Int = 0,
