+++ Project.toml
@@ -24,6 +24,7 @@ Optim = "0.17, 0.18, 0.19, 0.20, 0.21, 0.22, 1.2"
 OrderedCollections = "1"
 Polynomials = "1"
 RecipesBase = "1"
+SeasonalTrendLoess = "0.1"
 ShiftedArrays = "1"
 StatsBase = "0.27, 0.28, 0.29, 0.30, 0.31, 0.32, 0.33"
 julia = "1"
