+++ src/StateSpaceModels.jl
@@ -37,7 +37,6 @@ include("prints.jl")
 include("forecast.jl")
 include("cross_validation.jl")
 
-
 include("models/common.jl")
 include("models/locallevel.jl")
 include("models/locallevelcycle.jl")
