### 2023-09-07 #28 @luraess
Dear author @raphaelsaavedra , we are in the process of resuming the stalled JuliaCon reviews. If you are still interested in getting your work published, please start addressing the author-action-required items within the coming 3 weeks. If no action is taken within the coming 3 weeks, we will close the submission. Thank you for your understanding.
